Join RtData

Joins two nested lists in the format required by summaryWidget. This may be useful for merging estimates from disparate data sources or linking national level estimates with subnational estimates

joinRtData(rtData, rtData2)

Arguments

rtData
A nested list as required by summaryWidget

rtData2
A nested list as required by summaryWidget

Value

A nested list as required by summaryWidget

Examples

# \donttest{
base_url <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/master/"
subnational <-   national <- list("Cases" = readInEpiNow2(
path = paste0(base_url, "subnational/italy/cases/summary"),
region_var = "region"))

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

national <- list("Cases" = readInEpiNow2(
path = paste0(base_url, "national/cases/summary"),
region_var = "country"),
regions = "Italy")

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

out <- list()
out$Cases <- joinRtData(subnational$Cases, national$Cases)
# }