summaryWidget

Create an Rt visualisation using D3. Need convenience functions to define defaults

summaryWidget(
  geoData = NULL,
  rtData = NULL,
  data_ref = NULL,
  subregional_ref = NULL,
  ts_color_ref = NULL,
  ts_bar_color = "lightgrey",
  projection = "geoEquirectangular",
  map_legend_ref = NULL,
  credible_threshold = 10,
  width = NULL,
  activeArea = "United Kingdom",
  dryRun = FALSE
)

Arguments

geoData
sf object, map data

rtData
data.frame, rt estimates in the format 'Source':'rtData':x, 'casesInfectionData':x, 'casesReportData':x, 'obsCasesData':x, ...

data_ref
list, reference for input data column names. Specify the column holding geometry to be symbolized 'rtData':'geometry_name':'region', ...

subregional_ref
list, reference to subnational estimates in the format 'country_name':'url', ....

ts_color_ref
list, reference for colors for time series plots.

ts_bar_color
string, color of observed cases bars in time series plots.

projection
string, map projection, must be named in d3-geo-projection.

map_legend_ref
list, reference for map legend variables

credible_threshold
integer, Threshold for credible intervals, maximum observed cases * this value will be removed.

width
integer, Width of widget in pixels.

activeArea
string, Area to symbolize first.

dryRun
Logical, defaults to FALSE. Should the function be tested without the widget being created. Useful for checking the integrity of input data.