Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

rtData

#> $summaryData
#>              region New confirmed cases by infection date
#>   1:    Afghanistan                       181 (20 -- 527)
#>   2:        Albania                      340 (146 -- 549)
#>   3:        Algeria                      431 (204 -- 645)
#>   4:        Andorra                       211 (31 -- 560)
#>   5:         Angola                      330 (142 -- 609)
#>  ---                                                     
#> 193:        Vietnam                           3 (0 -- 13)
#> 194: Western Sahara                          1 (0 -- 289)
#> 195:          Yemen                            0 (0 -- 0)
#> 196:         Zambia                         38 (16 -- 80)
#> 197:       Zimbabwe                          14 (0 -- 41)
#>      Expected change in daily cases Effective reproduction no.
#>   1:              Likely increasing           1.2 (0.8 -- 1.6)
#>   2:                         Unsure             1 (0.7 -- 1.3)
#>   3:              Likely increasing           1.1 (0.9 -- 1.3)
#>   4:              Likely increasing           1.2 (0.9 -- 1.5)
#>   5:              Likely increasing           1.1 (0.9 -- 1.3)
#>  ---                                                          
#> 193:                         Unsure           0.9 (0.4 -- 1.5)
#> 194:                         Unsure             0.6 (0 -- 1.4)
#> 195:                     Decreasing             0.8 (0.5 -- 1)
#> 196:                         Unsure           0.9 (0.8 -- 1.2)
#> 197:                         Unsure           0.9 (0.4 -- 1.4)
#>             Rate of growth Doubling/halving time (days)
#>   1:  0.05 (-0.05 -- 0.16)            15.4 (4.2 -- -15)
#>   2:  0.01 (-0.07 -- 0.07)          135.1 (9.9 -- -9.5)
#>   3:   0.03 (-0.03 -- 0.1)          20.2 (7.1 -- -26.1)
#>   4:  0.05 (-0.02 -- 0.13)          14.7 (5.2 -- -29.1)
#>   5:  0.03 (-0.01 -- 0.07)          25.9 (9.3 -- -50.1)
#>  ---                                                   
#> 193: -0.02 (-0.18 -- 0.14)          -39.2 (4.9 -- -3.8)
#> 194:  -0.1 (-0.36 -- 0.14)           -6.7 (4.8 -- -1.9)
#> 195: -0.06 (-0.15 -- 0.01)         -12.4 (85.7 -- -4.6)
#> 196: -0.02 (-0.05 -- 0.05)          -40 (14.6 -- -13.7)
#> 197: -0.03 (-0.17 -- 0.12)              -27.6 (6 -- -4)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-22 estimate   0.93 0.93 0.07     0.92     0.95
#>     2: Afghanistan 2020-08-23 estimate   0.93 0.93 0.07     0.92     0.94
#>     3: Afghanistan 2020-08-24 estimate   0.93 0.93 0.07     0.91     0.94
#>     4: Afghanistan 2020-08-25 estimate   0.94 0.94 0.07     0.93     0.96
#>     5: Afghanistan 2020-08-26 estimate   0.94 0.94 0.07     0.94     0.96
#>    ---                                                                   
#> 16349:    Zimbabwe 2020-11-10 forecast   0.91 0.93 0.33     0.79     0.92
#> 16350:    Zimbabwe 2020-11-11 forecast   0.91 0.93 0.33     0.79     0.92
#> 16351:    Zimbabwe 2020-11-12 forecast   0.91 0.93 0.33     0.79     0.92
#> 16352:    Zimbabwe 2020-11-13 forecast   0.91 0.93 0.33     0.79     0.92
#> 16353:    Zimbabwe 2020-11-14 forecast   0.91 0.93 0.33     0.79     0.92
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.90     0.97     0.82     1.02
#>     2:     0.90     0.97     0.82     1.03
#>     3:     0.90     0.97     0.83     1.04
#>     4:     0.90     0.97     0.83     1.04
#>     5:     0.90     0.98     0.84     1.05
#>    ---                                    
#> 16349:     0.73     1.09     0.38     1.39
#> 16350:     0.73     1.09     0.38     1.39
#> 16351:     0.73     1.09     0.38     1.39
#> 16352:     0.73     1.09     0.38     1.39
#> 16353:     0.73     1.09     0.38     1.39
#> 
#> $casesInfectionData
#>             region       date     type median mean    sd lower_20 upper_20
#>     1: Afghanistan 2020-08-22 estimate     46   47    12       43       48
#>     2: Afghanistan 2020-08-23 estimate     45   46    12       38       43
#>     3: Afghanistan 2020-08-24 estimate     44   45    12       36       42
#>     4: Afghanistan 2020-08-25 estimate     44   45    12       38       43
#>     5: Afghanistan 2020-08-26 estimate     43   44    12       37       43
#>    ---                                                                    
#> 16349:    Zimbabwe 2020-11-10 forecast     11  159  2646        0        3
#> 16350:    Zimbabwe 2020-11-11 forecast     11  237  4457        0        2
#> 16351:    Zimbabwe 2020-11-12 forecast     10  360  7545        0        2
#> 16352:    Zimbabwe 2020-11-13 forecast     10  561 12829        0        2
#> 16353:    Zimbabwe 2020-11-14 forecast     10  890 21899        0        2
#>        lower_50 upper_50 lower_90 upper_90
#>     1:       37       52       26       65
#>     2:       36       51       26       64
#>     3:       36       51       26       64
#>     4:       35       50       26       64
#>     5:       34       48       25       62
#>    ---                                    
#> 16349:        0       11        0       88
#> 16350:        0       11        0       95
#> 16351:        0       10        0      102
#> 16352:        0       10        0      110
#> 16353:        0       10        0      118
#> 
#> $casesReportData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-22 estimate     20   39   55        0        4
#>     2: Afghanistan 2020-08-23 estimate     30   57   77        0        6
#>     3: Afghanistan 2020-08-24 estimate     23   42   55        0        5
#>     4: Afghanistan 2020-08-25 estimate     37   67   86        0        8
#>     5: Afghanistan 2020-08-26 estimate     29   56   82        0        6
#>    ---                                                                   
#> 16349:    Zimbabwe 2020-11-10 forecast     13   39  242        0        4
#> 16350:    Zimbabwe 2020-11-11 forecast     10   36  213        0        3
#> 16351:    Zimbabwe 2020-11-12 forecast      9   44  383        1        3
#> 16352:    Zimbabwe 2020-11-13 forecast     10   73 1054        0        3
#> 16353:    Zimbabwe 2020-11-14 forecast     11   92 1727        0        3
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0       20        0      101
#>     2:        0       30        0      144
#>     3:        0       23        0      108
#>     4:        0       37        0      168
#>     5:        0       29        0      139
#>    ---                                    
#> 16349:        0       13        0       63
#> 16350:        0       10        0       56
#> 16351:        0        9        0       54
#> 16352:        0       10        0       67
#> 16353:        0       11        0       72
#> 
#> $obsCasesData
#>                 region       date confirm
#>     1:     Afghanistan 2020-08-08      78
#>     2:         Albania 2020-08-08     135
#>     3:         Algeria 2020-08-08     529
#>     4:         Andorra 2020-08-08      11
#>     5:          Angola 2020-08-08      88
#>    ---                                   
#> 17795: Wallis & Futuna 2020-10-31       0
#> 17796:  Western Sahara 2020-10-31       0
#> 17797:           Yemen 2020-10-31       0
#> 17798:          Zambia 2020-10-31      90
#> 17799:        Zimbabwe 2020-10-31      13
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France               61735 (29379 -- 111441)                     Increasing
#>    Effective reproduction no.  Rate of growth Doubling/halving time (days)
#> 1:             1.2 (1 -- 1.4) 0.04 (0 -- 0.1)           16.7 (7 -- -282.7)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-08-22                       estimate   1.17 1.18 0.04     1.16
#>  2: France 2020-08-23                       estimate   1.17 1.17 0.04     1.16
#>  3: France 2020-08-24                       estimate   1.17 1.17 0.04     1.14
#>  4: France 2020-08-25                       estimate   1.16 1.17 0.04     1.15
#>  5: France 2020-08-26                       estimate   1.16 1.16 0.04     1.15
#>  6: France 2020-08-27                       estimate   1.16 1.16 0.04     1.14
#>  7: France 2020-08-28                       estimate   1.15 1.15 0.04     1.14
#>  8: France 2020-08-29                       estimate   1.15 1.15 0.04     1.14
#>  9: France 2020-08-30                       estimate   1.15 1.15 0.04     1.14
#> 10: France 2020-08-31                       estimate   1.14 1.14 0.04     1.14
#> 11: France 2020-09-01                       estimate   1.14 1.14 0.04     1.14
#> 12: France 2020-09-02                       estimate   1.14 1.14 0.04     1.14
#> 13: France 2020-09-03                       estimate   1.14 1.13 0.04     1.14
#> 14: France 2020-09-04                       estimate   1.13 1.13 0.04     1.14
#> 15: France 2020-09-05                       estimate   1.13 1.13 0.04     1.13
#> 16: France 2020-09-06                       estimate   1.13 1.12 0.04     1.14
#> 17: France 2020-09-07                       estimate   1.13 1.12 0.04     1.14
#> 18: France 2020-09-08                       estimate   1.12 1.12 0.04     1.13
#> 19: France 2020-09-09                       estimate   1.12 1.11 0.04     1.13
#> 20: France 2020-09-10                       estimate   1.12 1.11 0.04     1.13
#> 21: France 2020-09-11                       estimate   1.12 1.11 0.04     1.13
#> 22: France 2020-09-12                       estimate   1.12 1.11 0.05     1.13
#> 23: France 2020-09-13                       estimate   1.11 1.11 0.05     1.12
#> 24: France 2020-09-14                       estimate   1.11 1.10 0.05     1.12
#> 25: France 2020-09-15                       estimate   1.11 1.10 0.05     1.12
#> 26: France 2020-09-16                       estimate   1.11 1.10 0.05     1.12
#> 27: France 2020-09-17                       estimate   1.11 1.10 0.05     1.12
#> 28: France 2020-09-18                       estimate   1.11 1.10 0.05     1.12
#> 29: France 2020-09-19                       estimate   1.11 1.10 0.05     1.12
#> 30: France 2020-09-20                       estimate   1.11 1.10 0.05     1.12
#> 31: France 2020-09-21                       estimate   1.11 1.10 0.05     1.11
#> 32: France 2020-09-22                       estimate   1.11 1.10 0.05     1.11
#> 33: France 2020-09-23                       estimate   1.11 1.10 0.05     1.12
#> 34: France 2020-09-24                       estimate   1.11 1.11 0.05     1.12
#> 35: France 2020-09-25                       estimate   1.11 1.11 0.05     1.13
#> 36: France 2020-09-26                       estimate   1.12 1.11 0.04     1.13
#> 37: France 2020-09-27                       estimate   1.12 1.11 0.04     1.13
#> 38: France 2020-09-28                       estimate   1.12 1.12 0.04     1.13
#> 39: France 2020-09-29                       estimate   1.12 1.12 0.04     1.13
#> 40: France 2020-09-30                       estimate   1.13 1.12 0.04     1.12
#> 41: France 2020-10-01                       estimate   1.13 1.13 0.04     1.12
#> 42: France 2020-10-02                       estimate   1.13 1.13 0.04     1.14
#> 43: France 2020-10-03                       estimate   1.14 1.14 0.04     1.13
#> 44: France 2020-10-04                       estimate   1.14 1.14 0.04     1.13
#> 45: France 2020-10-05                       estimate   1.14 1.14 0.04     1.13
#> 46: France 2020-10-06                       estimate   1.14 1.15 0.05     1.13
#> 47: France 2020-10-07                       estimate   1.15 1.15 0.05     1.14
#> 48: France 2020-10-08                       estimate   1.15 1.16 0.05     1.13
#> 49: France 2020-10-09                       estimate   1.15 1.16 0.05     1.13
#> 50: France 2020-10-10                       estimate   1.15 1.16 0.05     1.13
#> 51: France 2020-10-11                       estimate   1.16 1.16 0.05     1.13
#> 52: France 2020-10-12                       estimate   1.16 1.17 0.06     1.13
#> 53: France 2020-10-13                       estimate   1.16 1.17 0.06     1.13
#> 54: France 2020-10-14                       estimate   1.16 1.17 0.06     1.13
#> 55: France 2020-10-15                       estimate   1.16 1.17 0.06     1.13
#> 56: France 2020-10-16                       estimate   1.16 1.17 0.06     1.13
#> 57: France 2020-10-17                       estimate   1.16 1.17 0.06     1.13
#> 58: France 2020-10-18                       estimate   1.16 1.17 0.07     1.13
#> 59: France 2020-10-19                       estimate   1.16 1.17 0.07     1.13
#> 60: France 2020-10-20                       estimate   1.16 1.17 0.07     1.13
#> 61: France 2020-10-21 estimate based on partial data   1.16 1.17 0.08     1.13
#> 62: France 2020-10-22 estimate based on partial data   1.16 1.17 0.08     1.13
#> 63: France 2020-10-23 estimate based on partial data   1.16 1.17 0.08     1.13
#> 64: France 2020-10-24 estimate based on partial data   1.16 1.17 0.09     1.13
#> 65: France 2020-10-25 estimate based on partial data   1.16 1.17 0.09     1.13
#> 66: France 2020-10-26 estimate based on partial data   1.16 1.17 0.10     1.13
#> 67: France 2020-10-27 estimate based on partial data   1.16 1.17 0.10     1.13
#> 68: France 2020-10-28 estimate based on partial data   1.15 1.17 0.11     1.13
#> 69: France 2020-10-29 estimate based on partial data   1.15 1.17 0.12     1.13
#> 70: France 2020-10-30 estimate based on partial data   1.15 1.17 0.12     1.13
#> 71: France 2020-10-31 estimate based on partial data   1.15 1.17 0.13     1.13
#> 72: France 2020-11-01                       forecast   1.15 1.17 0.13     1.13
#> 73: France 2020-11-02                       forecast   1.15 1.17 0.13     1.13
#> 74: France 2020-11-03                       forecast   1.15 1.17 0.13     1.13
#> 75: France 2020-11-04                       forecast   1.15 1.17 0.13     1.13
#> 76: France 2020-11-05                       forecast   1.15 1.17 0.13     1.13
#> 77: France 2020-11-06                       forecast   1.15 1.17 0.13     1.13
#> 78: France 2020-11-07                       forecast   1.15 1.17 0.13     1.13
#> 79: France 2020-11-08                       forecast   1.15 1.17 0.13     1.13
#> 80: France 2020-11-09                       forecast   1.15 1.17 0.13     1.13
#> 81: France 2020-11-10                       forecast   1.15 1.17 0.13     1.13
#> 82: France 2020-11-11                       forecast   1.15 1.17 0.13     1.13
#> 83: France 2020-11-12                       forecast   1.15 1.17 0.13     1.13
#> 84: France 2020-11-13                       forecast   1.15 1.17 0.13     1.13
#> 85: France 2020-11-14                       forecast   1.15 1.17 0.13     1.13
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.17     1.14     1.18     1.12     1.25
#>  2:     1.17     1.14     1.18     1.12     1.25
#>  3:     1.16     1.14     1.18     1.11     1.23
#>  4:     1.16     1.14     1.18     1.10     1.23
#>  5:     1.16     1.14     1.18     1.10     1.22
#>  6:     1.15     1.13     1.17     1.10     1.23
#>  7:     1.15     1.13     1.17     1.09     1.22
#>  8:     1.16     1.13     1.17     1.08     1.21
#>  9:     1.15     1.13     1.17     1.08     1.21
#> 10:     1.15     1.13     1.17     1.08     1.20
#> 11:     1.15     1.12     1.16     1.07     1.20
#> 12:     1.15     1.12     1.16     1.06     1.20
#> 13:     1.15     1.12     1.16     1.06     1.20
#> 14:     1.15     1.12     1.16     1.05     1.19
#> 15:     1.15     1.11     1.16     1.05     1.19
#> 16:     1.15     1.11     1.16     1.05     1.19
#> 17:     1.15     1.11     1.15     1.04     1.18
#> 18:     1.15     1.11     1.15     1.04     1.18
#> 19:     1.15     1.11     1.16     1.04     1.18
#> 20:     1.15     1.10     1.15     1.04     1.17
#> 21:     1.15     1.10     1.15     1.04     1.17
#> 22:     1.14     1.10     1.15     1.04     1.17
#> 23:     1.14     1.10     1.15     1.03     1.17
#> 24:     1.14     1.10     1.15     1.03     1.17
#> 25:     1.14     1.10     1.15     1.03     1.17
#> 26:     1.14     1.10     1.15     1.02     1.17
#> 27:     1.14     1.09     1.15     1.02     1.17
#> 28:     1.14     1.10     1.15     1.02     1.17
#> 29:     1.14     1.10     1.15     1.02     1.17
#> 30:     1.14     1.10     1.15     1.03     1.17
#> 31:     1.13     1.10     1.15     1.03     1.17
#> 32:     1.13     1.09     1.15     1.03     1.17
#> 33:     1.14     1.10     1.15     1.03     1.17
#> 34:     1.14     1.10     1.15     1.03     1.17
#> 35:     1.14     1.10     1.15     1.03     1.17
#> 36:     1.14     1.11     1.15     1.04     1.18
#> 37:     1.15     1.11     1.15     1.04     1.18
#> 38:     1.15     1.11     1.15     1.05     1.18
#> 39:     1.14     1.11     1.15     1.05     1.18
#> 40:     1.14     1.11     1.15     1.05     1.19
#> 41:     1.14     1.11     1.15     1.06     1.19
#> 42:     1.15     1.11     1.15     1.06     1.20
#> 43:     1.15     1.11     1.16     1.07     1.21
#> 44:     1.15     1.12     1.16     1.07     1.21
#> 45:     1.15     1.12     1.16     1.07     1.21
#> 46:     1.15     1.12     1.16     1.08     1.23
#> 47:     1.15     1.12     1.16     1.08     1.23
#> 48:     1.15     1.12     1.17     1.09     1.24
#> 49:     1.15     1.12     1.17     1.09     1.24
#> 50:     1.15     1.12     1.17     1.08     1.24
#> 51:     1.15     1.12     1.18     1.08     1.24
#> 52:     1.15     1.12     1.18     1.08     1.25
#> 53:     1.15     1.12     1.18     1.08     1.26
#> 54:     1.15     1.12     1.18     1.08     1.26
#> 55:     1.15     1.12     1.19     1.08     1.27
#> 56:     1.15     1.12     1.19     1.08     1.28
#> 57:     1.15     1.12     1.19     1.08     1.28
#> 58:     1.15     1.12     1.19     1.07     1.29
#> 59:     1.16     1.12     1.19     1.07     1.29
#> 60:     1.16     1.11     1.19     1.07     1.31
#> 61:     1.16     1.11     1.19     1.06     1.31
#> 62:     1.16     1.10     1.19     1.06     1.32
#> 63:     1.16     1.11     1.19     1.05     1.32
#> 64:     1.16     1.11     1.19     1.04     1.33
#> 65:     1.16     1.11     1.19     1.04     1.34
#> 66:     1.16     1.10     1.19     1.04     1.35
#> 67:     1.16     1.10     1.19     1.03     1.35
#> 68:     1.16     1.10     1.20     1.02     1.36
#> 69:     1.16     1.11     1.21     1.00     1.36
#> 70:     1.16     1.10     1.21     1.00     1.37
#> 71:     1.16     1.09     1.20     0.99     1.38
#> 72:     1.16     1.09     1.20     0.99     1.38
#> 73:     1.16     1.09     1.20     0.99     1.38
#> 74:     1.16     1.09     1.20     0.99     1.38
#> 75:     1.16     1.09     1.20     0.99     1.38
#> 76:     1.16     1.09     1.20     0.99     1.38
#> 77:     1.16     1.09     1.20     0.99     1.38
#> 78:     1.16     1.09     1.20     0.99     1.38
#> 79:     1.16     1.09     1.20     0.99     1.38
#> 80:     1.16     1.09     1.20     0.99     1.38
#> 81:     1.16     1.09     1.20     0.99     1.38
#> 82:     1.16     1.09     1.20     0.99     1.38
#> 83:     1.16     1.09     1.20     0.99     1.38
#> 84:     1.16     1.09     1.20     0.99     1.38
#> 85:     1.16     1.09     1.20     0.99     1.38
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData
#>     region       date                           type median   mean     sd
#>  1: France 2020-08-22                       estimate   5084   5187    728
#>  2: France 2020-08-23                       estimate   5302   5412    765
#>  3: France 2020-08-24                       estimate   5526   5640    802
#>  4: France 2020-08-25                       estimate   5765   5873    838
#>  5: France 2020-08-26                       estimate   5991   6109    874
#>  6: France 2020-08-27                       estimate   6235   6349    909
#>  7: France 2020-08-28                       estimate   6470   6592    943
#>  8: France 2020-08-29                       estimate   6710   6838    977
#>  9: France 2020-08-30                       estimate   6950   7087   1009
#> 10: France 2020-08-31                       estimate   7201   7337   1041
#> 11: France 2020-09-01                       estimate   7452   7590   1071
#> 12: France 2020-09-02                       estimate   7702   7845   1099
#> 13: France 2020-09-03                       estimate   7951   8101   1126
#> 14: France 2020-09-04                       estimate   8223   8358   1151
#> 15: France 2020-09-05                       estimate   8479   8617   1174
#> 16: France 2020-09-06                       estimate   8746   8877   1196
#> 17: France 2020-09-07                       estimate   9022   9138   1215
#> 18: France 2020-09-08                       estimate   9281   9401   1233
#> 19: France 2020-09-09                       estimate   9543   9665   1251
#> 20: France 2020-09-10                       estimate   9811   9930   1270
#> 21: France 2020-09-11                       estimate  10087  10198   1290
#> 22: France 2020-09-12                       estimate  10345  10470   1312
#> 23: France 2020-09-13                       estimate  10623  10744   1337
#> 24: France 2020-09-14                       estimate  10891  11024   1366
#> 25: France 2020-09-15                       estimate  11199  11309   1399
#> 26: France 2020-09-16                       estimate  11486  11601   1438
#> 27: France 2020-09-17                       estimate  11812  11902   1483
#> 28: France 2020-09-18                       estimate  12147  12212   1534
#> 29: France 2020-09-19                       estimate  12491  12534   1591
#> 30: France 2020-09-20                       estimate  12846  12869   1654
#> 31: France 2020-09-21                       estimate  13208  13220   1724
#> 32: France 2020-09-22                       estimate  13592  13588   1799
#> 33: France 2020-09-23                       estimate  13998  13975   1880
#> 34: France 2020-09-24                       estimate  14403  14384   1964
#> 35: France 2020-09-25                       estimate  14827  14817   2052
#> 36: France 2020-09-26                       estimate  15286  15277   2142
#> 37: France 2020-09-27                       estimate  15806  15765   2234
#> 38: France 2020-09-28                       estimate  16323  16285   2327
#> 39: France 2020-09-29                       estimate  16891  16838   2419
#> 40: France 2020-09-30                       estimate  17461  17428   2511
#> 41: France 2020-10-01                       estimate  18075  18057   2602
#> 42: France 2020-10-02                       estimate  18729  18727   2691
#> 43: France 2020-10-03                       estimate  19428  19441   2779
#> 44: France 2020-10-04                       estimate  20187  20201   2867
#> 45: France 2020-10-05                       estimate  20955  21010   2956
#> 46: France 2020-10-06                       estimate  21788  21870   3048
#> 47: France 2020-10-07                       estimate  22689  22783   3145
#> 48: France 2020-10-08                       estimate  23643  23752   3250
#> 49: France 2020-10-09                       estimate  24610  24778   3367
#> 50: France 2020-10-10                       estimate  25670  25864   3501
#> 51: France 2020-10-11                       estimate  26770  27011   3657
#> 52: France 2020-10-12                       estimate  27954  28221   3841
#> 53: France 2020-10-13                       estimate  29227  29495   4059
#> 54: France 2020-10-14                       estimate  30490  30836   4318
#> 55: France 2020-10-15                       estimate  31841  32244   4626
#> 56: France 2020-10-16                       estimate  33231  33723   4993
#> 57: France 2020-10-17                       estimate  34654  35275   5430
#> 58: France 2020-10-18                       estimate  36201  36904   5951
#> 59: France 2020-10-19                       estimate  37791  38614   6573
#> 60: France 2020-10-20                       estimate  39513  40411   7315
#> 61: France 2020-10-21 estimate based on partial data  41160  42303   8196
#> 62: France 2020-10-22 estimate based on partial data  43015  44297   9238
#> 63: France 2020-10-23 estimate based on partial data  44848  46405  10466
#> 64: France 2020-10-24 estimate based on partial data  46655  48638  11908
#> 65: France 2020-10-25 estimate based on partial data  48507  51011  13598
#> 66: France 2020-10-26 estimate based on partial data  50511  53541  15582
#> 67: France 2020-10-27 estimate based on partial data  52625  56248  17915
#> 68: France 2020-10-28 estimate based on partial data  54898  59157  20671
#> 69: France 2020-10-29 estimate based on partial data  57084  62296  23949
#> 70: France 2020-10-30 estimate based on partial data  59438  65699  27886
#> 71: France 2020-10-31 estimate based on partial data  61735  69410  32674
#> 72: France 2020-11-01                       forecast  64229  73406  38178
#> 73: France 2020-11-02                       forecast  66770  77745  44773
#> 74: France 2020-11-03                       forecast  69489  82467  52725
#> 75: France 2020-11-04                       forecast  72426  87622  62385
#> 76: France 2020-11-05                       forecast  75344  93264  74218
#> 77: France 2020-11-06                       forecast  78478  99461  88832
#> 78: France 2020-11-07                       forecast  81755 106290 107031
#> 79: France 2020-11-08                       forecast  85028 113845 129873
#> 80: France 2020-11-09                       forecast  88413 122240 158751
#> 81: France 2020-11-10                       forecast  92178 131609 195503
#> 82: France 2020-11-11                       forecast  96110 142118 242551
#> 83: France 2020-11-12                       forecast 100022 153969 303091
#> 84: France 2020-11-13                       forecast 104185 167409 381344
#> 85: France 2020-11-14                       forecast 108383 182744 482887
#>     region       date                           type median   mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     4598     4942     4410     5319     4072     6269
#>  2:     4787     5151     4631     5591     4268     6591
#>  3:     5109     5490     4876     5892     4412     6863
#>  4:     5485     5882     5078     6137     4550     7128
#>  5:     5541     5957     5300     6407     4813     7525
#>  6:     5976     6407     5507     6657     5000     7822
#>  7:     5892     6344     5734     6928     5183     8079
#>  8:     6171     6629     5893     7127     5274     8266
#>  9:     6354     6814     6168     7434     5485     8570
#> 10:     6677     7143     6414     7717     5700     8913
#> 11:     7022     7511     6468     7806     5884     9165
#> 12:     7275     7774     6728     8087     6127     9481
#> 13:     7516     8023     7063     8454     6346     9814
#> 14:     7869     8386     7235     8646     6534    10031
#> 15:     8079     8614     7501     8934     6690    10263
#> 16:     8494     9045     7875     9322     6880    10549
#> 17:     8239     8812     8062     9536     7325    11052
#> 18:     8912     9488     8316     9821     7571    11359
#> 19:     9404     9979     8500    10020     7830    11716
#> 20:     9563    10146     8874    10416     7917    11865
#> 21:     9668    10248     9038    10610     8245    12255
#> 22:     9993    10576     9325    10936     8314    12434
#> 23:    10244    10841     9521    11168     8766    13010
#> 24:    10287    10898     9864    11554     8836    13190
#> 25:    10734    11351    10236    11980     8920    13431
#> 26:    11227    11876    10446    12224     9075    13745
#> 27:    11570    12248    10710    12573     9543    14342
#> 28:    11925    12632    10885    12827     9855    14790
#> 29:    12457    13214    11351    13385     9998    15174
#> 30:    12774    13561    11750    13886    10077    15470
#> 31:    13339    14169    12228    14460    10401    16009
#> 32:    13719    14578    12303    14655    10578    16411
#> 33:    14051    14948    12639    15104    10922    16986
#> 34:    14009    14957    13045    15618    11213    17556
#> 35:    14508    15500    13435    16150    11411    18032
#> 36:    15021    16068    13736    16573    11783    18715
#> 37:    15535    16613    14487    17463    12129    19347
#> 38:    16166    17305    14963    18045    12437    19973
#> 39:    16656    17841    15550    18759    12907    20752
#> 40:    17190    18422    16155    19517    13563    21727
#> 41:    17839    19117    16466    19923    13844    22313
#> 42:    18300    19617    16953    20523    14465    23200
#> 43:    19043    20414    17440    21132    14802    23896
#> 44:    19400    20803    18211    21993    15423    24824
#> 45:    20536    21959    18882    22761    16361    26044
#> 46:    21618    23070    19437    23429    17132    27092
#> 47:    22165    23661    20263    24325    17887    28137
#> 48:    23083    24641    21141    25334    18424    29001
#> 49:    23952    25535    21951    26283    19203    30142
#> 50:    24688    26340    23297    27775    20336    31623
#> 51:    25607    27319    24189    28861    21281    33043
#> 52:    26603    28380    25728    30631    21679    33912
#> 53:    27875    29763    26160    31335    22792    35629
#> 54:    29188    31200    27438    32939    23991    37537
#> 55:    29812    31948    28708    34538    25204    39761
#> 56:    30890    33160    30148    36397    25530    41213
#> 57:    32269    34653    31639    38391    26523    43628
#> 58:    33704    36303    32233    39405    27469    45938
#> 59:    35003    37839    33193    40985    28198    48667
#> 60:    36100    39224    34234    42839    29198    51942
#> 61:    37790    41164    35807    45254    29672    55056
#> 62:    39657    43460    36677    47049    29563    58033
#> 63:    40910    45116    37648    49055    29337    61420
#> 64:    39437    44071    38521    51036    30147    66323
#> 65:    40338    45351    39177    53052    29760    70497
#> 66:    41532    47049    40721    55952    30717    76554
#> 67:    42705    48695    41265    58244    30891    82507
#> 68:    44198    50807    43048    61894    30594    88692
#> 69:    45609    52783    43635    64250    30392    95876
#> 70:    46596    54448    44981    67526    29379   102748
#> 71:    48850    57469    46778    71567    29379   111441
#> 72:    52004    61441    46292    73434    29352   120981
#> 73:    56942    67168    47531    77070    28494   130177
#> 74:    58105    69271    47793    79942    26532   139341
#> 75:    59234    71231    49793    84617    23176   147971
#> 76:    60914    73824    50226    88001    23238   161468
#> 77:    62490    76609    52045    92864    22224   175233
#> 78:    64869    80103    52140    96307    20832   189208
#> 79:    67133    83412    53345   101213    21969   206426
#> 80:    69026    86392    55593   107106    19577   221034
#> 81:    71896    90621    56716   112100    18671   238179
#> 82:    74024    94005    57748   117415    17885   257775
#> 83:    76188    97600    52185   115737    16439   278195
#> 84:    76922    99975    53045   120994    15452   300629
#> 85:    78619   103205    53849   125886     5614   314163
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesReportData
#>     region       date                           type median   mean     sd
#>  1: France 2020-08-22                       estimate   3700   4310   2899
#>  2: France 2020-08-23                       estimate   3490   4123   2720
#>  3: France 2020-08-24                       estimate   3138   3701   2439
#>  4: France 2020-08-25                       estimate   2278   2696   1821
#>  5: France 2020-08-26                       estimate   2818   3293   2208
#>  6: France 2020-08-27                       estimate   4122   4800   3173
#>  7: France 2020-08-28                       estimate   4855   5654   3759
#>  8: France 2020-08-29                       estimate   5079   6030   4058
#>  9: France 2020-08-30                       estimate   4808   5626   3765
#> 10: France 2020-08-31                       estimate   4335   5094   3427
#> 11: France 2020-09-01                       estimate   3112   3731   2594
#> 12: France 2020-09-02                       estimate   3753   4372   2944
#> 13: France 2020-09-03                       estimate   5314   6374   4359
#> 14: France 2020-09-04                       estimate   6451   7515   5038
#> 15: France 2020-09-05                       estimate   6906   8040   5392
#> 16: France 2020-09-06                       estimate   6292   7406   4926
#> 17: France 2020-09-07                       estimate   5791   6748   4544
#> 18: France 2020-09-08                       estimate   4059   4824   3237
#> 19: France 2020-09-09                       estimate   4886   5806   3929
#> 20: France 2020-09-10                       estimate   6994   8257   5639
#> 21: France 2020-09-11                       estimate   8076   9485   6432
#> 22: France 2020-09-12                       estimate   8741  10321   7203
#> 23: France 2020-09-13                       estimate   7756   9251   6368
#> 24: France 2020-09-14                       estimate   6986   8314   5616
#> 25: France 2020-09-15                       estimate   5096   5977   4047
#> 26: France 2020-09-16                       estimate   5977   7022   4863
#> 27: France 2020-09-17                       estimate   8688  10162   6906
#> 28: France 2020-09-18                       estimate   9826  11650   7822
#> 29: France 2020-09-19                       estimate  10702  12612   8699
#> 30: France 2020-09-20                       estimate   9576  11333   7608
#> 31: France 2020-09-21                       estimate   8524  10009   6892
#> 32: France 2020-09-22                       estimate   6103   7250   4838
#> 33: France 2020-09-23                       estimate   7379   8589   5695
#> 34: France 2020-09-24                       estimate  10376  12305   8137
#> 35: France 2020-09-25                       estimate  12040  14007   9269
#> 36: France 2020-09-26                       estimate  12516  14767  10040
#> 37: France 2020-09-27                       estimate  11674  13590   9258
#> 38: France 2020-09-28                       estimate  10543  12195   8318
#> 39: France 2020-09-29                       estimate   7544   8977   6217
#> 40: France 2020-09-30                       estimate   8872  10481   7247
#> 41: France 2020-10-01                       estimate  12417  14900  10213
#> 42: France 2020-10-02                       estimate  13998  16662  11494
#> 43: France 2020-10-03                       estimate  15634  18368  12399
#> 44: France 2020-10-04                       estimate  14316  16631  10895
#> 45: France 2020-10-05                       estimate  12535  14866   9851
#> 46: France 2020-10-06                       estimate   9103  10834   7342
#> 47: France 2020-10-07                       estimate  10824  12834   8718
#> 48: France 2020-10-08                       estimate  15967  18891  13029
#> 49: France 2020-10-09                       estimate  18004  21185  14464
#> 50: France 2020-10-10                       estimate  19668  23285  15606
#> 51: France 2020-10-11                       estimate  18790  22103  14853
#> 52: France 2020-10-12                       estimate  16269  19322  13300
#> 53: France 2020-10-13                       estimate  12081  14287   9953
#> 54: France 2020-10-14                       estimate  14404  16868  11289
#> 55: France 2020-10-15                       estimate  20282  23966  16227
#> 56: France 2020-10-16                       estimate  23718  27901  18675
#> 57: France 2020-10-17                       estimate  26490  30968  20755
#> 58: France 2020-10-18                       estimate  24280  28419  18888
#> 59: France 2020-10-19                       estimate  22498  26091  17326
#> 60: France 2020-10-20                       estimate  15734  18771  12884
#> 61: France 2020-10-21 estimate based on partial data  19278  22691  15062
#> 62: France 2020-10-22 estimate based on partial data  28288  33137  22251
#> 63: France 2020-10-23 estimate based on partial data  33231  38647  25194
#> 64: France 2020-10-24 estimate based on partial data  34616  41890  29161
#> 65: France 2020-10-25 estimate based on partial data  33474  39442  26446
#> 66: France 2020-10-26 estimate based on partial data  29918  35187  23600
#> 67: France 2020-10-27 estimate based on partial data  21342  25811  18213
#> 68: France 2020-10-28 estimate based on partial data  26402  31241  21170
#> 69: France 2020-10-29 estimate based on partial data  37893  45442  33178
#> 70: France 2020-10-30 estimate based on partial data  43690  52362  37292
#> 71: France 2020-10-31 estimate based on partial data  47478  57686  42365
#> 72: France 2020-11-01                       forecast  44580  54347  40152
#> 73: France 2020-11-02                       forecast  39740  48728  37203
#> 74: France 2020-11-03                       forecast  28772  35697  28326
#> 75: France 2020-11-04                       forecast  35310  43030  33660
#> 76: France 2020-11-05                       forecast  51594  65855  53733
#> 77: France 2020-11-06                       forecast  58920  74242  63333
#> 78: France 2020-11-07                       forecast  64449  82084  67479
#> 79: France 2020-11-08                       forecast  60318  79557  73989
#> 80: France 2020-11-09                       forecast  53670  71693  72115
#> 81: France 2020-11-10                       forecast  38696  52566  51699
#> 82: France 2020-11-11                       forecast  47899  66415  86997
#> 83: France 2020-11-12                       forecast  68159  95675 112405
#> 84: France 2020-11-13                       forecast  80571 113929 142328
#> 85: France 2020-11-14                       forecast  85806 126540 180128
#>     region       date                           type median   mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1976     3065     1036     4023      565     8344
#>  2:     1987     3026     1207     4086      329     8059
#>  3:     1881     2791     1002     3563      320     7168
#>  4:     1072     1768      811     2678      248     5270
#>  5:     1082     1937     1075     3338      239     6127
#>  6:     2586     3835     1479     4818      311     9175
#>  7:     2969     4394     1589     5513      581    10896
#>  8:     2524     4005     1988     6105      342    11344
#>  9:     3081     4521     2055     5956      289    10545
#> 10:     2759     4058     1476     5046      406     9714
#> 11:     1712     2599     1188     3729      198     7099
#> 12:     2278     3440     1087     4201      291     8396
#> 13:     3142     4758     2127     6472      415    12158
#> 14:     3511     5430     2336     7444      598    14359
#> 15:     2724     4843     2556     8088     1091    15655
#> 16:     3417     5294     2200     7463      590    14013
#> 17:     4087     5906     1919     6660      394    12703
#> 18:     2443     3633     1385     4773      462     9326
#> 19:     2245     3699     2007     5991      546    11339
#> 20:     2843     4945     2402     8016      479    15864
#> 21:     5784     8276     2839     9352      925    18513
#> 22:     3757     6454     2927    10077      574    19664
#> 23:     4572     6844     2826     9180      838    17707
#> 24:     4463     6518     2590     8198      545    15854
#> 25:     2777     4270     1934     6130      464    11541
#> 26:     3273     5053     2394     7278      523    13439
#> 27:     5603     8246     2811     9825      587    19059
#> 28:     4668     7537     3568    11546     1431    22786
#> 29:     6115     9261     3321    12107     1008    24259
#> 30:     5887     8732     2881    10697     1106    21817
#> 31:     3688     6296     3685    10715      624    19074
#> 32:     4089     5944     2093     7142      513    13726
#> 33:     3998     6154     2473     8370      786    16373
#> 34:     5031     8156     3179    11612     1565    24056
#> 35:     6721    10271     3980    13881     1224    26753
#> 36:     7716    11486     4282    14588     1439    28505
#> 37:     6594    10008     4624    13883      971    25412
#> 38:     6177     9417     3729    12356     1149    23194
#> 39:     4868     7079     2789     8899      739    17194
#> 40:     4861     7583     2848    10116      770    20042
#> 41:     8600    12248     5176    15249     1372    28920
#> 42:     8391    12716     4393    16006     1522    32279
#> 43:     6912    11478     6233    19045     1783    35137
#> 44:     7359    11678     5875    17286     1564    31791
#> 45:     8088    11777     3722    13893     2013    29276
#> 46:     5266     7873     3661    10968      973    20843
#> 47:     6071     9334     3095    12002     1368    25110
#> 48:     8096    12919     6116    19288     1574    36921
#> 49:     9566    14965     6724    21651     1718    40614
#> 50:    11403    17089     7193    23289     2018    44087
#> 51:     9194    14697     7439    22367     1655    42569
#> 52:    10737    15704     5149    18686     1810    38255
#> 53:     5573     9184     4048    13921     1255    27303
#> 54:     7717    11919     5147    16571     1118    31887
#> 55:    10769    16925     6634    22905     1412    45294
#> 56:    14535    21104     9179    28359     2242    53434
#> 57:    15889    23507    10688    32211     3877    59765
#> 58:    14407    21611     8150    27974     2523    54893
#> 59:    12550    19338     9438    27795     2371    49844
#> 60:     6703    11390     6429    19196     1721    36169
#> 61:    11117    16505     5934    21675     1664    42860
#> 62:    14147    22753     8856    32223     1561    62498
#> 63:    19132    29182    10417    37150     3858    73199
#> 64:    17117    27432    12084    40148     3728    80583
#> 65:    19338    29237    10797    37492     2381    74106
#> 66:    16565    25266    12161    37116     2302    67001
#> 67:     9578    15893     7335    24989     1608    49790
#> 68:    16424    24307     8795    30440     3369    59959
#> 69:    25712    37416    11899    43625     2856    87240
#> 70:    17876    31274    11192    47635     3908   101225
#> 71:    28444    43073    15221    53920     3375   110179
#> 72:    19359    33382    11910    49667     4392   107885
#> 73:    19661    31445    10295    43809     3301    95507
#> 74:    12937    21811     6801    31364     2693    70458
#> 75:    16891    27403    11630    41382     2566    84448
#> 76:    21476    37061    14006    58173     4599   131502
#> 77:    29949    48448    14172    63597     2322   142918
#> 78:    41044    60877    17346    72490     3776   160560
#> 79:    23117    41623    16613    67967     5400   157256
#> 80:    21947    38848     8727    56636     1445   140014
#> 81:    17258    29197     8252    42002     2035   106446
#> 82:    15742    30862    11423    53136     2573   132903
#> 83:    22359    43716    17331    76576     2963   189944
#> 84:    30875    56981    14803    87820     3165   224302
#> 85:    28169    55025    21165    97679     2085   249562
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-08-08    2288
#>  2: France 2020-08-09       0
#>  3: France 2020-08-10       0
#>  4: France 2020-08-11    4854
#>  5: France 2020-08-12    1397
#>  6: France 2020-08-13    2524
#>  7: France 2020-08-14    2669
#>  8: France 2020-08-15    2846
#>  9: France 2020-08-16    3310
#> 10: France 2020-08-17    3015
#> 11: France 2020-08-18     493
#> 12: France 2020-08-19    2238
#> 13: France 2020-08-20    3776
#> 14: France 2020-08-21    4771
#> 15: France 2020-08-22    4586
#> 16: France 2020-08-23    3602
#> 17: France 2020-08-24    4897
#> 18: France 2020-08-25    1955
#> 19: France 2020-08-26    3304
#> 20: France 2020-08-27    5429
#> 21: France 2020-08-28    6111
#> 22: France 2020-08-29    7379
#> 23: France 2020-08-30    5453
#> 24: France 2020-08-31    5413
#> 25: France 2020-09-01    3082
#> 26: France 2020-09-02    4982
#> 27: France 2020-09-03    7017
#> 28: France 2020-09-04    7157
#> 29: France 2020-09-05    8975
#> 30: France 2020-09-06    8550
#> 31: France 2020-09-07    7071
#> 32: France 2020-09-08    4203
#> 33: France 2020-09-09    6544
#> 34: France 2020-09-10    8577
#> 35: France 2020-09-11    9843
#> 36: France 2020-09-12    9406
#> 37: France 2020-09-13   10561
#> 38: France 2020-09-14    7183
#> 39: France 2020-09-15    6158
#> 40: France 2020-09-16    7852
#> 41: France 2020-09-17    9784
#> 42: France 2020-09-18   10593
#> 43: France 2020-09-19   13215
#> 44: France 2020-09-20   13498
#> 45: France 2020-09-21   11569
#> 46: France 2020-09-22    4298
#> 47: France 2020-09-23   10008
#> 48: France 2020-09-24   13072
#> 49: France 2020-09-25   16096
#> 50: France 2020-09-26   15797
#> 51: France 2020-09-27   14412
#> 52: France 2020-09-28   11123
#> 53: France 2020-09-29    4070
#> 54: France 2020-09-30    8051
#> 55: France 2020-10-01   12845
#> 56: France 2020-10-02   13970
#> 57: France 2020-10-03   12148
#> 58: France 2020-10-04   16972
#> 59: France 2020-10-05   12545
#> 60: France 2020-10-06    5104
#> 61: France 2020-10-07   10489
#> 62: France 2020-10-08   18746
#> 63: France 2020-10-09   18129
#> 64: France 2020-10-10   20339
#> 65: France 2020-10-11   26896
#> 66: France 2020-10-12   16101
#> 67: France 2020-10-13    8505
#> 68: France 2020-10-14   12993
#> 69: France 2020-10-15   22591
#> 70: France 2020-10-16   30621
#> 71: France 2020-10-17   25086
#> 72: France 2020-10-18   32427
#> 73: France 2020-10-19   29837
#> 74: France 2020-10-20   13243
#> 75: France 2020-10-21   20468
#> 76: France 2020-10-22   26676
#> 77: France 2020-10-23   41622
#> 78: France 2020-10-24   42032
#> 79: France 2020-10-25   45422
#> 80: France 2020-10-26   52010
#> 81: France 2020-10-27   26771
#> 82: France 2020-10-28   33417
#> 83: France 2020-10-29   36437
#> 84: France 2020-10-30   47637
#> 85: France 2020-10-31   49215
#>     region       date confirm
#>