Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

rtData

#> $summaryData
#>           region New confirmed cases by infection date
#>   1: Afghanistan                         49 (6 -- 150)
#>   2:     Albania                      243 (122 -- 395)
#>   3:     Algeria                      248 (132 -- 382)
#>   4:     Andorra                          31 (0 -- 90)
#>   5:      Angola                       117 (27 -- 234)
#>  ---                                                  
#> 176:   Venezuela                    1855 (294 -- 3960)
#> 177:     Vietnam                            0 (0 -- 1)
#> 178:       Yemen                            3 (0 -- 6)
#> 179:      Zambia                        105 (0 -- 473)
#> 180:    Zimbabwe                          28 (0 -- 79)
#>      Expected change in daily cases Effective reproduction no.
#>   1:                         Unsure         0.95 (0.73 -- 1.4)
#>   2:                         Unsure        1.05 (0.83 -- 1.28)
#>   3:                         Unsure         0.9 (0.73 -- 1.14)
#>   4:                         Unsure         1.13 (0.7 -- 1.49)
#>   5:                         Unsure         1.06 (0.8 -- 1.35)
#>  ---                                                          
#> 176:                         Unsure        1.03 (0.79 -- 1.28)
#> 177:              Likely decreasing        0.47 (0.01 -- 1.22)
#> 178:              Likely decreasing        0.85 (0.39 -- 1.16)
#> 179:                         Unsure        0.87 (0.44 -- 1.45)
#> 180:              Likely decreasing        0.73 (0.35 -- 1.11)
#>             Rate of growth Doubling/halving time (days)
#>   1:  -0.01 (-0.09 -- 0.1)          -49.4 (7.2 -- -8.1)
#>   2:  0.01 (-0.05 -- 0.07)          50.3 (9.5 -- -13.5)
#>   3: -0.03 (-0.08 -- 0.04)         -25.7 (18.3 -- -8.9)
#>   4:  0.04 (-0.09 -- 0.13)             19.6 (5.5 -- -8)
#>   5:  0.02 (-0.05 -- 0.09)          44.3 (7.5 -- -12.8)
#>  ---                                                   
#> 176:  0.01 (-0.06 -- 0.07)          75.7 (9.3 -- -11.3)
#> 177: -0.16 (-0.36 -- 0.06)          -4.3 (10.9 -- -1.9)
#> 178: -0.04 (-0.17 -- 0.07)         -16.2 (10.7 -- -4.2)
#> 179: -0.04 (-0.17 -- 0.12)            -19 (5.8 -- -4.1)
#> 180:  -0.08 (-0.2 -- 0.03)              -9 (24 -- -3.4)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-07-09 estimate   0.90 0.89 0.06     0.91     0.92
#>     2: Afghanistan 2020-07-10 estimate   0.90 0.88 0.06     0.90     0.92
#>     3: Afghanistan 2020-07-11 estimate   0.89 0.88 0.07     0.90     0.92
#>     4: Afghanistan 2020-07-12 estimate   0.89 0.88 0.07     0.90     0.92
#>     5: Afghanistan 2020-07-13 estimate   0.89 0.87 0.07     0.90     0.92
#>    ---                                                                   
#> 15109:    Zimbabwe 2020-09-27 forecast   0.73 0.74 0.24     0.64     0.74
#> 15110:    Zimbabwe 2020-09-28 forecast   0.73 0.74 0.24     0.64     0.74
#> 15111:    Zimbabwe 2020-09-29 forecast   0.73 0.74 0.24     0.64     0.74
#> 15112:    Zimbabwe 2020-09-30 forecast   0.73 0.74 0.24     0.64     0.74
#> 15113:    Zimbabwe 2020-10-01 forecast   0.73 0.74 0.24     0.64     0.74
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.88     0.93     0.80     0.98
#>     2:     0.88     0.93     0.79     0.97
#>     3:     0.88     0.93     0.78     0.97
#>     4:     0.88     0.93     0.76     0.96
#>     5:     0.87     0.93     0.75     0.97
#>    ---                                    
#> 15109:     0.58     0.88     0.35     1.11
#> 15110:     0.58     0.88     0.35     1.11
#> 15111:     0.58     0.88     0.35     1.11
#> 15112:     0.58     0.88     0.35     1.11
#> 15113:     0.58     0.88     0.35     1.11
#> 
#> $casesInfectionData

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

#>             region       date     type median          mean            sd
#>     1: Afghanistan 2020-07-09 estimate    192 1.007894e-321 2.470328e-322
#>     2: Afghanistan 2020-07-10 estimate    186 9.733093e-322 2.322109e-322
#>     3: Afghanistan 2020-07-11 estimate    180 9.337841e-322 2.173889e-322
#>     4: Afghanistan 2020-07-12 estimate    175 8.991995e-322 2.075076e-322
#>     5: Afghanistan 2020-07-13 estimate    169 8.646149e-322 1.976263e-322
#>    ---                                                                   
#> 15109:    Zimbabwe 2020-09-27 forecast     13 2.371515e-322 1.837924e-321
#> 15110:    Zimbabwe 2020-09-28 forecast     12 2.667954e-322 2.613607e-321
#> 15111:    Zimbabwe 2020-09-29 forecast     11 3.013800e-322 3.745018e-321
#> 15112:    Zimbabwe 2020-09-30 forecast     11 3.557273e-322 5.390256e-321
#> 15113:    Zimbabwe 2020-10-01 forecast     10 4.298371e-322 7.781534e-321
#>        lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>     1:      179      193      161      203      136      285
#>     2:      172      186      156      198      130      269
#>     3:      164      178      156      196      127      259
#>     4:      163      177      151      190      121      247
#>     5:      159      172      147      184      116      236
#>    ---                                                      
#> 15109:        0        3        0       13        0       81
#> 15110:        0        3        0       12        0       81
#> 15111:        0        2        0       12        0       80
#> 15112:        0        2        0       11        0       80
#> 15113:        0        2        0       10        0       81
#> 
#> $casesReportData
#>             region       date     type median mean  sd lower_20 upper_20
#>     1: Afghanistan 2020-07-09 estimate    108  199 271        0       26
#>     2: Afghanistan 2020-07-10 estimate    106  180 225        0       23
#>     3: Afghanistan 2020-07-11 estimate     99  179 231        0       23
#>     4: Afghanistan 2020-07-12 estimate     76  141 198        0       18
#>     5: Afghanistan 2020-07-13 estimate     78  136 164        0       19
#>    ---                                                                  
#> 15109:    Zimbabwe 2020-09-27 forecast     17   34  70        1        6
#> 15110:    Zimbabwe 2020-09-28 forecast     11   24  67        0        4
#> 15111:    Zimbabwe 2020-09-29 forecast     16   38 131        0        5
#> 15112:    Zimbabwe 2020-09-30 forecast     12   41 518        0        4
#> 15113:    Zimbabwe 2020-10-01 forecast     13   36 191        0        4
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0      108        0      511
#>     2:        0      106        0      443
#>     3:        0       99        0      460
#>     4:        0       76        0      348
#>     5:        0       78        0      330
#>    ---                                    
#> 15109:        0       17        0       77
#> 15110:        0       11        0       53
#> 15111:        0       16        0       78
#> 15112:        0       12        0       63
#> 15113:        0       13        0       72
#> 
#> $obsCasesData
#>             region       date confirm
#>     1: Afghanistan 2020-06-25     234
#>     2:     Albania 2020-06-25      67
#>     3:     Algeria 2020-06-25     172
#>     4:     Andorra 2020-06-25       0
#>     5:      Angola 2020-06-25       3
#>    ---                               
#> 15636:   Venezuela 2020-09-17     761
#> 15637:     Vietnam 2020-09-17       0
#> 15638:       Yemen 2020-09-17       4
#> 15639:      Zambia 2020-09-17      68
#> 15640:    Zimbabwe 2020-09-17      22
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France                  32443 (708 -- 98691)              Likely increasing
#>    Effective reproduction no.       Rate of growth Doubling/halving time (days)
#> 1:        1.18 (0.82 -- 1.54) 0.05 (-0.05 -- 0.14)          14.6 (4.9 -- -14.3)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-07-09                       estimate   1.14 1.13 0.08     1.15
#>  2: France 2020-07-10                       estimate   1.15 1.13 0.08     1.16
#>  3: France 2020-07-11                       estimate   1.15 1.13 0.08     1.15
#>  4: France 2020-07-12                       estimate   1.15 1.14 0.08     1.16
#>  5: France 2020-07-13                       estimate   1.15 1.14 0.08     1.15
#>  6: France 2020-07-14                       estimate   1.16 1.15 0.08     1.16
#>  7: France 2020-07-15                       estimate   1.16 1.15 0.08     1.16
#>  8: France 2020-07-16                       estimate   1.16 1.15 0.08     1.16
#>  9: France 2020-07-17                       estimate   1.16 1.16 0.08     1.16
#> 10: France 2020-07-18                       estimate   1.17 1.16 0.08     1.16
#> 11: France 2020-07-19                       estimate   1.17 1.17 0.08     1.16
#> 12: France 2020-07-20                       estimate   1.17 1.17 0.08     1.16
#> 13: France 2020-07-21                       estimate   1.17 1.18 0.08     1.15
#> 14: France 2020-07-22                       estimate   1.18 1.18 0.08     1.16
#> 15: France 2020-07-23                       estimate   1.18 1.19 0.08     1.16
#> 16: France 2020-07-24                       estimate   1.19 1.19 0.08     1.16
#> 17: France 2020-07-25                       estimate   1.19 1.20 0.08     1.16
#> 18: France 2020-07-26                       estimate   1.19 1.20 0.08     1.16
#> 19: France 2020-07-27                       estimate   1.20 1.21 0.09     1.17
#> 20: France 2020-07-28                       estimate   1.20 1.21 0.09     1.16
#> 21: France 2020-07-29                       estimate   1.20 1.22 0.09     1.16
#> 22: France 2020-07-30                       estimate   1.20 1.22 0.10     1.16
#> 23: France 2020-07-31                       estimate   1.21 1.23 0.10     1.16
#> 24: France 2020-08-01                       estimate   1.21 1.23 0.10     1.17
#> 25: France 2020-08-02                       estimate   1.21 1.24 0.10     1.17
#> 26: France 2020-08-03                       estimate   1.22 1.24 0.10     1.17
#> 27: France 2020-08-04                       estimate   1.22 1.24 0.11     1.17
#> 28: France 2020-08-05                       estimate   1.22 1.25 0.11     1.17
#> 29: France 2020-08-06                       estimate   1.22 1.25 0.11     1.17
#> 30: France 2020-08-07                       estimate   1.22 1.25 0.11     1.18
#> 31: France 2020-08-08                       estimate   1.23 1.25 0.11     1.18
#> 32: France 2020-08-09                       estimate   1.23 1.25 0.11     1.18
#> 33: France 2020-08-10                       estimate   1.23 1.25 0.11     1.17
#> 34: France 2020-08-11                       estimate   1.23 1.25 0.11     1.17
#> 35: France 2020-08-12                       estimate   1.23 1.25 0.11     1.17
#> 36: France 2020-08-13                       estimate   1.23 1.25 0.10     1.17
#> 37: France 2020-08-14                       estimate   1.23 1.25 0.10     1.18
#> 38: France 2020-08-15                       estimate   1.22 1.25 0.10     1.18
#> 39: France 2020-08-16                       estimate   1.22 1.24 0.10     1.18
#> 40: France 2020-08-17                       estimate   1.22 1.24 0.10     1.16
#> 41: France 2020-08-18                       estimate   1.22 1.24 0.10     1.17
#> 42: France 2020-08-19                       estimate   1.22 1.23 0.10     1.16
#> 43: France 2020-08-20                       estimate   1.21 1.23 0.10     1.18
#> 44: France 2020-08-21                       estimate   1.21 1.23 0.10     1.17
#> 45: France 2020-08-22                       estimate   1.21 1.22 0.10     1.17
#> 46: France 2020-08-23                       estimate   1.21 1.22 0.10     1.16
#> 47: France 2020-08-24                       estimate   1.20 1.21 0.10     1.17
#> 48: France 2020-08-25                       estimate   1.20 1.21 0.10     1.17
#> 49: France 2020-08-26                       estimate   1.20 1.21 0.10     1.17
#> 50: France 2020-08-27                       estimate   1.19 1.20 0.10     1.16
#> 51: France 2020-08-28                       estimate   1.19 1.20 0.10     1.16
#> 52: France 2020-08-29                       estimate   1.19 1.20 0.11     1.16
#> 53: France 2020-08-30                       estimate   1.19 1.19 0.11     1.16
#> 54: France 2020-08-31                       estimate   1.19 1.19 0.11     1.16
#> 55: France 2020-09-01                       estimate   1.19 1.19 0.12     1.16
#> 56: France 2020-09-02                       estimate   1.18 1.18 0.12     1.16
#> 57: France 2020-09-03                       estimate   1.18 1.18 0.12     1.16
#> 58: France 2020-09-04                       estimate   1.18 1.18 0.13     1.15
#> 59: France 2020-09-05 estimate based on partial data   1.18 1.18 0.14     1.16
#> 60: France 2020-09-06 estimate based on partial data   1.18 1.18 0.14     1.15
#> 61: France 2020-09-07 estimate based on partial data   1.18 1.18 0.15     1.15
#> 62: France 2020-09-08 estimate based on partial data   1.18 1.17 0.16     1.15
#> 63: France 2020-09-09 estimate based on partial data   1.18 1.17 0.16     1.15
#> 64: France 2020-09-10 estimate based on partial data   1.18 1.17 0.17     1.15
#> 65: France 2020-09-11 estimate based on partial data   1.18 1.17 0.18     1.15
#> 66: France 2020-09-12 estimate based on partial data   1.18 1.17 0.18     1.16
#> 67: France 2020-09-13 estimate based on partial data   1.18 1.17 0.19     1.17
#> 68: France 2020-09-14 estimate based on partial data   1.18 1.17 0.20     1.16
#> 69: France 2020-09-15 estimate based on partial data   1.18 1.17 0.21     1.16
#> 70: France 2020-09-16 estimate based on partial data   1.18 1.18 0.21     1.16
#> 71: France 2020-09-17 estimate based on partial data   1.18 1.18 0.22     1.15
#> 72: France 2020-09-18                       forecast   1.18 1.18 0.22     1.15
#> 73: France 2020-09-19                       forecast   1.18 1.18 0.22     1.15
#> 74: France 2020-09-20                       forecast   1.18 1.18 0.22     1.15
#> 75: France 2020-09-21                       forecast   1.18 1.18 0.22     1.15
#> 76: France 2020-09-22                       forecast   1.18 1.18 0.22     1.15
#> 77: France 2020-09-23                       forecast   1.18 1.18 0.22     1.15
#> 78: France 2020-09-24                       forecast   1.18 1.18 0.22     1.15
#> 79: France 2020-09-25                       forecast   1.18 1.18 0.22     1.15
#> 80: France 2020-09-26                       forecast   1.18 1.18 0.22     1.15
#> 81: France 2020-09-27                       forecast   1.18 1.18 0.22     1.15
#> 82: France 2020-09-28                       forecast   1.18 1.18 0.22     1.15
#> 83: France 2020-09-29                       forecast   1.18 1.18 0.22     1.15
#> 84: France 2020-09-30                       forecast   1.18 1.18 0.22     1.15
#> 85: France 2020-10-01                       forecast   1.18 1.18 0.22     1.15
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.17     1.12     1.20     1.00     1.23
#>  2:     1.18     1.12     1.19     1.00     1.23
#>  3:     1.18     1.13     1.20     1.01     1.24
#>  4:     1.19     1.13     1.20     1.01     1.25
#>  5:     1.18     1.13     1.20     1.00     1.25
#>  6:     1.18     1.13     1.20     1.00     1.25
#>  7:     1.18     1.13     1.20     1.01     1.26
#>  8:     1.18     1.14     1.20     1.02     1.27
#>  9:     1.18     1.14     1.20     1.03     1.28
#> 10:     1.18     1.14     1.21     1.04     1.30
#> 11:     1.18     1.14     1.21     1.04     1.30
#> 12:     1.18     1.14     1.21     1.05     1.31
#> 13:     1.18     1.14     1.21     1.05     1.31
#> 14:     1.18     1.14     1.21     1.06     1.32
#> 15:     1.18     1.14     1.21     1.05     1.32
#> 16:     1.18     1.14     1.21     1.06     1.32
#> 17:     1.19     1.15     1.22     1.06     1.33
#> 18:     1.19     1.14     1.22     1.07     1.34
#> 19:     1.19     1.15     1.22     1.09     1.37
#> 20:     1.19     1.14     1.22     1.09     1.38
#> 21:     1.19     1.14     1.23     1.08     1.37
#> 22:     1.19     1.15     1.23     1.09     1.39
#> 23:     1.19     1.14     1.23     1.10     1.40
#> 24:     1.20     1.15     1.24     1.11     1.41
#> 25:     1.20     1.15     1.24     1.11     1.42
#> 26:     1.20     1.14     1.23     1.11     1.43
#> 27:     1.20     1.15     1.24     1.11     1.43
#> 28:     1.20     1.15     1.25     1.10     1.42
#> 29:     1.20     1.15     1.25     1.10     1.42
#> 30:     1.21     1.14     1.24     1.11     1.42
#> 31:     1.21     1.15     1.25     1.11     1.43
#> 32:     1.21     1.15     1.25     1.11     1.43
#> 33:     1.21     1.16     1.25     1.12     1.43
#> 34:     1.20     1.15     1.25     1.12     1.43
#> 35:     1.20     1.16     1.26     1.11     1.42
#> 36:     1.21     1.15     1.25     1.12     1.42
#> 37:     1.21     1.16     1.26     1.11     1.41
#> 38:     1.21     1.15     1.25     1.10     1.40
#> 39:     1.21     1.15     1.25     1.10     1.39
#> 40:     1.20     1.16     1.26     1.09     1.39
#> 41:     1.20     1.15     1.25     1.11     1.40
#> 42:     1.19     1.15     1.24     1.11     1.40
#> 43:     1.21     1.15     1.25     1.09     1.39
#> 44:     1.21     1.14     1.24     1.08     1.39
#> 45:     1.21     1.14     1.24     1.07     1.37
#> 46:     1.20     1.14     1.24     1.07     1.38
#> 47:     1.20     1.15     1.25     1.06     1.38
#> 48:     1.20     1.14     1.24     1.07     1.39
#> 49:     1.21     1.13     1.23     1.06     1.38
#> 50:     1.19     1.14     1.24     1.04     1.38
#> 51:     1.19     1.13     1.23     1.04     1.38
#> 52:     1.20     1.13     1.24     1.03     1.38
#> 53:     1.19     1.13     1.24     1.02     1.38
#> 54:     1.20     1.13     1.24     1.02     1.39
#> 55:     1.20     1.13     1.24     0.99     1.38
#> 56:     1.20     1.12     1.24     1.01     1.40
#> 57:     1.20     1.13     1.26     0.97     1.38
#> 58:     1.19     1.13     1.26     0.96     1.39
#> 59:     1.21     1.12     1.26     0.95     1.40
#> 60:     1.19     1.13     1.26     0.95     1.42
#> 61:     1.20     1.13     1.27     0.94     1.44
#> 62:     1.20     1.12     1.26     0.92     1.45
#> 63:     1.19     1.10     1.25     0.90     1.46
#> 64:     1.19     1.11     1.26     0.90     1.48
#> 65:     1.20     1.11     1.26     0.86     1.46
#> 66:     1.21     1.10     1.26     0.86     1.47
#> 67:     1.22     1.10     1.26     0.84     1.47
#> 68:     1.20     1.09     1.26     0.83     1.48
#> 69:     1.21     1.10     1.27     0.80     1.48
#> 70:     1.20     1.09     1.27     0.81     1.51
#> 71:     1.20     1.08     1.26     0.82     1.54
#> 72:     1.20     1.08     1.26     0.82     1.54
#> 73:     1.20     1.08     1.26     0.82     1.54
#> 74:     1.20     1.08     1.26     0.82     1.54
#> 75:     1.20     1.08     1.26     0.82     1.54
#> 76:     1.20     1.08     1.26     0.82     1.54
#> 77:     1.20     1.08     1.26     0.82     1.54
#> 78:     1.20     1.08     1.26     0.82     1.54
#> 79:     1.20     1.08     1.26     0.82     1.54
#> 80:     1.20     1.08     1.26     0.82     1.54
#> 81:     1.20     1.08     1.26     0.82     1.54
#> 82:     1.20     1.08     1.26     0.82     1.54
#> 83:     1.20     1.08     1.26     0.82     1.54
#> 84:     1.20     1.08     1.26     0.82     1.54
#> 85:     1.20     1.08     1.26     0.82     1.54
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

#>     region       date                           type median          mean
#>  1: France 2020-07-09                       estimate   1004 4.960419e-321
#>  2: France 2020-07-10                       estimate   1039 5.148164e-321
#>  3: France 2020-07-11                       estimate   1077 5.345790e-321
#>  4: France 2020-07-12                       estimate   1118 5.558239e-321
#>  5: France 2020-07-13                       estimate   1162 5.785509e-321
#>  6: France 2020-07-14                       estimate   1207 6.022660e-321
#>  7: France 2020-07-15                       estimate   1258 6.279574e-321
#>  8: France 2020-07-16                       estimate   1312 6.556251e-321
#>  9: France 2020-07-17                       estimate   1373 6.847750e-321
#> 10: France 2020-07-18                       estimate   1433 7.159011e-321
#> 11: France 2020-07-19                       estimate   1499 7.499917e-321
#> 12: France 2020-07-20                       estimate   1568 7.860584e-321
#> 13: France 2020-07-21                       estimate   1642 8.245956e-321
#> 14: France 2020-07-22                       estimate   1727 8.660971e-321
#> 15: France 2020-07-23                       estimate   1819 9.110571e-321
#> 16: France 2020-07-24                       estimate   1911 9.589814e-321
#> 17: France 2020-07-25                       estimate   2011 1.010364e-320
#> 18: France 2020-07-26                       estimate   2114 1.066194e-320
#> 19: France 2020-07-27                       estimate   2226 1.125976e-320
#> 20: France 2020-07-28                       estimate   2354 1.190698e-320
#> 21: France 2020-07-29                       estimate   2494 1.260361e-320
#> 22: France 2020-07-30                       estimate   2632 1.335459e-320
#> 23: France 2020-07-31                       estimate   2785 1.416486e-320
#> 24: France 2020-08-01                       estimate   2951 1.504430e-320
#> 25: France 2020-08-02                       estimate   3121 1.598796e-320
#> 26: France 2020-08-03                       estimate   3306 1.701068e-320
#> 27: France 2020-08-04                       estimate   3509 1.810751e-320
#> 28: France 2020-08-05                       estimate   3727 1.929326e-320
#> 29: France 2020-08-06                       estimate   3962 2.056301e-320
#> 30: France 2020-08-07                       estimate   4214 2.192663e-320
#> 31: France 2020-08-08                       estimate   4507 2.338907e-320
#> 32: France 2020-08-09                       estimate   4774 2.495526e-320
#> 33: France 2020-08-10                       estimate   5055 2.662026e-320
#> 34: France 2020-08-11                       estimate   5359 2.839889e-320
#> 35: France 2020-08-12                       estimate   5707 3.028128e-320
#> 36: France 2020-08-13                       estimate   6047 3.228225e-320
#> 37: France 2020-08-14                       estimate   6421 3.439191e-320
#> 38: France 2020-08-15                       estimate   6808 3.661521e-320
#> 39: France 2020-08-16                       estimate   7229 3.895708e-320
#> 40: France 2020-08-17                       estimate   7629 4.140764e-320
#> 41: France 2020-08-18                       estimate   8083 4.396690e-320
#> 42: France 2020-08-19                       estimate   8591 4.663486e-320
#> 43: France 2020-08-20                       estimate   9098 4.940656e-320
#> 44: France 2020-08-21                       estimate   9625 5.228203e-320
#> 45: France 2020-08-22                       estimate  10151 5.526124e-320
#> 46: France 2020-08-23                       estimate  10740 5.834421e-320
#> 47: France 2020-08-24                       estimate  11314 6.153094e-320
#> 48: France 2020-08-25                       estimate  11931 6.483623e-320
#> 49: France 2020-08-26                       estimate  12564 6.825517e-320
#> 50: France 2020-08-27                       estimate  13154 7.180750e-320
#> 51: France 2020-08-28                       estimate  13824 7.549817e-320
#> 52: France 2020-08-29                       estimate  14421 7.934694e-320
#> 53: France 2020-08-30                       estimate  15090 8.336864e-320
#> 54: France 2020-08-31                       estimate  15780 8.758302e-320
#> 55: France 2020-09-01                       estimate  16541 9.202467e-320
#> 56: France 2020-09-02                       estimate  17263 9.671335e-320
#> 57: France 2020-09-03                       estimate  17924 1.016935e-319
#> 58: France 2020-09-04                       estimate  18664 1.070146e-319
#> 59: France 2020-09-05 estimate based on partial data  19500 1.127310e-319
#> 60: France 2020-09-06 estimate based on partial data  20415 1.189068e-319
#> 61: France 2020-09-07 estimate based on partial data  21236 1.256162e-319
#> 62: France 2020-09-08 estimate based on partial data  22113 1.329432e-319
#> 63: France 2020-09-09 estimate based on partial data  23133 1.409915e-319
#> 64: France 2020-09-10 estimate based on partial data  24244 1.498748e-319
#> 65: France 2020-09-11 estimate based on partial data  25197 1.597166e-319
#> 66: France 2020-09-12 estimate based on partial data  26278 1.706799e-319
#> 67: France 2020-09-13 estimate based on partial data  27483 1.829525e-319
#> 68: France 2020-09-14 estimate based on partial data  28657 1.967616e-319
#> 69: France 2020-09-15 estimate based on partial data  29789 2.124038e-319
#> 70: France 2020-09-16 estimate based on partial data  31080 2.302593e-319
#> 71: France 2020-09-17 estimate based on partial data  32443 2.508470e-319
#> 72: France 2020-09-18                       forecast  33911 2.742015e-319
#> 73: France 2020-09-19                       forecast  35311 3.010688e-319
#> 74: France 2020-09-20                       forecast  36949 3.321455e-319
#> 75: France 2020-09-21                       forecast  38772 3.683210e-319
#> 76: France 2020-09-22                       forecast  40464 4.107316e-319
#> 77: France 2020-09-23                       forecast  42477 4.608348e-319
#> 78: France 2020-09-24                       forecast  44652 5.205476e-319
#> 79: France 2020-09-25                       forecast  46412 5.923897e-319
#> 80: France 2020-09-26                       forecast  48509 6.797256e-319
#> 81: France 2020-09-27                       forecast  50651 7.871009e-319
#> 82: France 2020-09-28                       forecast  52932 9.207160e-319
#> 83: France 2020-09-29                       forecast  55362 1.089138e-318
#> 84: France 2020-09-30                       forecast  57626 1.304299e-318
#> 85: France 2020-10-01                       forecast  60227 1.583006e-318
#>     region       date                           type median          mean
#>                sd lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1: 1.422909e-321     1038     1178      814     1194      548     1432
#>  2: 1.506900e-321     1072     1223      844     1250      545     1482
#>  3: 1.590891e-321     1119     1281      859     1296      531     1512
#>  4: 1.679823e-321     1204     1381      898     1369      550     1586
#>  5: 1.768755e-321     1208     1399      943     1445      559     1658
#>  6: 1.867568e-321     1254     1459      988     1520      656     1816
#>  7: 1.966381e-321     1301     1517     1003     1576      627     1849
#>  8: 2.070135e-321     1353     1580     1039     1640      653     1941
#>  9: 2.183770e-321     1430     1674     1127     1753      647     2014
#> 10: 2.297405e-321     1230     1484     1117     1772      719     2170
#> 11: 2.415981e-321     1220     1482     1221     1904      747     2267
#> 12: 2.539497e-321     1347     1615     1263     1977      762     2356
#> 13: 2.667954e-321     1397     1674     1372     2112      803     2467
#> 14: 2.806293e-321     1454     1741     1421     2186      840     2586
#> 15: 2.949572e-321     1572     1871     1459     2258      936     2783
#> 16: 3.102732e-321     1600     1920     1600     2426      877     2815
#> 17: 3.260833e-321     1685     2014     1672     2535      945     2965
#> 18: 3.428816e-321     2219     2559     1696     2591      965     3089
#> 19: 3.606679e-321     1863     2216     1693     2626     1059     3286
#> 20: 3.799365e-321     2197     2555     1887     2853     1114     3447
#> 21: 4.001932e-321     2312     2678     1977     2972     1187     3633
#> 22: 4.229202e-321     2483     2866     2066     3105     1348     3946
#> 23: 4.471294e-321     2593     2989     2113     3198     1472     4210
#> 24: 4.743030e-321     2769     3180     2257     3382     1634     4515
#> 25: 5.044410e-321     2995     3440     2425     3597     1707     4757
#> 26: 5.380375e-321     2831     3288     2522     3770     1871     5100
#> 27: 5.760805e-321     3011     3483     2634     3966     1977     5383
#> 28: 6.190643e-321     3404     3902     2713     4112     2107     5690
#> 29: 6.669886e-321     3413     3947     2956     4425     2229     6087
#> 30: 7.208418e-321     3580     4157     3125     4679     2329     6442
#> 31: 7.801297e-321     3755     4353     3259     4908     2468     6851
#> 32: 8.458404e-321     3958     4599     3508     5271     2676     7377
#> 33: 9.179740e-321     4044     4714     3722     5619     2736     7870
#> 34: 9.965304e-321     4321     5046     3942     5969     2959     8444
#> 35: 1.082004e-320     4452     5224     4011     6184     2980     8880
#> 36: 1.175382e-320     4713     5564     4284     6611     3080     9488
#> 37: 1.276666e-320     5411     6320     4476     6989     3294    10155
#> 38: 1.385854e-320     5397     6343     4888     7584     3604    11025
#> 39: 1.502454e-320     5831     6837     5085     7978     4027    11947
#> 40: 1.626464e-320     6071     7152     5358     8451     4128    12596
#> 41: 1.757886e-320     6659     7826     5796     9042     4319    13379
#> 42: 1.895730e-320     6985     8259     6019     9472     4418    14211
#> 43: 2.039503e-320     6779     8169     6391    10072     4761    15200
#> 44: 2.188711e-320     7112     8592     6589    10555     5095    16285
#> 45: 2.342365e-320     7178     8753     6452    10738     5356    17140
#> 46: 2.500960e-320     8581    10193     6778    11330     5648    18166
#> 47: 2.664496e-320     8860    10612     7384    12298     5988    19257
#> 48: 2.835937e-320     9321    11188     8165    13422     6239    20351
#> 49: 3.018741e-320     9818    11844     8590    14164     6241    21256
#> 50: 3.217355e-320     9996    12165     9148    15018     6657    22644
#> 51: 3.436227e-320     9466    11746     9504    15735     6872    23834
#> 52: 3.679307e-320    10685    13108     9814    16521     6241    24423
#> 53: 3.950055e-320    10730    13215    10320    17474     7042    26571
#> 54: 4.253905e-320    11527    14178     9540    17090     6582    27485
#> 55: 4.595799e-320    11740    14545     9869    17870     7302    29379
#> 56: 4.983640e-320    12932    16005    10587    19118     7150    30811
#> 57: 5.427805e-320    12876    16144    10069    19106     7458    33325
#> 58: 5.940645e-320    13141    16687    10490    20218     7579    35436
#> 59: 6.537477e-320    13475    17334    10657    21127     8114    38234
#> 60: 7.236580e-320    14560    18660    11301    22645     7976    40643
#> 61: 8.059693e-320    14953    19355    11953    24113     8077    44176
#> 62: 9.033002e-320    13587    18310    11868    24912     7936    47675
#> 63: 1.018665e-319    13731    18904    12098    26165     7565    51181
#> 64: 1.155817e-319    13183    18688    13027    28371     5924    54547
#> 65: 1.319402e-319    14463    20304    12636    29169     5573    59367
#> 66: 1.515349e-319    13485    19817    11641    29473     3755    62654
#> 67: 1.751413e-319    13569    20418    13092    32227     3130    68918
#> 68: 2.038169e-319    13628    21058    11883    32510     2836    75994
#> 69: 2.390092e-319    14533    22450    12087    34750     2979    84595
#> 70: 2.828032e-319    12963    21517    10327    34975     2294    92574
#> 71: 3.382818e-319    13438    22444     9303    35951      708    98691
#> 72: 4.060627e-319    13807    23549    12559    41429     1760   110285
#> 73: 4.914224e-319    13630    24069     7915    39138      477   118523
#> 74: 5.998649e-319    15018    26260     7885    41189      574   131186
#> 75: 7.391321e-319    15479    27296     7355    42864      595   143362
#> 76: 9.202368e-319    15737    28405     7006    45068      261   158401
#> 77: 1.159152e-318    15686    29143     6540    47123      213   170798
#> 78: 1.479450e-318    15151    29710     7607    50534      174   188336
#> 79: 1.916421e-318    16559    32101     5578    51263      142   209872
#> 80: 2.523421e-318    16858    33362     5082    53237      116   234735
#> 81: 3.381588e-318    17447    34989     1500    51978       95   258497
#> 82: 4.614573e-318    15859    34646      960    53736       77   285393
#> 83: 6.410773e-318    15782    35750      347    55589       63   317088
#> 84: 9.056925e-318    16202    37308      109    57676       52   350400
#> 85: 1.298891e-317       42    21604       42    60228       42   390084
#>                sd lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesReportData
#>     region       date                           type median   mean      sd
#>  1: France 2020-07-09                       estimate    242    662    1144
#>  2: France 2020-07-10                       estimate    190    614    1094
#>  3: France 2020-07-11                       estimate    284    875    1618
#>  4: France 2020-07-12                       estimate     82    287     616
#>  5: France 2020-07-13                       estimate     86    262     482
#>  6: France 2020-07-14                       estimate    368    999    1616
#>  7: France 2020-07-15                       estimate    171    564    1202
#>  8: France 2020-07-16                       estimate    271    800    1411
#>  9: France 2020-07-17                       estimate    237    708    1223
#> 10: France 2020-07-18                       estimate    366   1145    1988
#> 11: France 2020-07-19                       estimate    119    406     795
#> 12: France 2020-07-20                       estimate    104    365     738
#> 13: France 2020-07-21                       estimate    482   1441    2398
#> 14: France 2020-07-22                       estimate    224    750    1370
#> 15: France 2020-07-23                       estimate    348   1151    2187
#> 16: France 2020-07-24                       estimate    350   1069    2108
#> 17: France 2020-07-25                       estimate    527   1526    2774
#> 18: France 2020-07-26                       estimate    149    521    1070
#> 19: France 2020-07-27                       estimate    142    473     893
#> 20: France 2020-07-28                       estimate    579   1885    3228
#> 21: France 2020-07-29                       estimate    306   1011    1949
#> 22: France 2020-07-30                       estimate    479   1495    2964
#> 23: France 2020-07-31                       estimate    418   1401    2463
#> 24: France 2020-08-01                       estimate    672   2114    4163
#> 25: France 2020-08-02                       estimate    208    746    1714
#> 26: France 2020-08-03                       estimate    208    722    1506
#> 27: France 2020-08-04                       estimate    902   2709    4476
#> 28: France 2020-08-05                       estimate    500   1468    2787
#> 29: France 2020-08-06                       estimate    717   2250    4109
#> 30: France 2020-08-07                       estimate    710   2142    3763
#> 31: France 2020-08-08                       estimate    986   3033    5451
#> 32: France 2020-08-09                       estimate    352   1042    1811
#> 33: France 2020-08-10                       estimate    272    952    1784
#> 34: France 2020-08-11                       estimate   1256   4315    8162
#> 35: France 2020-08-12                       estimate    669   2142    3915
#> 36: France 2020-08-13                       estimate   1022   3160    5707
#> 37: France 2020-08-14                       estimate   1030   3370    6421
#> 38: France 2020-08-15                       estimate   1618   5008    9084
#> 39: France 2020-08-16                       estimate    520   1624    3359
#> 40: France 2020-08-17                       estimate    498   1524    2713
#> 41: France 2020-08-18                       estimate   2036   6790   12299
#> 42: France 2020-08-19                       estimate   1037   3369    7064
#> 43: France 2020-08-20                       estimate   1580   5205    9457
#> 44: France 2020-08-21                       estimate   1738   5229    9792
#> 45: France 2020-08-22                       estimate   2118   7422   14506
#> 46: France 2020-08-23                       estimate    754   2382    4591
#> 47: France 2020-08-24                       estimate    740   2288    4546
#> 48: France 2020-08-25                       estimate   3348  10461   18194
#> 49: France 2020-08-26                       estimate   1622   4908    8922
#> 50: France 2020-08-27                       estimate   2757   7893   14511
#> 51: France 2020-08-28                       estimate   2480   7469   15216
#> 52: France 2020-08-29                       estimate   3487  11684   23462
#> 53: France 2020-08-30                       estimate   1180   3599    7250
#> 54: France 2020-08-31                       estimate   1128   3499    6866
#> 55: France 2020-09-01                       estimate   4892  16604   36957
#> 56: France 2020-09-02                       estimate   2426   7688   15138
#> 57: France 2020-09-03                       estimate   3478  11198   20475
#> 58: France 2020-09-04                       estimate   3538  11389   22987
#> 59: France 2020-09-05 estimate based on partial data   5186  16440   31810
#> 60: France 2020-09-06 estimate based on partial data   1670   5097    9502
#> 61: France 2020-09-07 estimate based on partial data   1494   4763    9522
#> 62: France 2020-09-08 estimate based on partial data   6441  21795   41259
#> 63: France 2020-09-09 estimate based on partial data   3052  10337   22271
#> 64: France 2020-09-10 estimate based on partial data   5240  16430   32580
#> 65: France 2020-09-11 estimate based on partial data   4812  15160   30150
#> 66: France 2020-09-12 estimate based on partial data   7459  24302   47569
#> 67: France 2020-09-13 estimate based on partial data   2234   7132   15779
#> 68: France 2020-09-14 estimate based on partial data   1984   7309   16184
#> 69: France 2020-09-15 estimate based on partial data   8586  33478   83028
#> 70: France 2020-09-16 estimate based on partial data   4550  15586   31628
#> 71: France 2020-09-17 estimate based on partial data   7402  26224   62299
#> 72: France 2020-09-18                       forecast   7243  27185   66450
#> 73: France 2020-09-19                       forecast   9478  37720  127012
#> 74: France 2020-09-20                       forecast   2896  11258   25953
#> 75: France 2020-09-21                       forecast   2781  11952   32953
#> 76: France 2020-09-22                       forecast  11658  52857  129145
#> 77: France 2020-09-23                       forecast   5036  25160   65534
#> 78: France 2020-09-24                       forecast   8617  49790  217191
#> 79: France 2020-09-25                       forecast   8508  42664  122347
#> 80: France 2020-09-26                       forecast  12230  65392  211185
#> 81: France 2020-09-27                       forecast   3754  20731   93103
#> 82: France 2020-09-28                       forecast   3664  21381   78659
#> 83: France 2020-09-29                       forecast  14284 127459 1052670
#> 84: France 2020-09-30                       forecast   8262  56715  266375
#> 85: France 2020-10-01                       forecast  10877  74921  241024
#>     region       date                           type median   mean      sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:        0       22        0      242        0     1796
#>  2:        0       19        0      190        0     1756
#>  3:        0       30        0      284        0     2282
#>  4:        0        8        0       82        0      772
#>  5:        0        8        0       86        0      708
#>  6:        0       35        0      368        0     2771
#>  7:        0       16        0      171        0     1514
#>  8:        0       26        0      271        0     2141
#>  9:        0       23        0      237        0     1961
#> 10:        0       34        0      367        0     3154
#> 11:        0       10        0      119        0     1070
#> 12:        0        9        0      104        0      997
#> 13:        0       49        0      483        0     3929
#> 14:        0       22        0      224        0     2045
#> 15:        0       38        0      348        0     3141
#> 16:        0       30        0      350        0     2817
#> 17:        0       56        0      528        0     4058
#> 18:        0       11        0      149        0     1403
#> 19:        0       12        0      143        0     1301
#> 20:        0       53        0      580        0     5520
#> 21:        0       31        0      307        0     2647
#> 22:        0       42        0      479        0     3964
#> 23:        0       38        0      418        0     3977
#> 24:        0       66        0      674        0     5669
#> 25:        0       18        0      208        0     1982
#> 26:        0       19        0      208        0     1838
#> 27:        0       81        0      903        0     7675
#> 28:        0       47        0      501        0     3779
#> 29:        0       72        0      718        0     5998
#> 30:        0       65        0      710        0     6091
#> 31:        0       86        0      988        0     7929
#> 32:        0       34        0      353        0     2909
#> 33:        0       27        0      273        0     2552
#> 34:        0      126        0     1257        0    11556
#> 35:        0       58        0      669        0     5863
#> 36:        0      108        0     1023        0     8721
#> 37:        0       84        0     1031        0     9051
#> 38:        0      162        0     1619        0    13550
#> 39:        0       50        0      521        0     4195
#> 40:        0       40        0      498        0     4078
#> 41:        0      187        0     2039        0    19044
#> 42:        0      103        0     1037        0     9205
#> 43:        0      160        0     1581        0    14276
#> 44:        0      177        0     1739        0    14048
#> 45:        0      181        0     2119        0    19633
#> 46:        0       77        0      754        0     6206
#> 47:        0       68        0      742        0     6134
#> 48:        0      305        0     3356        0    28654
#> 49:        0      135        0     1631        0    13689
#> 50:        0      249        0     2760        0    21659
#> 51:        0      221        0     2481        0    19027
#> 52:        0      312        0     3505        0    30586
#> 53:        0      106        0     1182        0    10003
#> 54:        0       92        0     1128        0     9967
#> 55:        0      503        0     4895        0    44197
#> 56:        0      238        0     2433        0    19660
#> 57:        0      322        0     3482        0    30980
#> 58:        0      351        0     3540        0    29935
#> 59:        0      460        0     5206        0    43451
#> 60:        0      154        0     1672        0    13646
#> 61:        0      146        0     1494        0    12500
#> 62:        0      589        0     6458        0    62778
#> 63:        0      257        0     3064        0    27948
#> 64:        0      528        0     5247        0    41870
#> 65:        0      422        0     4817        0    39159
#> 66:        0      556        0     7471        0    65526
#> 67:        0      212        0     2241        0    18154
#> 68:        0      180        0     1987        0    20078
#> 69:        0      777        0     8589        0    97169
#> 70:        0      403        0     4550        0    42168
#> 71:        0      564        0     7422        0    66746
#> 72:        0      588        0     7246        0    65303
#> 73:        0      964        0     9485        0    88119
#> 74:        0      293        0     2901        0    28024
#> 75:        0      228        0     2782        0    27766
#> 76:        0     1045        0    11664        0   124920
#> 77:        0      440        0     5038        0    60158
#> 78:        0      681        0     8618        0   107302
#> 79:        0      755        0     8513        0    97953
#> 80:        0     1057        0    12276        0   148459
#> 81:        0      358        0     3758        0    42886
#> 82:        0      366        0     3672        0    44758
#> 83:        0     1187        0    14292        0   212041
#> 84:        0      653        0     8264        0    97695
#> 85:        0      910        0    10905        0   165035
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-06-25      81
#>  2: France 2020-06-26       0
#>  3: France 2020-06-27    1588
#>  4: France 2020-06-28       0
#>  5: France 2020-06-29       0
#>  6: France 2020-06-30    1324
#>  7: France 2020-07-01     541
#>  8: France 2020-07-02     918
#>  9: France 2020-07-03     659
#> 10: France 2020-07-04     582
#> 11: France 2020-07-05       0
#> 12: France 2020-07-06       0
#> 13: France 2020-07-07    1375
#> 14: France 2020-07-08     475
#> 15: France 2020-07-09     663
#> 16: France 2020-07-10     621
#> 17: France 2020-07-11     658
#> 18: France 2020-07-12       0
#> 19: France 2020-07-13       0
#> 20: France 2020-07-14    1625
#> 21: France 2020-07-15       0
#> 22: France 2020-07-16     927
#> 23: France 2020-07-17     534
#> 24: France 2020-07-18     836
#> 25: France 2020-07-19       0
#> 26: France 2020-07-20       0
#> 27: France 2020-07-21    2080
#> 28: France 2020-07-22     584
#> 29: France 2020-07-23     998
#> 30: France 2020-07-24    1062
#> 31: France 2020-07-25    1130
#> 32: France 2020-07-26       0
#> 33: France 2020-07-27       0
#> 34: France 2020-07-28    2551
#> 35: France 2020-07-29     725
#> 36: France 2020-07-30    1392
#> 37: France 2020-07-31    1377
#> 38: France 2020-08-01    1346
#> 39: France 2020-08-02       0
#> 40: France 2020-08-03       0
#> 41: France 2020-08-04    3376
#> 42: France 2020-08-05    1039
#> 43: France 2020-08-06    1695
#> 44: France 2020-08-07    1604
#> 45: France 2020-08-08    2288
#> 46: France 2020-08-09       0
#> 47: France 2020-08-10       0
#> 48: France 2020-08-11    4854
#> 49: France 2020-08-12    1397
#> 50: France 2020-08-13    2524
#> 51: France 2020-08-14    2669
#> 52: France 2020-08-15    2846
#> 53: France 2020-08-16    3310
#> 54: France 2020-08-17    3015
#> 55: France 2020-08-18     493
#> 56: France 2020-08-19    2238
#> 57: France 2020-08-20    3776
#> 58: France 2020-08-21    4771
#> 59: France 2020-08-22    4586
#> 60: France 2020-08-23    3602
#> 61: France 2020-08-24    4897
#> 62: France 2020-08-25    1955
#> 63: France 2020-08-26    3304
#> 64: France 2020-08-27    5429
#> 65: France 2020-08-28    6111
#> 66: France 2020-08-29    7379
#> 67: France 2020-08-30    5453
#> 68: France 2020-08-31    5413
#> 69: France 2020-09-01    3082
#> 70: France 2020-09-02    4982
#> 71: France 2020-09-03    7017
#> 72: France 2020-09-04    7157
#> 73: France 2020-09-05    8975
#> 74: France 2020-09-06    8550
#> 75: France 2020-09-07    7071
#> 76: France 2020-09-08    4203
#> 77: France 2020-09-09    6544
#> 78: France 2020-09-10    8577
#> 79: France 2020-09-11    9843
#> 80: France 2020-09-12    9406
#> 81: France 2020-09-13   10561
#> 82: France 2020-09-14    7183
#> 83: France 2020-09-15    6158
#> 84: France 2020-09-16    7852
#> 85: France 2020-09-17    9784
#>     region       date confirm
#>