Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

rtData

#> $summaryData
#>              region New confirmed cases by infection date
#>   1:    Afghanistan                          35 (6 -- 95)
#>   2:        Albania                       155 (98 -- 223)
#>   3:        Algeria                        98 (67 -- 134)
#>   4:        Andorra                         69 (9 -- 174)
#>   5:         Angola                       164 (56 -- 332)
#>  ---                                                     
#> 192:        Vietnam                            1 (0 -- 5)
#> 193: Western Sahara                          1 (0 -- 289)
#> 194:          Yemen                            1 (0 -- 3)
#> 195:         Zambia                         48 (6 -- 126)
#> 196:       Zimbabwe                           4 (0 -- 13)
#>      Expected change in daily cases Effective reproduction no.
#>   1:                         Unsure             1 (0.8 -- 1.3)
#>   2:                         Unsure             1 (0.9 -- 1.2)
#>   3:              Likely decreasing             0.9 (0.8 -- 1)
#>   4:              Likely increasing           1.2 (0.9 -- 1.4)
#>   5:              Likely increasing           1.1 (0.9 -- 1.3)
#>  ---                                                          
#> 192:                         Unsure             0.7 (0 -- 1.4)
#> 193:                         Unsure             0.6 (0 -- 1.4)
#> 194:              Likely decreasing           0.9 (0.7 -- 1.2)
#> 195:                         Unsure           0.9 (0.6 -- 1.3)
#> 196:              Likely decreasing           0.8 (0.4 -- 1.2)
#>             Rate of growth Doubling/halving time (days)
#>   1: -0.01 (-0.06 -- 0.08)        -117.6 (8.2 -- -11.2)
#>   2:  0.01 (-0.03 -- 0.05)        107.5 (14.2 -- -24.2)
#>   3:    -0.03 (-0.06 -- 0)       -23.9 (570.7 -- -11.6)
#>   4:  0.04 (-0.03 -- 0.12)          17.1 (5.9 -- -20.3)
#>   5:  0.02 (-0.02 -- 0.07)          31.7 (9.6 -- -27.7)
#>  ---                                                   
#> 192:  -0.1 (-0.32 -- 0.12)             -7 (5.6 -- -2.2)
#> 193:  -0.1 (-0.36 -- 0.14)           -6.7 (4.8 -- -1.9)
#> 194: -0.03 (-0.09 -- 0.05)         -24.9 (13.1 -- -7.8)
#> 195:  -0.02 (-0.1 -- 0.08)          -32.1 (8.2 -- -6.8)
#> 196: -0.07 (-0.18 -- 0.05)             -10 (14.3 -- -4)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-07-30 estimate   0.92 0.91 0.05     0.92     0.93
#>     2: Afghanistan 2020-07-31 estimate   0.92 0.91 0.05     0.92     0.94
#>     3: Afghanistan 2020-08-01 estimate   0.92 0.92 0.05     0.92     0.93
#>     4: Afghanistan 2020-08-02 estimate   0.92 0.92 0.06     0.92     0.93
#>     5: Afghanistan 2020-08-03 estimate   0.92 0.92 0.06     0.92     0.93
#>    ---                                                                   
#> 16264:    Zimbabwe 2020-10-18 forecast   0.76 0.80 0.27     0.70     0.78
#> 16265:    Zimbabwe 2020-10-19 forecast   0.76 0.80 0.27     0.70     0.78
#> 16266:    Zimbabwe 2020-10-20 forecast   0.76 0.80 0.27     0.70     0.78
#> 16267:    Zimbabwe 2020-10-21 forecast   0.76 0.80 0.27     0.70     0.78
#> 16268:    Zimbabwe 2020-10-22 forecast   0.76 0.80 0.27     0.70     0.78
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.90     0.95     0.84     0.98
#>     2:     0.90     0.94     0.83     0.98
#>     3:     0.90     0.95     0.83     0.99
#>     4:     0.90     0.95     0.83     0.99
#>     5:     0.90     0.95     0.83     1.00
#>    ---                                    
#> 16264:     0.62     0.85     0.40     1.15
#> 16265:     0.62     0.85     0.40     1.15
#> 16266:     0.62     0.85     0.40     1.15
#> 16267:     0.62     0.85     0.40     1.15
#> 16268:     0.62     0.85     0.40     1.15
#> 
#> $casesInfectionData

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

#>             region       date     type median mean     sd lower_20 upper_20
#>     1: Afghanistan 2020-07-30 estimate     77   77     16       77       83
#>     2: Afghanistan 2020-07-31 estimate     75   75     16       75       82
#>     3: Afghanistan 2020-08-01 estimate     73   73     16       74       80
#>     4: Afghanistan 2020-08-02 estimate     72   72     15       71       78
#>     5: Afghanistan 2020-08-03 estimate     70   70     15       71       77
#>    ---                                                                     
#> 16264:    Zimbabwe 2020-10-18 forecast      2  151   7695        0        1
#> 16265:    Zimbabwe 2020-10-19 forecast      2  277  15020        0        1
#> 16266:    Zimbabwe 2020-10-20 forecast      2  519  29325        0        1
#> 16267:    Zimbabwe 2020-10-21 forecast      2  984  57262        0        1
#> 16268:    Zimbabwe 2020-10-22 forecast      2 1883 111825        0        0
#>        lower_50 upper_50 lower_90      upper_90
#>     1:       69       88       50 4.940656e-322
#>     2:       66       84       48 4.792437e-322
#>     3:       64       82       46 4.644217e-322
#>     4:       62       80       46 4.644217e-322
#>     5:       60       79       44 4.495997e-322
#>    ---                                         
#> 16264:        0        2        0 7.905050e-323
#> 16265:        0        2        0 7.905050e-323
#> 16266:        0        2        0 8.399116e-323
#> 16267:        0        2        0 8.399116e-323
#> 16268:        0        2        0 8.399116e-323
#> 
#> $casesReportData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-07-30 estimate     54  106  144        0       11
#>     2: Afghanistan 2020-07-31 estimate     33   64   91        0        6
#>     3: Afghanistan 2020-08-01 estimate     40   79  111        0        8
#>     4: Afghanistan 2020-08-02 estimate     47   89  124        0       10
#>     5: Afghanistan 2020-08-03 estimate     34   68  100        0        6
#>    ---                                                                   
#> 16264:    Zimbabwe 2020-10-18 forecast      4   17  514        0        1
#> 16265:    Zimbabwe 2020-10-19 forecast      2    8   55        0        0
#> 16266:    Zimbabwe 2020-10-20 forecast      4   43 1879        0        1
#> 16267:    Zimbabwe 2020-10-21 forecast      3   20  472        0        1
#> 16268:    Zimbabwe 2020-10-22 forecast      2   60 2598        0        0
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0       54        0      265
#>     2:        0       33        0      163
#>     3:        0       40        0      199
#>     4:        0       47        0      225
#>     5:        0       34        0      171
#>    ---                                    
#> 16264:        0        4        0       17
#> 16265:        0        2        0       12
#> 16266:        0        4        0       20
#> 16267:        0        3        0       16
#> 16268:        0        2        0       14
#> 
#> $obsCasesData
#>                region       date confirm
#>     1:    Afghanistan 2020-07-16     254
#>     2:        Albania 2020-07-16      85
#>     3:        Algeria 2020-07-16     554
#>     4:        Andorra 2020-07-16       1
#>     5:         Angola 2020-07-16      51
#>    ---                                  
#> 17761:        Vietnam 2020-10-08       1
#> 17762: Western Sahara 2020-10-08       0
#> 17763:          Yemen 2020-10-08       8
#> 17764:         Zambia 2020-10-08      54
#> 17765:       Zimbabwe 2020-10-08       4
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France                 25278 (3547 -- 52922)              Likely increasing
#>    Effective reproduction no.       Rate of growth Doubling/halving time (days)
#> 1:           1.1 (0.8 -- 1.4) 0.03 (-0.06 -- 0.09)          20.3 (7.7 -- -11.7)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-07-30                       estimate   1.20 1.20 0.06     1.18
#>  2: France 2020-07-31                       estimate   1.20 1.20 0.06     1.18
#>  3: France 2020-08-01                       estimate   1.20 1.20 0.06     1.19
#>  4: France 2020-08-02                       estimate   1.20 1.20 0.06     1.18
#>  5: France 2020-08-03                       estimate   1.20 1.20 0.07     1.18
#>  6: France 2020-08-04                       estimate   1.20 1.20 0.07     1.18
#>  7: France 2020-08-05                       estimate   1.20 1.20 0.07     1.18
#>  8: France 2020-08-06                       estimate   1.20 1.20 0.07     1.18
#>  9: France 2020-08-07                       estimate   1.20 1.20 0.07     1.18
#> 10: France 2020-08-08                       estimate   1.20 1.20 0.07     1.18
#> 11: France 2020-08-09                       estimate   1.20 1.21 0.07     1.18
#> 12: France 2020-08-10                       estimate   1.20 1.21 0.07     1.18
#> 13: France 2020-08-11                       estimate   1.20 1.21 0.07     1.18
#> 14: France 2020-08-12                       estimate   1.20 1.21 0.08     1.17
#> 15: France 2020-08-13                       estimate   1.20 1.21 0.08     1.17
#> 16: France 2020-08-14                       estimate   1.19 1.21 0.08     1.17
#> 17: France 2020-08-15                       estimate   1.19 1.21 0.08     1.17
#> 18: France 2020-08-16                       estimate   1.19 1.21 0.09     1.17
#> 19: France 2020-08-17                       estimate   1.19 1.21 0.09     1.17
#> 20: France 2020-08-18                       estimate   1.19 1.21 0.09     1.17
#> 21: France 2020-08-19                       estimate   1.19 1.20 0.08     1.17
#> 22: France 2020-08-20                       estimate   1.19 1.20 0.08     1.17
#> 23: France 2020-08-21                       estimate   1.18 1.20 0.07     1.16
#> 24: France 2020-08-22                       estimate   1.18 1.19 0.07     1.16
#> 25: France 2020-08-23                       estimate   1.18 1.19 0.07     1.16
#> 26: France 2020-08-24                       estimate   1.18 1.18 0.06     1.16
#> 27: France 2020-08-25                       estimate   1.17 1.18 0.06     1.17
#> 28: France 2020-08-26                       estimate   1.17 1.17 0.07     1.17
#> 29: France 2020-08-27                       estimate   1.17 1.17 0.07     1.16
#> 30: France 2020-08-28                       estimate   1.17 1.16 0.08     1.16
#> 31: France 2020-08-29                       estimate   1.16 1.16 0.08     1.16
#> 32: France 2020-08-30                       estimate   1.16 1.16 0.09     1.16
#> 33: France 2020-08-31                       estimate   1.16 1.15 0.09     1.16
#> 34: France 2020-09-01                       estimate   1.16 1.15 0.09     1.15
#> 35: France 2020-09-02                       estimate   1.15 1.14 0.09     1.16
#> 36: France 2020-09-03                       estimate   1.15 1.14 0.09     1.16
#> 37: France 2020-09-04                       estimate   1.15 1.14 0.08     1.16
#> 38: France 2020-09-05                       estimate   1.15 1.14 0.08     1.16
#> 39: France 2020-09-06                       estimate   1.14 1.13 0.07     1.16
#> 40: France 2020-09-07                       estimate   1.14 1.13 0.07     1.14
#> 41: France 2020-09-08                       estimate   1.14 1.13 0.07     1.14
#> 42: France 2020-09-09                       estimate   1.13 1.13 0.06     1.15
#> 43: France 2020-09-10                       estimate   1.13 1.13 0.06     1.14
#> 44: France 2020-09-11                       estimate   1.13 1.13 0.07     1.14
#> 45: France 2020-09-12                       estimate   1.13 1.13 0.08     1.14
#> 46: France 2020-09-13                       estimate   1.13 1.12 0.08     1.14
#> 47: France 2020-09-14                       estimate   1.13 1.12 0.09     1.14
#> 48: France 2020-09-15                       estimate   1.13 1.12 0.09     1.14
#> 49: France 2020-09-16                       estimate   1.12 1.12 0.10     1.14
#> 50: France 2020-09-17                       estimate   1.12 1.12 0.10     1.14
#> 51: France 2020-09-18                       estimate   1.12 1.11 0.09     1.13
#> 52: France 2020-09-19                       estimate   1.12 1.11 0.09     1.13
#> 53: France 2020-09-20                       estimate   1.12 1.10 0.09     1.13
#> 54: France 2020-09-21                       estimate   1.12 1.10 0.09     1.12
#> 55: France 2020-09-22                       estimate   1.11 1.10 0.10     1.13
#> 56: France 2020-09-23                       estimate   1.11 1.09 0.10     1.14
#> 57: France 2020-09-24                       estimate   1.11 1.09 0.11     1.15
#> 58: France 2020-09-25                       estimate   1.11 1.09 0.11     1.15
#> 59: France 2020-09-26                       estimate   1.11 1.09 0.12     1.15
#> 60: France 2020-09-27                       estimate   1.11 1.09 0.12     1.15
#> 61: France 2020-09-28 estimate based on partial data   1.11 1.09 0.13     1.14
#> 62: France 2020-09-29 estimate based on partial data   1.11 1.09 0.13     1.14
#> 63: France 2020-09-30 estimate based on partial data   1.11 1.09 0.14     1.15
#> 64: France 2020-10-01 estimate based on partial data   1.11 1.09 0.14     1.14
#> 65: France 2020-10-02 estimate based on partial data   1.11 1.09 0.14     1.15
#> 66: France 2020-10-03 estimate based on partial data   1.11 1.10 0.14     1.14
#> 67: France 2020-10-04 estimate based on partial data   1.12 1.10 0.14     1.14
#> 68: France 2020-10-05 estimate based on partial data   1.12 1.10 0.14     1.13
#> 69: France 2020-10-06 estimate based on partial data   1.12 1.11 0.15     1.14
#> 70: France 2020-10-07 estimate based on partial data   1.13 1.12 0.18     1.15
#> 71: France 2020-10-08 estimate based on partial data   1.13 1.13 0.21     1.14
#> 72: France 2020-10-09                       forecast   1.13 1.13 0.21     1.14
#> 73: France 2020-10-10                       forecast   1.13 1.13 0.21     1.14
#> 74: France 2020-10-11                       forecast   1.13 1.13 0.21     1.14
#> 75: France 2020-10-12                       forecast   1.13 1.13 0.21     1.14
#> 76: France 2020-10-13                       forecast   1.13 1.13 0.21     1.14
#> 77: France 2020-10-14                       forecast   1.13 1.13 0.21     1.14
#> 78: France 2020-10-15                       forecast   1.13 1.13 0.21     1.14
#> 79: France 2020-10-16                       forecast   1.13 1.13 0.21     1.14
#> 80: France 2020-10-17                       forecast   1.13 1.13 0.21     1.14
#> 81: France 2020-10-18                       forecast   1.13 1.13 0.21     1.14
#> 82: France 2020-10-19                       forecast   1.13 1.13 0.21     1.14
#> 83: France 2020-10-20                       forecast   1.13 1.13 0.21     1.14
#> 84: France 2020-10-21                       forecast   1.13 1.13 0.21     1.14
#> 85: France 2020-10-22                       forecast   1.13 1.13 0.21     1.14
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.20     1.17     1.22     1.12     1.29
#>  2:     1.20     1.17     1.22     1.12     1.30
#>  3:     1.21     1.17     1.22     1.10     1.29
#>  4:     1.20     1.17     1.23     1.10     1.31
#>  5:     1.20     1.17     1.22     1.08     1.30
#>  6:     1.20     1.17     1.23     1.09     1.32
#>  7:     1.20     1.16     1.22     1.09     1.32
#>  8:     1.20     1.16     1.22     1.07     1.30
#>  9:     1.20     1.16     1.22     1.09     1.32
#> 10:     1.20     1.16     1.22     1.09     1.32
#> 11:     1.20     1.16     1.22     1.10     1.33
#> 12:     1.20     1.16     1.22     1.11     1.34
#> 13:     1.20     1.16     1.22     1.11     1.34
#> 14:     1.19     1.16     1.22     1.11     1.35
#> 15:     1.19     1.16     1.22     1.08     1.35
#> 16:     1.19     1.16     1.22     1.08     1.34
#> 17:     1.19     1.16     1.22     1.09     1.34
#> 18:     1.19     1.16     1.22     1.09     1.33
#> 19:     1.19     1.16     1.22     1.08     1.32
#> 20:     1.19     1.15     1.21     1.09     1.32
#> 21:     1.19     1.15     1.21     1.09     1.31
#> 22:     1.19     1.15     1.21     1.09     1.31
#> 23:     1.18     1.15     1.21     1.09     1.30
#> 24:     1.18     1.15     1.21     1.09     1.31
#> 25:     1.19     1.14     1.20     1.09     1.30
#> 26:     1.18     1.14     1.21     1.09     1.28
#> 27:     1.19     1.14     1.20     1.07     1.27
#> 28:     1.19     1.14     1.20     1.06     1.27
#> 29:     1.19     1.14     1.20     1.03     1.27
#> 30:     1.19     1.14     1.20     1.05     1.29
#> 31:     1.18     1.14     1.20     1.03     1.27
#> 32:     1.18     1.13     1.20     1.04     1.28
#> 33:     1.18     1.13     1.20     1.04     1.27
#> 34:     1.18     1.13     1.20     1.04     1.27
#> 35:     1.18     1.13     1.20     1.03     1.26
#> 36:     1.18     1.13     1.20     1.02     1.24
#> 37:     1.18     1.12     1.19     1.02     1.25
#> 38:     1.18     1.13     1.20     1.02     1.25
#> 39:     1.18     1.12     1.19     1.02     1.24
#> 40:     1.17     1.12     1.19     1.02     1.23
#> 41:     1.17     1.11     1.19     1.02     1.22
#> 42:     1.17     1.10     1.18     1.02     1.22
#> 43:     1.17     1.11     1.19     1.02     1.22
#> 44:     1.17     1.10     1.19     1.00     1.23
#> 45:     1.17     1.10     1.18     0.99     1.23
#> 46:     1.17     1.10     1.19     0.98     1.23
#> 47:     1.17     1.10     1.19     0.97     1.22
#> 48:     1.17     1.10     1.19     0.97     1.23
#> 49:     1.17     1.10     1.19     0.96     1.23
#> 50:     1.17     1.09     1.19     0.96     1.23
#> 51:     1.17     1.09     1.19     0.95     1.23
#> 52:     1.17     1.09     1.19     0.95     1.23
#> 53:     1.17     1.09     1.19     0.96     1.23
#> 54:     1.15     1.09     1.19     0.96     1.24
#> 55:     1.17     1.09     1.19     0.95     1.23
#> 56:     1.18     1.08     1.19     0.93     1.24
#> 57:     1.19     1.08     1.19     0.90     1.24
#> 58:     1.19     1.08     1.20     0.89     1.24
#> 59:     1.19     1.08     1.20     0.88     1.24
#> 60:     1.19     1.08     1.20     0.88     1.25
#> 61:     1.19     1.08     1.20     0.88     1.26
#> 62:     1.19     1.08     1.20     0.87     1.27
#> 63:     1.19     1.08     1.21     0.86     1.28
#> 64:     1.19     1.08     1.20     0.84     1.27
#> 65:     1.20     1.07     1.20     0.83     1.26
#> 66:     1.19     1.07     1.20     0.86     1.29
#> 67:     1.19     1.07     1.21     0.86     1.29
#> 68:     1.18     1.07     1.21     0.86     1.30
#> 69:     1.18     1.07     1.21     0.82     1.31
#> 70:     1.20     1.07     1.21     0.82     1.34
#> 71:     1.18     1.06     1.21     0.80     1.36
#> 72:     1.18     1.06     1.21     0.80     1.36
#> 73:     1.18     1.06     1.21     0.80     1.36
#> 74:     1.18     1.06     1.21     0.80     1.36
#> 75:     1.18     1.06     1.21     0.80     1.36
#> 76:     1.18     1.06     1.21     0.80     1.36
#> 77:     1.18     1.06     1.21     0.80     1.36
#> 78:     1.18     1.06     1.21     0.80     1.36
#> 79:     1.18     1.06     1.21     0.80     1.36
#> 80:     1.18     1.06     1.21     0.80     1.36
#> 81:     1.18     1.06     1.21     0.80     1.36
#> 82:     1.18     1.06     1.21     0.80     1.36
#> 83:     1.18     1.06     1.21     0.80     1.36
#> 84:     1.18     1.06     1.21     0.80     1.36
#> 85:     1.18     1.06     1.21     0.80     1.36
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData

#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.

#>     region       date                           type median   mean     sd
#>  1: France 2020-07-30                       estimate   1728   1787    392
#>  2: France 2020-07-31                       estimate   1820   1880    410
#>  3: France 2020-08-01                       estimate   1916   1977    428
#>  4: France 2020-08-02                       estimate   2006   2080    448
#>  5: France 2020-08-03                       estimate   2112   2189    470
#>  6: France 2020-08-04                       estimate   2224   2304    495
#>  7: France 2020-08-05                       estimate   2337   2426    522
#>  8: France 2020-08-06                       estimate   2454   2556    552
#>  9: France 2020-08-07                       estimate   2581   2693    585
#> 10: France 2020-08-08                       estimate   2721   2838    621
#> 11: France 2020-08-09                       estimate   2872   2993    661
#> 12: France 2020-08-10                       estimate   3021   3157    703
#> 13: France 2020-08-11                       estimate   3192   3330    749
#> 14: France 2020-08-12                       estimate   3365   3513    799
#> 15: France 2020-08-13                       estimate   3548   3707    854
#> 16: France 2020-08-14                       estimate   3747   3912    916
#> 17: France 2020-08-15                       estimate   3948   4128    985
#> 18: France 2020-08-16                       estimate   4151   4354   1066
#> 19: France 2020-08-17                       estimate   4353   4592   1159
#> 20: France 2020-08-18                       estimate   4563   4840   1264
#> 21: France 2020-08-19                       estimate   4790   5096   1379
#> 22: France 2020-08-20                       estimate   5019   5359   1498
#> 23: France 2020-08-21                       estimate   5254   5628   1613
#> 24: France 2020-08-22                       estimate   5514   5899   1714
#> 25: France 2020-08-23                       estimate   5754   6172   1793
#> 26: France 2020-08-24                       estimate   6018   6444   1847
#> 27: France 2020-08-25                       estimate   6286   6717   1877
#> 28: France 2020-08-26                       estimate   6574   6990   1892
#> 29: France 2020-08-27                       estimate   6854   7265   1903
#> 30: France 2020-08-28                       estimate   7148   7545   1921
#> 31: France 2020-08-29                       estimate   7447   7830   1958
#> 32: France 2020-08-30                       estimate   7753   8123   2018
#> 33: France 2020-08-31                       estimate   8027   8424   2104
#> 34: France 2020-09-01                       estimate   8312   8734   2211
#> 35: France 2020-09-02                       estimate   8599   9052   2335
#> 36: France 2020-09-03                       estimate   8897   9378   2466
#> 37: France 2020-09-04                       estimate   9205   9710   2596
#> 38: France 2020-09-05                       estimate   9544  10047   2715
#> 39: France 2020-09-06                       estimate   9871  10389   2816
#> 40: France 2020-09-07                       estimate  10227  10735   2894
#> 41: France 2020-09-08                       estimate  10578  11085   2950
#> 42: France 2020-09-09                       estimate  10925  11439   2988
#> 43: France 2020-09-10                       estimate  11302  11798   3014
#> 44: France 2020-09-11                       estimate  11685  12165   3039
#> 45: France 2020-09-12                       estimate  12069  12540   3075
#> 46: France 2020-09-13                       estimate  12474  12924   3134
#> 47: France 2020-09-14                       estimate  12850  13319   3227
#> 48: France 2020-09-15                       estimate  13265  13723   3361
#> 49: France 2020-09-16                       estimate  13619  14136   3535
#> 50: France 2020-09-17                       estimate  13977  14554   3742
#> 51: France 2020-09-18                       estimate  14374  14974   3972
#> 52: France 2020-09-19                       estimate  14754  15394   4212
#> 53: France 2020-09-20                       estimate  15169  15811   4460
#> 54: France 2020-09-21                       estimate  15524  16228   4722
#> 55: France 2020-09-22                       estimate  15900  16648   5012
#> 56: France 2020-09-23                       estimate  16223  17078   5348
#> 57: France 2020-09-24                       estimate  16587  17523   5744
#> 58: France 2020-09-25                       estimate  17032  17992   6204
#> 59: France 2020-09-26                       estimate  17494  18491   6730
#> 60: France 2020-09-27                       estimate  17935  19027   7318
#> 61: France 2020-09-28 estimate based on partial data  18453  19604   7968
#> 62: France 2020-09-29 estimate based on partial data  18935  20229   8683
#> 63: France 2020-09-30 estimate based on partial data  19419  20906   9468
#> 64: France 2020-10-01 estimate based on partial data  19929  21641  10336
#> 65: France 2020-10-02 estimate based on partial data  20615  22441  11305
#> 66: France 2020-10-03 estimate based on partial data  21191  23314  12398
#> 67: France 2020-10-04 estimate based on partial data  21865  24270  13652
#> 68: France 2020-10-05 estimate based on partial data  22565  25326  15116
#> 69: France 2020-10-06 estimate based on partial data  23448  26504  16864
#> 70: France 2020-10-07 estimate based on partial data  24276  27839  19013
#> 71: France 2020-10-08 estimate based on partial data  25278  29393  21765
#> 72: France 2020-10-09                       forecast  26262  31025  25153
#> 73: France 2020-10-10                       forecast  27154  32856  29480
#> 74: France 2020-10-11                       forecast  28043  34922  35055
#> 75: France 2020-10-12                       forecast  29008  37266  42294
#> 76: France 2020-10-13                       forecast  29920  39947  51747
#> 77: France 2020-10-14                       forecast  30993  43038  64141
#> 78: France 2020-10-15                       forecast  32085  46629  80447
#> 79: France 2020-10-16                       forecast  33161  50840 101956
#> 80: France 2020-10-17                       forecast  34293  55819 130393
#> 81: France 2020-10-18                       forecast  35665  61757 168073
#> 82: France 2020-10-19                       forecast  36998  68899 218103
#> 83: France 2020-10-20                       forecast  38483  77562 284665
#> 84: France 2020-10-21                       forecast  39816  88151 373396
#> 85: France 2020-10-22                       forecast  41107 101194 491903
#>     region       date                           type median   mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90      upper_90
#>  1:     1585     1725     1486     1874     1144 1.154631e-320
#>  2:     1639     1787     1566     1978     1301 1.268267e-320
#>  3:     1739     1897     1652     2090     1344 1.320637e-320
#>  4:     1831     1994     1755     2222     1390 1.372020e-320
#>  5:     1872     2045     1825     2313     1492 1.449095e-320
#>  6:     1978     2161     1848     2364     1564 1.519252e-320
#>  7:     2026     2221     1988     2537     1646 1.606701e-320
#>  8:     2177     2378     2085     2667     1723 1.688716e-320
#>  9:     2277     2492     2173     2792     1804 1.774684e-320
#> 10:     2362     2591     2297     2957     1907 1.867568e-320
#> 11:     2477     2720     2415     3122     2050 1.977745e-320
#> 12:     2667     2927     2487     3239     2115 2.067171e-320
#> 13:     2799     3072     2651     3446     2281 2.219343e-320
#> 14:     3009     3308     2714     3561     2278 2.292959e-320
#> 15:     3200     3530     2956     3876     2515 2.486632e-320
#> 16:     3297     3661     3050     4047     2615 2.613113e-320
#> 17:     3468     3855     3203     4272     2794 2.825067e-320
#> 18:     3597     4001     3365     4516     2880 2.984651e-320
#> 19:     3858     4285     3468     4691     2934 3.155103e-320
#> 20:     4052     4506     3675     4973     3043 3.342848e-320
#> 21:     4182     4654     3927     5302     3280 3.567648e-320
#> 22:     4370     4862     4060     5502     3327 3.718338e-320
#> 23:     4540     5055     4231     5735     3339 3.832961e-320
#> 24:     4667     5201     4354     5920     3714 4.103709e-320
#> 25:     4959     5520     4682     6321     3924 4.290960e-320
#> 26:     5124     5718     4782     6500     4028 4.443626e-320
#> 27:     5455     6082     4930     6725     4201 4.609138e-320
#> 28:     5713     6384     5189     7063     4396 4.758840e-320
#> 29:     6133     6837     5480     7451     4711 4.985616e-320
#> 30:     6532     7282     5636     7714     4905 5.150634e-320
#> 31:     6471     7230     5825     7995     5016 5.271186e-320
#> 32:     6736     7522     6148     8352     5193 5.451026e-320
#> 33:     6873     7674     6332     8597     5430 5.650629e-320
#> 34:     7180     8008     6553     8893     5651 5.896179e-320
#> 35:     7419     8300     6738     9198     5725 6.054280e-320
#> 36:     7987     8899     7137     9702     5808 6.263764e-320
#> 37:     8085     9036     7526    10181     6011 6.502398e-320
#> 38:     8249     9231     7612    10386     6263 6.782039e-320
#> 39:     8517     9542     7888    10793     6160 6.888263e-320
#> 40:     8636     9692     8084    11075     6381 7.177292e-320
#> 41:     9030    10116     8523    11625     6688 7.432724e-320
#> 42:     9476    10641     8894    12062     7038 7.698531e-320
#> 43:     9693    10901     9196    12455     7241 7.878865e-320
#> 44:    10165    11413     9659    13038     7758 8.224217e-320
#> 45:    10446    11741     9968    13460     7861 8.430242e-320
#> 46:    10561    11916    10297    13859     7940 8.580438e-320
#> 47:    11447    12823    10509    14203     8482 9.028556e-320
#> 48:    11423    12887    10365    14224     8783 9.308691e-320
#> 49:    11981    13536    10981    15044     9120 9.694062e-320
#> 50:    12029    13635    11283    15639     9203 1.006511e-319
#> 51:    12431    14091    10927    15537     8895 1.020987e-319
#> 52:    12973    14734    11243    16107     9327 1.078150e-319
#> 53:    13531    15414    11385    16548     9102 1.104830e-319
#> 54:    13992    15935    12366    17807     9425 1.165353e-319
#> 55:    13985    16028    12466    18222     9140 1.195293e-319
#> 56:    14598    16798    11645    17717     9240 1.246775e-319
#> 57:    13576    15980    12421    19017     8957 1.291241e-319
#> 58:    13405    16100    12535    19762     8664 1.334620e-319
#> 59:    13246    16100    11858    19725     8332 1.392228e-319
#> 60:    15455    18614    12045    20536     7268 1.415992e-319
#> 61:    15549    18999    11947    21071     6879 1.484025e-319
#> 62:    15564    19326    12080    21957     6938 1.595140e-319
#> 63:    16774    20781    12111    22831     6316 1.672412e-319
#> 64:    17280    21644    11250    22775     5443 1.738123e-319
#> 65:    17898    22567    11195    23592     5501 1.849238e-319
#> 66:    18464    23561    10811    24218     5375 1.963318e-319
#> 67:    15489    20963    11500    25863     5404 2.084908e-319
#> 68:    15472    21396    11204    26525     4755 2.185549e-319
#> 69:    14423    20818    11182    27581     5260 2.360497e-319
#> 70:    15100    21723    12330    30115     3741 2.432631e-319
#> 71:    18582    25808    12783    31755     3547 2.614694e-319
#> 72:    17051    24950    11975    32373     3051 2.792657e-319
#> 73:    12406    20707    12067    33909     2843 2.989641e-319
#> 74:    12281    21012     9295    32586     2615 3.180646e-319
#> 75:    11138    20380    10066    34780     1802 3.390575e-319
#> 76:    12394    22059     9610    35904     1120 3.578320e-319
#> 77:    13644    23835     6152    33964      854 3.851143e-319
#> 78:    13522    24294     5645    34789      588 4.108255e-319
#> 79:    13691    25079     5403    35873      168 4.347778e-319
#> 80:    10555    22603     4870    36804      449 4.624356e-319
#> 81:    10575    23316     5294    38755       96 4.966298e-319
#> 82:     5244    18444     4363    39444       79 5.304635e-319
#> 83:     5003    18656     3940    40649       64 5.661103e-319
#> 84:     3698    17742     3698    42434       52 6.055516e-319
#> 85:     3608    17935     3780    44278       43 6.501410e-319
#>     lower_20 upper_20 lower_50 upper_50 lower_90      upper_90
#> 
#> $casesReportData
#>     region       date                           type median  mean     sd
#>  1: France 2020-07-30                       estimate    841  1413   1741
#>  2: France 2020-07-31                       estimate    788  1453   1922
#>  3: France 2020-08-01                       estimate    891  1568   2104
#>  4: France 2020-08-02                       estimate    592  1050   1308
#>  5: France 2020-08-03                       estimate    562  1040   1437
#>  6: France 2020-08-04                       estimate   1005  1747   2159
#>  7: France 2020-08-05                       estimate    700  1249   1630
#>  8: France 2020-08-06                       estimate   1182  2143   2765
#>  9: France 2020-08-07                       estimate   1148  2016   2562
#> 10: France 2020-08-08                       estimate   1350  2338   2951
#> 11: France 2020-08-09                       estimate    872  1601   2126
#> 12: France 2020-08-10                       estimate    789  1398   1768
#> 13: France 2020-08-11                       estimate   1429  2553   3207
#> 14: France 2020-08-12                       estimate   1011  1765   2268
#> 15: France 2020-08-13                       estimate   1732  3040   3878
#> 16: France 2020-08-14                       estimate   1628  2912   3731
#> 17: France 2020-08-15                       estimate   1852  3365   4325
#> 18: France 2020-08-16                       estimate   1286  2241   2859
#> 19: France 2020-08-17                       estimate   1094  1938   2523
#> 20: France 2020-08-18                       estimate   2098  3657   4704
#> 21: France 2020-08-19                       estimate   1444  2551   3303
#> 22: France 2020-08-20                       estimate   2433  4191   5337
#> 23: France 2020-08-21                       estimate   2364  4296   5556
#> 24: France 2020-08-22                       estimate   2818  4913   6080
#> 25: France 2020-08-23                       estimate   1730  3153   4017
#> 26: France 2020-08-24                       estimate   1631  2873   3823
#> 27: France 2020-08-25                       estimate   2925  5119   6594
#> 28: France 2020-08-26                       estimate   2170  3833   4781
#> 29: France 2020-08-27                       estimate   3358  6234   9442
#> 30: France 2020-08-28                       estimate   3346  5921   7521
#> 31: France 2020-08-29                       estimate   3756  6663   8918
#> 32: France 2020-08-30                       estimate   2466  4451   5543
#> 33: France 2020-08-31                       estimate   2242  3989   5216
#> 34: France 2020-09-01                       estimate   3974  7195   9471
#> 35: France 2020-09-02                       estimate   2996  5089   6339
#> 36: France 2020-09-03                       estimate   4702  8312  10930
#> 37: France 2020-09-04                       estimate   4515  7884  10038
#> 38: France 2020-09-05                       estimate   5088  8977  11777
#> 39: France 2020-09-06                       estimate   3354  5961   7727
#> 40: France 2020-09-07                       estimate   2999  5419   7407
#> 41: France 2020-09-08                       estimate   5218  9683  13182
#> 42: France 2020-09-09                       estimate   3690  6730   8773
#> 43: France 2020-09-10                       estimate   6330 11239  14233
#> 44: France 2020-09-11                       estimate   5761 10506  14011
#> 45: France 2020-09-12                       estimate   6434 11517  14762
#> 46: France 2020-09-13                       estimate   4271  7651  10321
#> 47: France 2020-09-14                       estimate   3810  6793   8779
#> 48: France 2020-09-15                       estimate   6879 12625  16546
#> 49: France 2020-09-16                       estimate   4516  8100  10433
#> 50: France 2020-09-17                       estimate   7732 13894  18030
#> 51: France 2020-09-18                       estimate   7479 13208  17174
#> 52: France 2020-09-19                       estimate   8162 14540  18363
#> 53: France 2020-09-20                       estimate   5282  9512  12128
#> 54: France 2020-09-21                       estimate   4910  8655  11356
#> 55: France 2020-09-22                       estimate   8338 15192  20118
#> 56: France 2020-09-23                       estimate   5676 10559  14230
#> 57: France 2020-09-24                       estimate   9902 17717  23504
#> 58: France 2020-09-25                       estimate   8996 16525  21080
#> 59: France 2020-09-26                       estimate  10408 18529  25361
#> 60: France 2020-09-27                       estimate   6472 11850  15820
#> 61: France 2020-09-28 estimate based on partial data   5716 10208  13074
#> 62: France 2020-09-29 estimate based on partial data  10440 19387  26253
#> 63: France 2020-09-30 estimate based on partial data   6984 12516  16831
#> 64: France 2020-10-01 estimate based on partial data  11260 20890  28047
#> 65: France 2020-10-02 estimate based on partial data  11320 20502  28880
#> 66: France 2020-10-03 estimate based on partial data  11965 21982  29768
#> 67: France 2020-10-04 estimate based on partial data   8067 14558  19246
#> 68: France 2020-10-05 estimate based on partial data   6952 13162  18293
#> 69: France 2020-10-06 estimate based on partial data  11488 23228  32945
#> 70: France 2020-10-07 estimate based on partial data   8244 15732  22669
#> 71: France 2020-10-08 estimate based on partial data  13762 26115  38830
#> 72: France 2020-10-09                       forecast  12982 25835  42200
#> 73: France 2020-10-10                       forecast  14460 29066  44476
#> 74: France 2020-10-11                       forecast   8944 18505  29809
#> 75: France 2020-10-12                       forecast   8287 16272  26894
#> 76: France 2020-10-13                       forecast  13404 29435  49885
#> 77: France 2020-10-14                       forecast   9494 20312  32613
#> 78: France 2020-10-15                       forecast  16692 36799  68328
#> 79: France 2020-10-16                       forecast  14426 33856  61074
#> 80: France 2020-10-17                       forecast  17996 40709  89364
#> 81: France 2020-10-18                       forecast  11302 26406  47504
#> 82: France 2020-10-19                       forecast  10026 23561  47198
#> 83: France 2020-10-20                       forecast  17654 44968 117449
#> 84: France 2020-10-21                       forecast  12757 31786  61065
#> 85: France 2020-10-22                       forecast  19386 61773 230778
#>     region       date                           type median  mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:        0      214        0      841        0     3445
#>  2:        0      184        0      789        0     3609
#>  3:        0      220        0      891        0     3799
#>  4:        0      145        0      592        0     2571
#>  5:        0      125        0      562        0     2550
#>  6:        0      251        0     1005        0     4330
#>  7:        0      176        0      700        0     3055
#>  8:        0      299        0     1183        0     5253
#>  9:        0      284        0     1148        0     4977
#> 10:        0      295        0     1350        0     5883
#> 11:        0      210        0      872        0     3885
#> 12:        0      207        0      790        0     3363
#> 13:        0      363        0     1430        0     6245
#> 14:        0      237        0     1011        0     4278
#> 15:        0      433        0     1733        0     7464
#> 16:        0      401        0     1628        0     7076
#> 17:        0      451        0     1852        0     8252
#> 18:        0      306        0     1287        0     5480
#> 19:        0      264        0     1094        0     4729
#> 20:        0      466        0     2098        0     9154
#> 21:        0      337        0     1444        0     6305
#> 22:        0      620        0     2433        0     9928
#> 23:        0      577        0     2365        0    10630
#> 24:        0      707        0     2819        0    12107
#> 25:        0      453        0     1730        0     7833
#> 26:        0      372        0     1632        0     6886
#> 27:        0      714        0     2926        0    12647
#> 28:        0      532        0     2170        0     9597
#> 29:        0      776        0     3360        0    14798
#> 30:        0      840        0     3347        0    14574
#> 31:        0      949        0     3757        0    15785
#> 32:        0      623        0     2467        0    11251
#> 33:        0      523        0     2242        0     9824
#> 34:        0      955        0     3975        0    17782
#> 35:        0      737        0     2996        0    12469
#> 36:        0     1191        0     4704        0    20240
#> 37:        0     1060        0     4517        0    19649
#> 38:        0     1205        0     5090        0    22126
#> 39:        0      848        0     3354        0    14551
#> 40:        0      687        0     3001        0    12691
#> 41:        0     1280        0     5223        0    24183
#> 42:        0      944        0     3690        0    16650
#> 43:        0     1473        0     6333        0    27455
#> 44:        0     1421        0     5762        0    25920
#> 45:        0     1589        0     6437        0    28318
#> 46:        0     1107        0     4276        0    18901
#> 47:        0      927        0     3811        0    16754
#> 48:        0     1696        0     6890        0    30502
#> 49:        0     1117        0     4517        0    19665
#> 50:        0     1939        0     7740        0    33265
#> 51:        0     1873        0     7480        0    32381
#> 52:        0     2063        0     8170        0    35981
#> 53:        0     1241        0     5284        0    23458
#> 54:        0     1182        0     4914        0    21567
#> 55:        0     2028        0     8344        0    37445
#> 56:        0     1341        0     5676        0    26669
#> 57:        0     2332        0     9902        0    44038
#> 58:        0     2193        0     9004        0    40945
#> 59:        0     2515        0    10409        0    45955
#> 60:        0     1704        0     6479        0    27919
#> 61:        0     1341        0     5719        0    25038
#> 62:        0     2686        0    10441        0    48689
#> 63:        0     1686        0     6985        0    30519
#> 64:        0     2957        0    11261        0    51764
#> 65:        0     2711        0    11332        0    50410
#> 66:        2     3058        2    11978        2    53791
#> 67:        0     1958        0     8068        0    36403
#> 68:        0     1709        0     6952        0    32868
#> 69:        0     2785        0    11489        0    57941
#> 70:        0     1918        0     8245        0    38432
#> 71:        0     3137        0    13764        0    61575
#> 72:        0     2973        0    12982        0    62502
#> 73:        0     3222        0    14460        0    73232
#> 74:        0     2281        0     8947        0    44697
#> 75:        0     1938        0     8287        0    40310
#> 76:        0     3192        0    13405        0    73031
#> 77:        0     2214        0     9494        0    50168
#> 78:        0     3817        0    16693        0    87047
#> 79:        0     3195        0    14431        0    82211
#> 80:        0     4095        0    18000        0    96280
#> 81:        0     2571        0    11302        0    65246
#> 82:        0     2217        0    10037        0    56169
#> 83:        0     3576        0    17690        0   102692
#> 84:        0     2525        0    12776        0    80231
#> 85:        0     4178        0    19387        0   129083
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-07-16     927
#>  2: France 2020-07-17     534
#>  3: France 2020-07-18     836
#>  4: France 2020-07-19       0
#>  5: France 2020-07-20       0
#>  6: France 2020-07-21    2080
#>  7: France 2020-07-22     584
#>  8: France 2020-07-23     998
#>  9: France 2020-07-24    1062
#> 10: France 2020-07-25    1130
#> 11: France 2020-07-26       0
#> 12: France 2020-07-27       0
#> 13: France 2020-07-28    2551
#> 14: France 2020-07-29     725
#> 15: France 2020-07-30    1392
#> 16: France 2020-07-31    1377
#> 17: France 2020-08-01    1346
#> 18: France 2020-08-02       0
#> 19: France 2020-08-03       0
#> 20: France 2020-08-04    3376
#> 21: France 2020-08-05    1039
#> 22: France 2020-08-06    1695
#> 23: France 2020-08-07    1604
#> 24: France 2020-08-08    2288
#> 25: France 2020-08-09       0
#> 26: France 2020-08-10       0
#> 27: France 2020-08-11    4854
#> 28: France 2020-08-12    1397
#> 29: France 2020-08-13    2524
#> 30: France 2020-08-14    2669
#> 31: France 2020-08-15    2846
#> 32: France 2020-08-16    3310
#> 33: France 2020-08-17    3015
#> 34: France 2020-08-18     493
#> 35: France 2020-08-19    2238
#> 36: France 2020-08-20    3776
#> 37: France 2020-08-21    4771
#> 38: France 2020-08-22    4586
#> 39: France 2020-08-23    3602
#> 40: France 2020-08-24    4897
#> 41: France 2020-08-25    1955
#> 42: France 2020-08-26    3304
#> 43: France 2020-08-27    5429
#> 44: France 2020-08-28    6111
#> 45: France 2020-08-29    7379
#> 46: France 2020-08-30    5453
#> 47: France 2020-08-31    5413
#> 48: France 2020-09-01    3082
#> 49: France 2020-09-02    4982
#> 50: France 2020-09-03    7017
#> 51: France 2020-09-04    7157
#> 52: France 2020-09-05    8975
#> 53: France 2020-09-06    8550
#> 54: France 2020-09-07    7071
#> 55: France 2020-09-08    4203
#> 56: France 2020-09-09    6544
#> 57: France 2020-09-10    8577
#> 58: France 2020-09-11    9843
#> 59: France 2020-09-12    9406
#> 60: France 2020-09-13   10561
#> 61: France 2020-09-14    7183
#> 62: France 2020-09-15    6158
#> 63: France 2020-09-16    7852
#> 64: France 2020-09-17    9784
#> 65: France 2020-09-18   10593
#> 66: France 2020-09-19   13215
#> 67: France 2020-09-20   13498
#> 68: France 2020-09-21   11569
#> 69: France 2020-09-22    4298
#> 70: France 2020-09-23   10008
#> 71: France 2020-09-24   13072
#> 72: France 2020-09-25   16096
#> 73: France 2020-09-26   15797
#> 74: France 2020-09-27   14412
#> 75: France 2020-09-28   11123
#> 76: France 2020-09-29    4070
#> 77: France 2020-09-30    8051
#> 78: France 2020-10-01   12845
#> 79: France 2020-10-02   13970
#> 80: France 2020-10-03   12148
#> 81: France 2020-10-04   16972
#> 82: France 2020-10-05   12545
#> 83: France 2020-10-06    5104
#> 84: France 2020-10-07   10489
#> 85: France 2020-10-08   18746
#>     region       date confirm
#>