Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

rtData

#> $summaryData
#>              region New confirmed cases by infection date
#>   1:    Afghanistan                         46 (9 -- 133)
#>   2:        Albania                      245 (140 -- 375)
#>   3:        Algeria                       173 (94 -- 295)
#>   4:        Andorra                         78 (5 -- 199)
#>   5:         Angola                       190 (60 -- 355)
#>  ---                                                     
#> 192:        Vietnam                           5 (0 -- 31)
#> 193: Western Sahara                          1 (0 -- 289)
#> 194:          Yemen                            0 (0 -- 1)
#> 195:         Zambia                         43 (15 -- 79)
#> 196:       Zimbabwe                          14 (2 -- 52)
#>      Expected change in daily cases Effective reproduction no.
#>   1:                         Unsure             1 (0.8 -- 1.3)
#>   2:              Likely increasing           1.1 (0.9 -- 1.3)
#>   3:                         Unsure           1.1 (0.9 -- 1.3)
#>   4:              Likely increasing           1.1 (0.8 -- 1.4)
#>   5:              Likely increasing           1.1 (0.9 -- 1.3)
#>  ---                                                          
#> 192:                         Unsure           1.2 (0.2 -- 2.5)
#> 193:                         Unsure             0.6 (0 -- 1.4)
#> 194:              Likely decreasing           0.8 (0.5 -- 1.2)
#> 195:              Likely decreasing           0.9 (0.8 -- 1.1)
#> 196:                         Unsure             1 (0.6 -- 1.5)
#>             Rate of growth Doubling/halving time (days)
#>   1:     0 (-0.06 -- 0.09)         -1239 (7.4 -- -11.9)
#>   2:  0.03 (-0.02 -- 0.08)          25.5 (8.5 -- -44.6)
#>   3:  0.01 (-0.03 -- 0.08)            47.6 (8.6 -- -24)
#>   4:  0.04 (-0.05 -- 0.11)          19.4 (6.4 -- -14.8)
#>   5:  0.02 (-0.03 -- 0.07)          32.2 (9.6 -- -22.3)
#>  ---                                                   
#> 192:   0.05 (-0.2 -- 0.39)           14.2 (1.8 -- -3.5)
#> 193:  -0.1 (-0.36 -- 0.14)           -6.7 (4.8 -- -1.9)
#> 194: -0.05 (-0.14 -- 0.04)           -14.3 (15.8 -- -5)
#> 195: -0.02 (-0.06 -- 0.04)            -33 (19.3 -- -11)
#> 196: -0.01 (-0.11 -- 0.14)         -121.9 (4.8 -- -6.4)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-07 estimate   0.96 0.95 0.05     0.95     0.97
#>     2: Afghanistan 2020-08-08 estimate   0.96 0.95 0.05     0.95     0.97
#>     3: Afghanistan 2020-08-09 estimate   0.95 0.95 0.05     0.95     0.97
#>     4: Afghanistan 2020-08-10 estimate   0.95 0.95 0.06     0.95     0.97
#>     5: Afghanistan 2020-08-11 estimate   0.95 0.95 0.06     0.95     0.97
#>    ---                                                                   
#> 16318:    Zimbabwe 2020-10-26 forecast   0.98 1.06 0.33     0.84     0.93
#> 16319:    Zimbabwe 2020-10-27 forecast   0.98 1.06 0.33     0.84     0.93
#> 16320:    Zimbabwe 2020-10-28 forecast   0.98 1.06 0.33     0.84     0.93
#> 16321:    Zimbabwe 2020-10-29 forecast   0.98 1.06 0.33     0.84     0.93
#> 16322:    Zimbabwe 2020-10-30 forecast   0.98 1.06 0.33     0.84     0.93
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.94     0.98     0.87     1.03
#>     2:     0.94     0.99     0.86     1.02
#>     3:     0.94     0.98     0.86     1.03
#>     4:     0.94     0.98     0.85     1.02
#>     5:     0.94     0.99     0.85     1.03
#>    ---                                    
#> 16318:     0.79     1.06     0.60     1.53
#> 16319:     0.79     1.06     0.60     1.53
#> 16320:     0.79     1.06     0.60     1.53
#> 16321:     0.79     1.06     0.60     1.53
#> 16322:     0.79     1.06     0.60     1.53
#> 
#> $casesInfectionData
#>             region       date     type median mean    sd lower_20 upper_20
#>     1: Afghanistan 2020-08-07 estimate     60   62    14       57       62
#>     2: Afghanistan 2020-08-08 estimate     60   61    14       57       62
#>     3: Afghanistan 2020-08-09 estimate     59   60    14       57       62
#>     4: Afghanistan 2020-08-10 estimate     58   59    14       55       61
#>     5: Afghanistan 2020-08-11 estimate     57   58    13       54       59
#>    ---                                                                    
#> 16318:    Zimbabwe 2020-10-26 forecast     13  185  1821        3        7
#> 16319:    Zimbabwe 2020-10-27 forecast     13  253  2858        3        6
#> 16320:    Zimbabwe 2020-10-28 forecast     13  353  4504        2        6
#> 16321:    Zimbabwe 2020-10-29 forecast     13  501  7123        3        6
#> 16322:    Zimbabwe 2020-10-30 forecast     13  721 11297        2        6
#>        lower_50 upper_50 lower_90 upper_90
#>     1:       55       69       40       81
#>     2:       53       68       39       81
#>     3:       51       66       38       79
#>     4:       50       65       37       77
#>     5:       51       65       37       77
#>    ---                                    
#> 16318:        1       13        0      150
#> 16319:        1       13        0      169
#> 16320:        1       13        0      190
#> 16321:        1       13        0      210
#> 16322:        0       13        0      235
#> 
#> $casesReportData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-07 estimate     26   53   78        0        4
#>     2: Afghanistan 2020-08-08 estimate     26   58   86        0        5
#>     3: Afghanistan 2020-08-09 estimate     33   70   99        0        6
#>     4: Afghanistan 2020-08-10 estimate     26   56   83        0        5
#>     5: Afghanistan 2020-08-11 estimate     40   83  121        0        7
#>    ---                                                                   
#> 16318:    Zimbabwe 2020-10-26 forecast      8   30  120        1        3
#> 16319:    Zimbabwe 2020-10-27 forecast     14   61  296        0        4
#> 16320:    Zimbabwe 2020-10-28 forecast     12   62  465        0        3
#> 16321:    Zimbabwe 2020-10-29 forecast     10   65  412        0        3
#> 16322:    Zimbabwe 2020-10-30 forecast     12   96 1022        0        3
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0       26        0      136
#>     2:        0       26        0      149
#>     3:        0       33        0      188
#>     4:        0       27        0      144
#>     5:        0       40        0      217
#>    ---                                    
#> 16318:        0        8        0       53
#> 16319:        0       14        0       94
#> 16320:        0       12        0       80
#> 16321:        0       10        0       72
#> 16322:        0       12        0       94
#> 
#> $obsCasesData
#>                region       date confirm
#>     1:    Afghanistan 2020-07-24      13
#>     2:        Albania 2020-07-24     108
#>     3:        Algeria 2020-07-24     612
#>     4:        Andorra 2020-07-24       0
#>     5:         Angola 2020-07-24      33
#>    ---                                  
#> 17762:        Vietnam 2020-10-16       2
#> 17763: Western Sahara 2020-10-16       0
#> 17764:          Yemen 2020-10-16       0
#> 17765:         Zambia 2020-10-16      43
#> 17766:       Zimbabwe 2020-10-16      20
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France                 30001 (6138 -- 60873)              Likely increasing
#>    Effective reproduction no.       Rate of growth Doubling/halving time (days)
#> 1:           1.1 (0.9 -- 1.4) 0.03 (-0.04 -- 0.09)            22 (7.5 -- -16.7)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-08-07                       estimate   1.22 1.22 0.05     1.20
#>  2: France 2020-08-08                       estimate   1.22 1.22 0.06     1.20
#>  3: France 2020-08-09                       estimate   1.21 1.22 0.06     1.20
#>  4: France 2020-08-10                       estimate   1.21 1.22 0.06     1.20
#>  5: France 2020-08-11                       estimate   1.21 1.22 0.06     1.19
#>  6: France 2020-08-12                       estimate   1.21 1.21 0.06     1.18
#>  7: France 2020-08-13                       estimate   1.21 1.21 0.06     1.18
#>  8: France 2020-08-14                       estimate   1.20 1.21 0.06     1.18
#>  9: France 2020-08-15                       estimate   1.20 1.21 0.06     1.18
#> 10: France 2020-08-16                       estimate   1.20 1.20 0.06     1.18
#> 11: France 2020-08-17                       estimate   1.19 1.20 0.06     1.18
#> 12: France 2020-08-18                       estimate   1.19 1.20 0.06     1.17
#> 13: France 2020-08-19                       estimate   1.19 1.19 0.06     1.18
#> 14: France 2020-08-20                       estimate   1.19 1.19 0.06     1.17
#> 15: France 2020-08-21                       estimate   1.18 1.19 0.06     1.17
#> 16: France 2020-08-22                       estimate   1.18 1.18 0.06     1.17
#> 17: France 2020-08-23                       estimate   1.18 1.18 0.06     1.16
#> 18: France 2020-08-24                       estimate   1.17 1.18 0.06     1.16
#> 19: France 2020-08-25                       estimate   1.17 1.17 0.06     1.16
#> 20: France 2020-08-26                       estimate   1.17 1.17 0.06     1.17
#> 21: France 2020-08-27                       estimate   1.17 1.16 0.06     1.17
#> 22: France 2020-08-28                       estimate   1.16 1.16 0.06     1.16
#> 23: France 2020-08-29                       estimate   1.16 1.16 0.06     1.15
#> 24: France 2020-08-30                       estimate   1.16 1.15 0.06     1.16
#> 25: France 2020-08-31                       estimate   1.15 1.15 0.06     1.15
#> 26: France 2020-09-01                       estimate   1.15 1.15 0.06     1.15
#> 27: France 2020-09-02                       estimate   1.15 1.14 0.06     1.15
#> 28: France 2020-09-03                       estimate   1.14 1.14 0.06     1.15
#> 29: France 2020-09-04                       estimate   1.14 1.14 0.06     1.14
#> 30: France 2020-09-05                       estimate   1.14 1.13 0.06     1.14
#> 31: France 2020-09-06                       estimate   1.13 1.13 0.06     1.13
#> 32: France 2020-09-07                       estimate   1.13 1.13 0.06     1.13
#> 33: France 2020-09-08                       estimate   1.13 1.12 0.06     1.12
#> 34: France 2020-09-09                       estimate   1.13 1.12 0.06     1.12
#> 35: France 2020-09-10                       estimate   1.12 1.12 0.06     1.11
#> 36: France 2020-09-11                       estimate   1.12 1.12 0.06     1.11
#> 37: France 2020-09-12                       estimate   1.12 1.11 0.06     1.11
#> 38: France 2020-09-13                       estimate   1.12 1.11 0.06     1.11
#> 39: France 2020-09-14                       estimate   1.12 1.11 0.06     1.11
#> 40: France 2020-09-15                       estimate   1.11 1.11 0.06     1.12
#> 41: France 2020-09-16                       estimate   1.11 1.11 0.06     1.10
#> 42: France 2020-09-17                       estimate   1.11 1.11 0.06     1.11
#> 43: France 2020-09-18                       estimate   1.11 1.10 0.06     1.11
#> 44: France 2020-09-19                       estimate   1.11 1.10 0.06     1.10
#> 45: France 2020-09-20                       estimate   1.11 1.10 0.06     1.10
#> 46: France 2020-09-21                       estimate   1.11 1.10 0.06     1.10
#> 47: France 2020-09-22                       estimate   1.11 1.10 0.06     1.09
#> 48: France 2020-09-23                       estimate   1.11 1.10 0.07     1.10
#> 49: France 2020-09-24                       estimate   1.11 1.10 0.07     1.10
#> 50: France 2020-09-25                       estimate   1.11 1.10 0.07     1.10
#> 51: France 2020-09-26                       estimate   1.11 1.10 0.07     1.10
#> 52: France 2020-09-27                       estimate   1.11 1.10 0.07     1.10
#> 53: France 2020-09-28                       estimate   1.11 1.10 0.07     1.12
#> 54: France 2020-09-29                       estimate   1.11 1.10 0.07     1.11
#> 55: France 2020-09-30                       estimate   1.11 1.11 0.08     1.12
#> 56: France 2020-10-01                       estimate   1.11 1.11 0.08     1.12
#> 57: France 2020-10-02                       estimate   1.11 1.11 0.08     1.12
#> 58: France 2020-10-03                       estimate   1.11 1.11 0.09     1.13
#> 59: France 2020-10-04                       estimate   1.11 1.11 0.09     1.13
#> 60: France 2020-10-05                       estimate   1.11 1.11 0.10     1.13
#> 61: France 2020-10-06 estimate based on partial data   1.11 1.11 0.10     1.13
#> 62: France 2020-10-07 estimate based on partial data   1.11 1.11 0.11     1.13
#> 63: France 2020-10-08 estimate based on partial data   1.11 1.11 0.12     1.13
#> 64: France 2020-10-09 estimate based on partial data   1.11 1.12 0.12     1.13
#> 65: France 2020-10-10 estimate based on partial data   1.11 1.12 0.13     1.13
#> 66: France 2020-10-11 estimate based on partial data   1.12 1.12 0.14     1.12
#> 67: France 2020-10-12 estimate based on partial data   1.12 1.12 0.15     1.12
#> 68: France 2020-10-13 estimate based on partial data   1.12 1.12 0.16     1.12
#> 69: France 2020-10-14 estimate based on partial data   1.12 1.12 0.17     1.13
#> 70: France 2020-10-15 estimate based on partial data   1.12 1.12 0.18     1.13
#> 71: France 2020-10-16 estimate based on partial data   1.12 1.12 0.19     1.12
#> 72: France 2020-10-17                       forecast   1.12 1.12 0.19     1.12
#> 73: France 2020-10-18                       forecast   1.12 1.12 0.19     1.12
#> 74: France 2020-10-19                       forecast   1.12 1.12 0.19     1.12
#> 75: France 2020-10-20                       forecast   1.12 1.12 0.19     1.12
#> 76: France 2020-10-21                       forecast   1.12 1.12 0.19     1.12
#> 77: France 2020-10-22                       forecast   1.12 1.12 0.19     1.12
#> 78: France 2020-10-23                       forecast   1.12 1.12 0.19     1.12
#> 79: France 2020-10-24                       forecast   1.12 1.12 0.19     1.12
#> 80: France 2020-10-25                       forecast   1.12 1.12 0.19     1.12
#> 81: France 2020-10-26                       forecast   1.12 1.12 0.19     1.12
#> 82: France 2020-10-27                       forecast   1.12 1.12 0.19     1.12
#> 83: France 2020-10-28                       forecast   1.12 1.12 0.19     1.12
#> 84: France 2020-10-29                       forecast   1.12 1.12 0.19     1.12
#> 85: France 2020-10-30                       forecast   1.12 1.12 0.19     1.12
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.22     1.18     1.24     1.15     1.31
#>  2:     1.22     1.18     1.24     1.14     1.31
#>  3:     1.22     1.18     1.24     1.14     1.31
#>  4:     1.22     1.18     1.24     1.13     1.31
#>  5:     1.21     1.17     1.23     1.13     1.31
#>  6:     1.21     1.17     1.23     1.12     1.31
#>  7:     1.20     1.17     1.22     1.12     1.30
#>  8:     1.20     1.17     1.23     1.11     1.30
#>  9:     1.20     1.16     1.22     1.12     1.30
#> 10:     1.20     1.16     1.22     1.11     1.30
#> 11:     1.20     1.16     1.22     1.11     1.30
#> 12:     1.19     1.16     1.21     1.11     1.29
#> 13:     1.20     1.16     1.21     1.10     1.29
#> 14:     1.19     1.15     1.21     1.11     1.30
#> 15:     1.19     1.15     1.20     1.10     1.28
#> 16:     1.19     1.15     1.20     1.10     1.28
#> 17:     1.18     1.15     1.20     1.09     1.27
#> 18:     1.18     1.15     1.20     1.08     1.27
#> 19:     1.18     1.14     1.20     1.08     1.27
#> 20:     1.19     1.14     1.20     1.07     1.26
#> 21:     1.19     1.13     1.19     1.07     1.25
#> 22:     1.18     1.13     1.19     1.07     1.25
#> 23:     1.17     1.13     1.19     1.06     1.25
#> 24:     1.18     1.13     1.19     1.06     1.24
#> 25:     1.17     1.13     1.19     1.06     1.24
#> 26:     1.17     1.13     1.19     1.05     1.23
#> 27:     1.18     1.12     1.18     1.05     1.23
#> 28:     1.17     1.12     1.19     1.04     1.23
#> 29:     1.17     1.11     1.18     1.04     1.22
#> 30:     1.17     1.11     1.18     1.04     1.22
#> 31:     1.15     1.11     1.17     1.03     1.21
#> 32:     1.15     1.11     1.17     1.03     1.21
#> 33:     1.15     1.10     1.17     1.03     1.21
#> 34:     1.14     1.10     1.17     1.02     1.21
#> 35:     1.14     1.09     1.16     1.02     1.21
#> 36:     1.13     1.09     1.16     1.02     1.21
#> 37:     1.14     1.09     1.16     1.02     1.20
#> 38:     1.14     1.09     1.16     1.02     1.20
#> 39:     1.14     1.08     1.16     1.01     1.20
#> 40:     1.14     1.09     1.16     1.01     1.20
#> 41:     1.13     1.09     1.16     1.02     1.21
#> 42:     1.14     1.08     1.16     1.01     1.20
#> 43:     1.14     1.08     1.15     1.00     1.20
#> 44:     1.13     1.08     1.15     1.00     1.20
#> 45:     1.13     1.08     1.15     1.00     1.20
#> 46:     1.12     1.08     1.15     1.00     1.20
#> 47:     1.12     1.08     1.16     1.00     1.20
#> 48:     1.13     1.08     1.16     0.99     1.20
#> 49:     1.13     1.07     1.16     0.99     1.20
#> 50:     1.13     1.07     1.16     0.99     1.20
#> 51:     1.13     1.07     1.16     0.99     1.21
#> 52:     1.13     1.07     1.16     0.98     1.20
#> 53:     1.15     1.07     1.16     0.98     1.20
#> 54:     1.15     1.07     1.16     0.99     1.21
#> 55:     1.16     1.07     1.17     0.98     1.21
#> 56:     1.15     1.08     1.18     0.98     1.22
#> 57:     1.16     1.07     1.17     0.97     1.22
#> 58:     1.17     1.06     1.17     0.96     1.22
#> 59:     1.17     1.07     1.18     0.97     1.25
#> 60:     1.17     1.07     1.18     0.95     1.24
#> 61:     1.17     1.06     1.18     0.95     1.27
#> 62:     1.18     1.06     1.17     0.94     1.28
#> 63:     1.18     1.06     1.18     0.93     1.29
#> 64:     1.18     1.06     1.19     0.93     1.30
#> 65:     1.18     1.06     1.19     0.91     1.31
#> 66:     1.17     1.06     1.19     0.91     1.33
#> 67:     1.17     1.06     1.19     0.91     1.34
#> 68:     1.17     1.04     1.18     0.88     1.33
#> 69:     1.18     1.05     1.19     0.86     1.33
#> 70:     1.18     1.04     1.19     0.86     1.35
#> 71:     1.18     1.04     1.19     0.85     1.35
#> 72:     1.18     1.04     1.19     0.85     1.35
#> 73:     1.18     1.04     1.19     0.85     1.35
#> 74:     1.18     1.04     1.19     0.85     1.35
#> 75:     1.18     1.04     1.19     0.85     1.35
#> 76:     1.18     1.04     1.19     0.85     1.35
#> 77:     1.18     1.04     1.19     0.85     1.35
#> 78:     1.18     1.04     1.19     0.85     1.35
#> 79:     1.18     1.04     1.19     0.85     1.35
#> 80:     1.18     1.04     1.19     0.85     1.35
#> 81:     1.18     1.04     1.19     0.85     1.35
#> 82:     1.18     1.04     1.19     0.85     1.35
#> 83:     1.18     1.04     1.19     0.85     1.35
#> 84:     1.18     1.04     1.19     0.85     1.35
#> 85:     1.18     1.04     1.19     0.85     1.35
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData
#>     region       date                           type median    mean        sd
#>  1: France 2020-08-07                       estimate   2541    2628       536
#>  2: France 2020-08-08                       estimate   2687    2777       571
#>  3: France 2020-08-09                       estimate   2841    2934       607
#>  4: France 2020-08-10                       estimate   2999    3097       645
#>  5: France 2020-08-11                       estimate   3167    3268       684
#>  6: France 2020-08-12                       estimate   3336    3447       725
#>  7: France 2020-08-13                       estimate   3514    3632       769
#>  8: France 2020-08-14                       estimate   3699    3825       815
#>  9: France 2020-08-15                       estimate   3894    4025       862
#> 10: France 2020-08-16                       estimate   4097    4233       912
#> 11: France 2020-08-17                       estimate   4303    4447       964
#> 12: France 2020-08-18                       estimate   4515    4668      1018
#> 13: France 2020-08-19                       estimate   4741    4896      1073
#> 14: France 2020-08-20                       estimate   4975    5130      1130
#> 15: France 2020-08-21                       estimate   5205    5371      1187
#> 16: France 2020-08-22                       estimate   5445    5617      1244
#> 17: France 2020-08-23                       estimate   5686    5868      1301
#> 18: France 2020-08-24                       estimate   5931    6125      1359
#> 19: France 2020-08-25                       estimate   6192    6386      1417
#> 20: France 2020-08-26                       estimate   6445    6652      1475
#> 21: France 2020-08-27                       estimate   6710    6923      1532
#> 22: France 2020-08-28                       estimate   6966    7198      1588
#> 23: France 2020-08-29                       estimate   7236    7476      1643
#> 24: France 2020-08-30                       estimate   7512    7758      1698
#> 25: France 2020-08-31                       estimate   7782    8043      1751
#> 26: France 2020-09-01                       estimate   8061    8332      1805
#> 27: France 2020-09-02                       estimate   8348    8624      1858
#> 28: France 2020-09-03                       estimate   8616    8918      1911
#> 29: France 2020-09-04                       estimate   8888    9215      1964
#> 30: France 2020-09-05                       estimate   9184    9515      2018
#> 31: France 2020-09-06                       estimate   9492    9817      2071
#> 32: France 2020-09-07                       estimate   9804   10122      2124
#> 33: France 2020-09-08                       estimate  10094   10430      2178
#> 34: France 2020-09-09                       estimate  10401   10740      2233
#> 35: France 2020-09-10                       estimate  10725   11054      2289
#> 36: France 2020-09-11                       estimate  11062   11371      2347
#> 37: France 2020-09-12                       estimate  11372   11693      2407
#> 38: France 2020-09-13                       estimate  11694   12018      2470
#> 39: France 2020-09-14                       estimate  12014   12349      2537
#> 40: France 2020-09-15                       estimate  12350   12685      2607
#> 41: France 2020-09-16                       estimate  12683   13027      2683
#> 42: France 2020-09-17                       estimate  13039   13377      2764
#> 43: France 2020-09-18                       estimate  13405   13734      2853
#> 44: France 2020-09-19                       estimate  13755   14101      2950
#> 45: France 2020-09-20                       estimate  14134   14478      3057
#> 46: France 2020-09-21                       estimate  14508   14867      3177
#> 47: France 2020-09-22                       estimate  14881   15269      3310
#> 48: France 2020-09-23                       estimate  15257   15685      3458
#> 49: France 2020-09-24                       estimate  15646   16116      3623
#> 50: France 2020-09-25                       estimate  16068   16564      3805
#> 51: France 2020-09-26                       estimate  16484   17030      4006
#> 52: France 2020-09-27                       estimate  16927   17516      4228
#> 53: France 2020-09-28                       estimate  17391   18023      4473
#> 54: France 2020-09-29                       estimate  17875   18552      4744
#> 55: France 2020-09-30                       estimate  18379   19107      5045
#> 56: France 2020-10-01                       estimate  18907   19688      5381
#> 57: France 2020-10-02                       estimate  19466   20300      5758
#> 58: France 2020-10-03                       estimate  20036   20944      6185
#> 59: France 2020-10-04                       estimate  20610   21626      6669
#> 60: France 2020-10-05                       estimate  21242   22349      7221
#> 61: France 2020-10-06 estimate based on partial data  21847   23118      7856
#> 62: France 2020-10-07 estimate based on partial data  22624   23941      8592
#> 63: France 2020-10-08 estimate based on partial data  23322   24825      9455
#> 64: France 2020-10-09 estimate based on partial data  24025   25780     10488
#> 65: France 2020-10-10 estimate based on partial data  24794   26820     11766
#> 66: France 2020-10-11 estimate based on partial data  25549   27965     13442
#> 67: France 2020-10-12 estimate based on partial data  26435   29241     15847
#> 68: France 2020-10-13 estimate based on partial data  27250   30693     19721
#> 69: France 2020-10-14 estimate based on partial data  28112   32391     26672
#> 70: France 2020-10-15 estimate based on partial data  29041   34457     39946
#> 71: France 2020-10-16 estimate based on partial data  30001   37104     65711
#> 72: France 2020-10-17                       forecast  30997   40472    110569
#> 73: France 2020-10-18                       forecast  31969   45084    190804
#> 74: France 2020-10-19                       forecast  32861   51720    333901
#> 75: France 2020-10-20                       forecast  33776   61727    589333
#> 76: France 2020-10-21                       forecast  34782   77436   1046336
#> 77: France 2020-10-22                       forecast  35921  102916   1866194
#> 78: France 2020-10-23                       forecast  37070  145294   3340958
#> 79: France 2020-10-24                       forecast  38115  217115   6000384
#> 80: France 2020-10-25                       forecast  39203  340534  10806840
#> 81: France 2020-10-26                       forecast  40449  554808  19510890
#> 82: France 2020-10-27                       forecast  41644  929687  35300460
#> 83: France 2020-10-28                       forecast  42767 1589422  63986695
#> 84: France 2020-10-29                       forecast  44090 2755852 116171403
#> 85: France 2020-10-30                       forecast  45403 4825877 211210304
#>     region       date                           type median    mean        sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     2330     2555     2114     2709     1852     3421
#>  2:     2398     2639     2204     2853     1939     3598
#>  3:     2532     2790     2349     3046     2054     3817
#>  4:     2624     2904     2458     3207     2158     4030
#>  5:     2826     3123     2557     3361     2266     4260
#>  6:     2960     3272     2643     3501     2378     4494
#>  7:     3208     3539     2837     3753     2484     4734
#>  8:     3305     3659     2975     3943     2567     4948
#>  9:     3429     3801     3176     4204     2728     5240
#> 10:     3592     3987     3386     4476     2857     5518
#> 11:     3775     4186     3572     4720     2993     5799
#> 12:     3991     4428     3758     4965     3005     5982
#> 13:     4301     4770     3934     5205     3233     6375
#> 14:     4240     4740     4202     5536     3479     6756
#> 15:     4754     5280     4318     5723     3546     6990
#> 16:     4990     5536     4409     5883     3780     7379
#> 17:     5033     5614     4712     6253     3910     7659
#> 18:     5348     5954     4947     6544     4116     8041
#> 19:     5716     6346     5104     6773     4258     8356
#> 20:     5779     6431     5268     7004     4441     8741
#> 21:     5861     6543     5366     7162     4616     9091
#> 22:     6305     7009     5639     7476     4830     9461
#> 23:     6384     7106     5894     7789     5141     9971
#> 24:     6363     7103     6146     8085     5268    10299
#> 25:     6807     7559     6360     8347     5264    10492
#> 26:     7106     7874     6460     8526     5734    11134
#> 27:     7420     8187     7007     9125     5896    11447
#> 28:     7687     8475     7224     9402     6145    11852
#> 29:     8011     8812     7519     9765     6249    12139
#> 30:     8081     8926     7529     9855     6653    12665
#> 31:     8589     9460     8131    10503     6780    12989
#> 32:     8491     9392     8054    10483     7074    13412
#> 33:     9334    10255     8377    10869     7138    13626
#> 34:     9377    10317     8654    11227     7369    14017
#> 35:     9684    10642     8970    11637     7644    14425
#> 36:    10192    11184     9064    11812     7730    14675
#> 37:    10634    11668     9381    12187     8074    15273
#> 38:    10844    11890     9726    12593     8271    15669
#> 39:    11055    12118     9910    12856     8474    16190
#> 40:    11311    12387    10112    13182     8731    16728
#> 41:    11622    12724    10607    13812     9011    17280
#> 42:    11851    12992    11107    14381     9356    17882
#> 43:    12017    13200    11448    14817     9639    18413
#> 44:    12423    13663    11573    15035     9743    18868
#> 45:    12506    13823    11751    15348     9830    19266
#> 46:    13025    14431    12166    15913    10245    20084
#> 47:    13593    15038    12271    16172    10508    20748
#> 48:    13702    15188    12437    16528    10174    20872
#> 49:    13604    15149    12895    17172    10601    21844
#> 50:    14354    15962    13235    17739    11065    22804
#> 51:    14615    16298    14017    18778    11071    23373
#> 52:    15867    17636    14162    19194    11572    24550
#> 53:    15887    17771    14461    19760    10830    24518
#> 54:    15974    17986    14299    19958    11632    26025
#> 55:    16336    18489    14774    20731    11447    26541
#> 56:    16522    18828    14135    20411    11515    27521
#> 57:    17039    19521    14827    21463    11370    28330
#> 58:    17143    19832    15518    22586    11467    29598
#> 59:    17527    20401    14963    22544    11417    30772
#> 60:    17666    20766    15476    23624    11348    32176
#> 61:    15807    19168    15604    24454    11329    33756
#> 62:    16766    20412    16021    25621    11345    35651
#> 63:    16231    20189    16282    26718    11250    37776
#> 64:    16707    21018    15963    27371    10836    39819
#> 65:    16826    21554    15974    28376    10598    42327
#> 66:    17637    22686    16526    29968     9073    44376
#> 67:    18144    23600    16382    31013     8996    47557
#> 68:    17880    23738    15227    31075     8116    50386
#> 69:    16465    22790    15140    32234     7770    53457
#> 70:    16143    22927    14853    33281     6904    56905
#> 71:    15803    23056    14713    34337     6138    60873
#> 72:    15752    23381    14895    35929     6343    66007
#> 73:    15868    24108    14695    37243     4353    69188
#> 74:    15706    24478    13990    38078     2992    72575
#> 75:    15506    24735    13347    39059     3175    77597
#> 76:    15493    25239    13671    41054     3447    83489
#> 77:    17157    27487    11856    40959     2114    88089
#> 78:    17603    28496    11459    42193     1153    94002
#> 79:    15177    26669    11090    43496     2084   102203
#> 80:    15106    27048    11019    45250     1492   108641
#> 81:    15190    27716    10879    46902      759   115923
#> 82:    14597    27693    10067    47966      399   124416
#> 83:    14343    28073     8676    48360      343   133611
#> 84:    14224    28564     6064    47594      109   143068
#> 85:    12211    27091     5738    49108       90   153157
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesReportData
#>     region       date                           type median  mean     sd
#>  1: France 2020-08-07                       estimate   1311  2010   2322
#>  2: France 2020-08-08                       estimate   1403  2120   2280
#>  3: France 2020-08-09                       estimate   1078  1698   1914
#>  4: France 2020-08-10                       estimate    954  1514   1741
#>  5: France 2020-08-11                       estimate   1350  2064   2241
#>  6: France 2020-08-12                       estimate   1051  1662   1916
#>  7: France 2020-08-13                       estimate   1666  2588   2904
#>  8: France 2020-08-14                       estimate   2018  3065   3415
#>  9: France 2020-08-15                       estimate   2030  3189   3751
#> 10: France 2020-08-16                       estimate   1618  2509   2862
#> 11: France 2020-08-17                       estimate   1427  2150   2330
#> 12: France 2020-08-18                       estimate   1950  3026   3353
#> 13: France 2020-08-19                       estimate   1572  2441   2733
#> 14: France 2020-08-20                       estimate   2493  3883   4528
#> 15: France 2020-08-21                       estimate   2876  4539   5136
#> 16: France 2020-08-22                       estimate   3036  4731   5561
#> 17: France 2020-08-23                       estimate   2306  3611   4153
#> 18: France 2020-08-24                       estimate   1994  3101   3423
#> 19: France 2020-08-25                       estimate   2788  4523   5159
#> 20: France 2020-08-26                       estimate   2309  3537   3847
#> 21: France 2020-08-27                       estimate   3452  5326   5928
#> 22: France 2020-08-28                       estimate   4035  6147   6768
#> 23: France 2020-08-29                       estimate   4198  6475   7259
#> 24: France 2020-08-30                       estimate   3244  4892   5247
#> 25: France 2020-08-31                       estimate   2817  4365   4921
#> 26: France 2020-09-01                       estimate   3948  6193   7091
#> 27: France 2020-09-02                       estimate   3145  4842   5284
#> 28: France 2020-09-03                       estimate   4665  7313   8369
#> 29: France 2020-09-04                       estimate   5329  8397   9648
#> 30: France 2020-09-05                       estimate   5588  8549   9498
#> 31: France 2020-09-06                       estimate   4266  6586   7405
#> 32: France 2020-09-07                       estimate   3576  5630   6234
#> 33: France 2020-09-08                       estimate   4874  7839   8819
#> 34: France 2020-09-09                       estimate   4166  6350   7190
#> 35: France 2020-09-10                       estimate   6295  9724  11132
#> 36: France 2020-09-11                       estimate   6760 10741  12009
#> 37: France 2020-09-12                       estimate   7214 11316  12681
#> 38: France 2020-09-13                       estimate   5408  8387   9320
#> 39: France 2020-09-14                       estimate   4620  7002   7704
#> 40: France 2020-09-15                       estimate   6468 10284  11781
#> 41: France 2020-09-16                       estimate   5058  7827   9131
#> 42: France 2020-09-17                       estimate   7710 11859  13551
#> 43: France 2020-09-18                       estimate   8777 13814  15497
#> 44: France 2020-09-19                       estimate   9276 14304  16334
#> 45: France 2020-09-20                       estimate   6616 10164  11307
#> 46: France 2020-09-21                       estimate   5788  8895   9564
#> 47: France 2020-09-22                       estimate   8240 12432  13846
#> 48: France 2020-09-23                       estimate   6090  9627  10766
#> 49: France 2020-09-24                       estimate   8984 14187  17222
#> 50: France 2020-09-25                       estimate  10088 16204  18078
#> 51: France 2020-09-26                       estimate  10928 16920  18557
#> 52: France 2020-09-27                       estimate   7864 12438  13818
#> 53: France 2020-09-28                       estimate   6919 10717  12098
#> 54: France 2020-09-29                       estimate   9616 15066  17766
#> 55: France 2020-09-30                       estimate   7374 11502  12816
#> 56: France 2020-10-01                       estimate  11106 17258  19114
#> 57: France 2020-10-02                       estimate  12262 18960  21157
#> 58: France 2020-10-03                       estimate  12538 19848  22729
#> 59: France 2020-10-04                       estimate   9557 15041  17197
#> 60: France 2020-10-05                       estimate   8024 12862  14496
#> 61: France 2020-10-06 estimate based on partial data  11258 18446  21627
#> 62: France 2020-10-07 estimate based on partial data   8538 13698  15806
#> 63: France 2020-10-08 estimate based on partial data  13246 21238  24952
#> 64: France 2020-10-09 estimate based on partial data  15631 24191  27636
#> 65: France 2020-10-10 estimate based on partial data  15886 25533  30226
#> 66: France 2020-10-11 estimate based on partial data  11631 18702  23611
#> 67: France 2020-10-12 estimate based on partial data   9724 15916  19168
#> 68: France 2020-10-13 estimate based on partial data  13866 22727  28095
#> 69: France 2020-10-14 estimate based on partial data  11062 17804  21898
#> 70: France 2020-10-15 estimate based on partial data  15609 25739  32252
#> 71: France 2020-10-16 estimate based on partial data  19234 31045  37795
#> 72: France 2020-10-17                       forecast  19050 31613  38628
#> 73: France 2020-10-18                       forecast  14472 24426  33161
#> 74: France 2020-10-19                       forecast  11814 21558  34137
#> 75: France 2020-10-20                       forecast  17116 30989  54468
#> 76: France 2020-10-21                       forecast  13611 24989  62584
#> 77: France 2020-10-22                       forecast  20006 39587 114602
#> 78: France 2020-10-23                       forecast  22086 43320  97484
#> 79: France 2020-10-24                       forecast  22506 51942 273033
#> 80: France 2020-10-25                       forecast  17645 36928 143189
#> 81: France 2020-10-26                       forecast  14823 39567 328720
#> 82: France 2020-10-27                       forecast  19982 62122 666571
#> 83: France 2020-10-28                       forecast  15662 48760 476972
#> 84: France 2020-10-29                       forecast  23217 68309 596793
#> 85: France 2020-10-30                       forecast  27909 92248 883583
#>     region       date                           type median  mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:        0      421        0     1311        0     4623
#>  2:        1      437        1     1404        1     4875
#>  3:        0      345        0     1079        0     3929
#>  4:        7      325        0      954        0     3491
#>  5:        1      399        0     1350        0     4935
#>  6:        0      324        0     1052        0     3852
#>  7:        5      497        1     1666        0     6052
#>  8:        0      618        0     2018        0     7114
#>  9:        0      600        0     2031        0     7331
#> 10:        0      488        0     1618        0     5851
#> 11:        0      438        0     1427        0     4985
#> 12:        3      643        2     1951        0     7028
#> 13:        0      487        0     1572        0     5637
#> 14:        0      770        0     2493        0     9182
#> 15:        0      896        1     2876        0    10853
#> 16:        0      895        0     3036        0    11063
#> 17:        0      658        0     2306        0     8649
#> 18:        5      626        0     1994        0     7328
#> 19:        3      876        0     2788        0    10696
#> 20:        4      694        1     2309        0     8314
#> 21:        6     1054        0     3452        0    12575
#> 22:        1     1173        1     4036        1    14192
#> 23:        0     1314        3     4201        0    14950
#> 24:       48     1049        0     3245        0    11647
#> 25:        1      832        1     2818        1    10374
#> 26:        0     1187        0     3949        0    14426
#> 27:        0      927        0     3146        0    11537
#> 28:        1     1415        1     4667        1    17049
#> 29:        0     1598        0     5330        0    19684
#> 30:        2     1692        2     5589        2    19510
#> 31:        0     1310        0     4266        0    15061
#> 32:        1     1143        1     3581        1    13311
#> 33:       13     1587        0     4878        0    18691
#> 34:        0     1244        0     4168        0    14740
#> 35:       19     1946        4     6299        0    22593
#> 36:       24     2197        0     6763        0    25769
#> 37:        0     2214        0     7215        0    26048
#> 38:        2     1625        2     5410        4    19402
#> 39:        0     1396        0     4620        0    16166
#> 40:       22     2037        3     6471        3    24048
#> 41:        0     1540        0     5060        0    18146
#> 42:       30     2458        2     7712        2    27505
#> 43:        2     2789        2     8779        2    32278
#> 44:       58     2905        2     9279        2    32938
#> 45:       12     2073        5     6618        1    23453
#> 46:        0     1802        5     5793        0    20552
#> 47:        6     2387        6     8251        6    28673
#> 48:        2     1908        2     6095        2    22750
#> 49:        1     2729        1     8987        1    32918
#> 50:        7     3230        7    10090        4    38033
#> 51:        3     3202        2    10932        2    40336
#> 52:        4     2454        4     7872        4    29802
#> 53:        8     1971        1     6919        1    25485
#> 54:        5     2829        0     9619        0    35002
#> 55:        8     2419        0     7378        0    26892
#> 56:        8     3462        8    11116        8    40600
#> 57:        1     3711        1    12266        1    44275
#> 58:        2     3885        2    12541        2    44884
#> 59:       28     2828        0     9559        0    35626
#> 60:        1     2356        1     8025        1    30326
#> 61:       30     3512        2    11261        2    44310
#> 62:        2     2697        2     8540        0    32574
#> 63:        4     4285        4    13246        4    49247
#> 64:        0     4543        0    15635        0    56968
#> 65:        2     4627        2    15887        2    60404
#> 66:        0     3548        0    11633        0    42332
#> 67:       20     2971       16     9737        2    37576
#> 68:        0     4026        0    13870        0    53239
#> 69:        1     3255        6    11067        1    41226
#> 70:        0     4736        0    15612        0    58667
#> 71:       34     5875        0    19234        0    72318
#> 72:       60     6093        1    19056        1    73227
#> 73:        1     4378        1    14476        1    57142
#> 74:       17     3608        0    11815        0    49307
#> 75:       34     5025        5    17140        5    71300
#> 76:       44     3881        3    13615        0    56117
#> 77:       50     5422       12    20014       12    86893
#> 78:       18     6207       12    22091        7    97914
#> 79:        0     6713        0    22507        0   107536
#> 80:       18     5071        2    17651        2    76845
#> 81:       18     3999        0    14839        0    68593
#> 82:       21     5434        0    19985        0   100486
#> 83:       13     4299        1    15680        1    81007
#> 84:        7     6307        7    23221        7   118104
#> 85:       21     6930       19    27910       18   146356
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-07-24    1062
#>  2: France 2020-07-25    1130
#>  3: France 2020-07-26       0
#>  4: France 2020-07-27       0
#>  5: France 2020-07-28    2551
#>  6: France 2020-07-29     725
#>  7: France 2020-07-30    1392
#>  8: France 2020-07-31    1377
#>  9: France 2020-08-01    1346
#> 10: France 2020-08-02       0
#> 11: France 2020-08-03       0
#> 12: France 2020-08-04    3376
#> 13: France 2020-08-05    1039
#> 14: France 2020-08-06    1695
#> 15: France 2020-08-07    1604
#> 16: France 2020-08-08    2288
#> 17: France 2020-08-09       0
#> 18: France 2020-08-10       0
#> 19: France 2020-08-11    4854
#> 20: France 2020-08-12    1397
#> 21: France 2020-08-13    2524
#> 22: France 2020-08-14    2669
#> 23: France 2020-08-15    2846
#> 24: France 2020-08-16    3310
#> 25: France 2020-08-17    3015
#> 26: France 2020-08-18     493
#> 27: France 2020-08-19    2238
#> 28: France 2020-08-20    3776
#> 29: France 2020-08-21    4771
#> 30: France 2020-08-22    4586
#> 31: France 2020-08-23    3602
#> 32: France 2020-08-24    4897
#> 33: France 2020-08-25    1955
#> 34: France 2020-08-26    3304
#> 35: France 2020-08-27    5429
#> 36: France 2020-08-28    6111
#> 37: France 2020-08-29    7379
#> 38: France 2020-08-30    5453
#> 39: France 2020-08-31    5413
#> 40: France 2020-09-01    3082
#> 41: France 2020-09-02    4982
#> 42: France 2020-09-03    7017
#> 43: France 2020-09-04    7157
#> 44: France 2020-09-05    8975
#> 45: France 2020-09-06    8550
#> 46: France 2020-09-07    7071
#> 47: France 2020-09-08    4203
#> 48: France 2020-09-09    6544
#> 49: France 2020-09-10    8577
#> 50: France 2020-09-11    9843
#> 51: France 2020-09-12    9406
#> 52: France 2020-09-13   10561
#> 53: France 2020-09-14    7183
#> 54: France 2020-09-15    6158
#> 55: France 2020-09-16    7852
#> 56: France 2020-09-17    9784
#> 57: France 2020-09-18   10593
#> 58: France 2020-09-19   13215
#> 59: France 2020-09-20   13498
#> 60: France 2020-09-21   11569
#> 61: France 2020-09-22    4298
#> 62: France 2020-09-23   10008
#> 63: France 2020-09-24   13072
#> 64: France 2020-09-25   16096
#> 65: France 2020-09-26   15797
#> 66: France 2020-09-27   14412
#> 67: France 2020-09-28   11123
#> 68: France 2020-09-29    4070
#> 69: France 2020-09-30    8051
#> 70: France 2020-10-01   12845
#> 71: France 2020-10-02   13970
#> 72: France 2020-10-03   12148
#> 73: France 2020-10-04   16972
#> 74: France 2020-10-05   12545
#> 75: France 2020-10-06    5104
#> 76: France 2020-10-07   10489
#> 77: France 2020-10-08   18746
#> 78: France 2020-10-09   18129
#> 79: France 2020-10-10   20339
#> 80: France 2020-10-11   26896
#> 81: France 2020-10-12   16101
#> 82: France 2020-10-13    8505
#> 83: France 2020-10-14   12993
#> 84: France 2020-10-15   22591
#> 85: France 2020-10-16   30621
#>     region       date confirm
#>