Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

rtData

#> $summaryData
#>              region New confirmed cases by infection date
#>   1:    Afghanistan                        42 (10 -- 129)
#>   2:        Albania                      180 (117 -- 270)
#>   3:        Algeria                       101 (75 -- 136)
#>   4:        Andorra                         63 (3 -- 160)
#>   5:         Angola                       177 (48 -- 344)
#>  ---                                                     
#> 192:        Vietnam                           2 (0 -- 16)
#> 193: Western Sahara                          1 (0 -- 289)
#> 194:          Yemen                            1 (0 -- 2)
#> 195:         Zambia                        47 (11 -- 106)
#> 196:       Zimbabwe                          12 (1 -- 50)
#>      Expected change in daily cases Effective reproduction no.
#>   1:                         Unsure             1 (0.8 -- 1.4)
#>   2:                         Unsure             1 (0.9 -- 1.2)
#>   3:              Likely decreasing             0.9 (0.8 -- 1)
#>   4:              Likely increasing           1.1 (0.8 -- 1.4)
#>   5:              Likely increasing           1.1 (0.8 -- 1.3)
#>  ---                                                          
#> 192:                         Unsure             1 (0.2 -- 2.1)
#> 193:                         Unsure             0.6 (0 -- 1.4)
#> 194:              Likely decreasing           0.9 (0.7 -- 1.2)
#> 195:                         Unsure           0.9 (0.7 -- 1.2)
#> 196:                         Unsure           0.9 (0.5 -- 1.6)
#>             Rate of growth Doubling/halving time (days)
#>   1:      0 (-0.05 -- 0.1)        1066.8 (6.6 -- -13.3)
#>   2:  0.01 (-0.02 -- 0.05)           55.3 (12.8 -- -31)
#>   3:    -0.02 (-0.05 -- 0)       -30.2 (258.3 -- -14.6)
#>   4:  0.03 (-0.06 -- 0.11)            20 (6.4 -- -12.2)
#>   5:  0.02 (-0.04 -- 0.07)          31.6 (9.6 -- -17.2)
#>  ---                                                   
#> 192:     0 (-0.25 -- 0.29)         -143.9 (2.4 -- -2.7)
#> 193:  -0.1 (-0.36 -- 0.14)           -6.7 (4.8 -- -1.9)
#> 194:  -0.03 (-0.1 -- 0.04)         -20.2 (16.2 -- -6.9)
#> 195: -0.02 (-0.09 -- 0.06)           -36.2 (11.6 -- -8)
#> 196: -0.02 (-0.12 -- 0.17)          -46.2 (4.1 -- -5.9)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-03 estimate   0.93 0.93 0.05     0.93     0.95
#>     2: Afghanistan 2020-08-04 estimate   0.93 0.93 0.06     0.93     0.95
#>     3: Afghanistan 2020-08-05 estimate   0.93 0.93 0.06     0.93     0.95
#>     4: Afghanistan 2020-08-06 estimate   0.93 0.93 0.06     0.92     0.94
#>     5: Afghanistan 2020-08-07 estimate   0.93 0.93 0.06     0.92     0.94
#>    ---                                                                   
#> 16256:    Zimbabwe 2020-10-22 forecast   0.95 1.05 0.40     0.83     0.93
#> 16257:    Zimbabwe 2020-10-23 forecast   0.95 1.05 0.40     0.83     0.93
#> 16258:    Zimbabwe 2020-10-24 forecast   0.95 1.05 0.40     0.83     0.93
#> 16259:    Zimbabwe 2020-10-25 forecast   0.95 1.05 0.40     0.83     0.93
#> 16260:    Zimbabwe 2020-10-26 forecast   0.95 1.05 0.40     0.83     0.93
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.91     0.96     0.85     1.02
#>     2:     0.91     0.96     0.84     1.01
#>     3:     0.91     0.96     0.83     1.01
#>     4:     0.91     0.96     0.83     1.01
#>     5:     0.91     0.96     0.83     1.01
#>    ---                                    
#> 16256:     0.71     1.01     0.52     1.57
#> 16257:     0.71     1.01     0.52     1.57
#> 16258:     0.71     1.01     0.52     1.57
#> 16259:     0.71     1.01     0.52     1.57
#> 16260:     0.71     1.01     0.52     1.57
#> 
#> $casesInfectionData
#>             region       date     type median mean    sd lower_20 upper_20
#>     1: Afghanistan 2020-08-03 estimate     69   70    16       68       74
#>     2: Afghanistan 2020-08-04 estimate     67   68    16       66       72
#>     3: Afghanistan 2020-08-05 estimate     66   67    15       65       71
#>     4: Afghanistan 2020-08-06 estimate     65   66    15       64       70
#>     5: Afghanistan 2020-08-07 estimate     64   65    15       63       69
#>    ---                                                                    
#> 16256:    Zimbabwe 2020-10-22 forecast     10  533  7072        2        5
#> 16257:    Zimbabwe 2020-10-23 forecast     10  832 12176        1        4
#> 16258:    Zimbabwe 2020-10-24 forecast     10 1320 21060        1        4
#> 16259:    Zimbabwe 2020-10-25 forecast     10 2120 36563        1        4
#> 16260:    Zimbabwe 2020-10-26 forecast     10 3446 63676        1        4
#>        lower_50 upper_50 lower_90 upper_90
#>     1:       59       76       43       91
#>     2:       58       75       42       89
#>     3:       56       73       43       89
#>     4:       55       72       41       87
#>     5:       55       72       40       86
#>    ---                                    
#> 16256:        0       10        0      161
#> 16257:        0       10        0      179
#> 16258:        0       10        0      203
#> 16259:        0       10        0      231
#> 16260:        0       10        0      263
#> 
#> $casesReportData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-03 estimate     38   73  102        0        7
#>     2: Afghanistan 2020-08-04 estimate     43   93  132        0        8
#>     3: Afghanistan 2020-08-05 estimate     42   84  115        0        8
#>     4: Afghanistan 2020-08-06 estimate     44   83  112        0        8
#>     5: Afghanistan 2020-08-07 estimate     30   61   86        0        5
#>    ---                                                                   
#> 16256:    Zimbabwe 2020-10-22 forecast      9   54  439        0        3
#> 16257:    Zimbabwe 2020-10-23 forecast      9   81  932        0        3
#> 16258:    Zimbabwe 2020-10-24 forecast     11  113  991        0        3
#> 16259:    Zimbabwe 2020-10-25 forecast     10  178 2257        0        3
#> 16260:    Zimbabwe 2020-10-26 forecast      7  189 2356        0        2
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0       39        0      183
#>     2:        0       43        0      234
#>     3:        0       42        0      218
#>     4:        0       44        0      216
#>     5:        0       30        0      160
#>    ---                                    
#> 16256:        0        9        0       56
#> 16257:        0        9        0       72
#> 16258:        0       11        0       89
#> 16259:        0       10        0      101
#> 16260:        0        7        0       81
#> 
#> $obsCasesData
#>                region       date confirm
#>     1:    Afghanistan 2020-07-20     174
#>     2:        Albania 2020-07-20      82
#>     3:        Algeria 2020-07-20     535
#>     4:        Andorra 2020-07-20       0
#>     5:         Angola 2020-07-20      18
#>    ---                                  
#> 17761:        Vietnam 2020-10-12       2
#> 17762: Western Sahara 2020-10-12       0
#> 17763:          Yemen 2020-10-12       0
#> 17764:         Zambia 2020-10-12     119
#> 17765:       Zimbabwe 2020-10-12       1
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France                 30317 (7729 -- 57230)              Likely increasing
#>    Effective reproduction no.       Rate of growth Doubling/halving time (days)
#> 1:           1.1 (0.9 -- 1.3) 0.04 (-0.03 -- 0.09)          19.3 (7.9 -- -21.1)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-08-04                       estimate   1.20 1.21 0.05     1.19
#>  2: France 2020-08-05                       estimate   1.20 1.21 0.05     1.18
#>  3: France 2020-08-06                       estimate   1.20 1.21 0.05     1.18
#>  4: France 2020-08-07                       estimate   1.20 1.21 0.05     1.18
#>  5: France 2020-08-08                       estimate   1.20 1.21 0.06     1.18
#>  6: France 2020-08-09                       estimate   1.20 1.21 0.06     1.18
#>  7: France 2020-08-10                       estimate   1.20 1.21 0.06     1.17
#>  8: France 2020-08-11                       estimate   1.20 1.21 0.06     1.18
#>  9: France 2020-08-12                       estimate   1.20 1.20 0.06     1.17
#> 10: France 2020-08-13                       estimate   1.19 1.20 0.06     1.17
#> 11: France 2020-08-14                       estimate   1.19 1.20 0.06     1.17
#> 12: France 2020-08-15                       estimate   1.19 1.20 0.06     1.17
#> 13: France 2020-08-16                       estimate   1.19 1.20 0.06     1.16
#> 14: France 2020-08-17                       estimate   1.19 1.19 0.06     1.17
#> 15: France 2020-08-18                       estimate   1.18 1.19 0.06     1.17
#> 16: France 2020-08-19                       estimate   1.18 1.19 0.05     1.17
#> 17: France 2020-08-20                       estimate   1.18 1.19 0.05     1.17
#> 18: France 2020-08-21                       estimate   1.18 1.18 0.05     1.17
#> 19: France 2020-08-22                       estimate   1.18 1.18 0.05     1.16
#> 20: France 2020-08-23                       estimate   1.18 1.18 0.05     1.17
#> 21: France 2020-08-24                       estimate   1.17 1.17 0.05     1.17
#> 22: France 2020-08-25                       estimate   1.17 1.17 0.05     1.16
#> 23: France 2020-08-26                       estimate   1.17 1.17 0.05     1.16
#> 24: France 2020-08-27                       estimate   1.17 1.17 0.05     1.16
#> 25: France 2020-08-28                       estimate   1.16 1.16 0.05     1.16
#> 26: France 2020-08-29                       estimate   1.16 1.16 0.05     1.16
#> 27: France 2020-08-30                       estimate   1.16 1.16 0.05     1.15
#> 28: France 2020-08-31                       estimate   1.16 1.15 0.05     1.16
#> 29: France 2020-09-01                       estimate   1.16 1.15 0.05     1.16
#> 30: France 2020-09-02                       estimate   1.15 1.15 0.05     1.16
#> 31: France 2020-09-03                       estimate   1.15 1.15 0.05     1.16
#> 32: France 2020-09-04                       estimate   1.15 1.14 0.05     1.16
#> 33: France 2020-09-05                       estimate   1.15 1.14 0.05     1.15
#> 34: France 2020-09-06                       estimate   1.14 1.14 0.05     1.15
#> 35: France 2020-09-07                       estimate   1.14 1.13 0.06     1.15
#> 36: France 2020-09-08                       estimate   1.14 1.13 0.06     1.15
#> 37: France 2020-09-09                       estimate   1.14 1.13 0.06     1.15
#> 38: France 2020-09-10                       estimate   1.14 1.13 0.06     1.16
#> 39: France 2020-09-11                       estimate   1.13 1.13 0.06     1.15
#> 40: France 2020-09-12                       estimate   1.13 1.12 0.06     1.12
#> 41: France 2020-09-13                       estimate   1.13 1.12 0.06     1.15
#> 42: France 2020-09-14                       estimate   1.13 1.12 0.06     1.15
#> 43: France 2020-09-15                       estimate   1.13 1.12 0.06     1.15
#> 44: France 2020-09-16                       estimate   1.12 1.12 0.06     1.15
#> 45: France 2020-09-17                       estimate   1.12 1.12 0.06     1.15
#> 46: France 2020-09-18                       estimate   1.12 1.12 0.06     1.14
#> 47: France 2020-09-19                       estimate   1.12 1.12 0.06     1.15
#> 48: France 2020-09-20                       estimate   1.12 1.11 0.06     1.15
#> 49: France 2020-09-21                       estimate   1.12 1.11 0.07     1.15
#> 50: France 2020-09-22                       estimate   1.12 1.11 0.07     1.15
#> 51: France 2020-09-23                       estimate   1.12 1.11 0.07     1.14
#> 52: France 2020-09-24                       estimate   1.12 1.11 0.07     1.14
#> 53: France 2020-09-25                       estimate   1.12 1.12 0.07     1.15
#> 54: France 2020-09-26                       estimate   1.12 1.12 0.07     1.15
#> 55: France 2020-09-27                       estimate   1.12 1.12 0.08     1.14
#> 56: France 2020-09-28                       estimate   1.12 1.12 0.08     1.15
#> 57: France 2020-09-29                       estimate   1.12 1.12 0.08     1.14
#> 58: France 2020-09-30                       estimate   1.12 1.12 0.09     1.15
#> 59: France 2020-10-01                       estimate   1.13 1.12 0.09     1.15
#> 60: France 2020-10-02 estimate based on partial data   1.13 1.12 0.09     1.15
#> 61: France 2020-10-03 estimate based on partial data   1.13 1.12 0.10     1.15
#> 62: France 2020-10-04 estimate based on partial data   1.13 1.12 0.10     1.14
#> 63: France 2020-10-05 estimate based on partial data   1.13 1.13 0.11     1.14
#> 64: France 2020-10-06 estimate based on partial data   1.13 1.13 0.12     1.13
#> 65: France 2020-10-07 estimate based on partial data   1.13 1.13 0.12     1.14
#> 66: France 2020-10-08 estimate based on partial data   1.13 1.13 0.13     1.14
#> 67: France 2020-10-09 estimate based on partial data   1.13 1.13 0.13     1.15
#> 68: France 2020-10-10 estimate based on partial data   1.13 1.13 0.14     1.14
#> 69: France 2020-10-11 estimate based on partial data   1.13 1.13 0.15     1.14
#> 70: France 2020-10-12 estimate based on partial data   1.13 1.13 0.16     1.14
#> 71: France 2020-10-13                       forecast   1.13 1.13 0.16     1.14
#> 72: France 2020-10-14                       forecast   1.13 1.13 0.16     1.14
#> 73: France 2020-10-15                       forecast   1.13 1.13 0.16     1.14
#> 74: France 2020-10-16                       forecast   1.13 1.13 0.16     1.14
#> 75: France 2020-10-17                       forecast   1.13 1.13 0.16     1.14
#> 76: France 2020-10-18                       forecast   1.13 1.13 0.16     1.14
#> 77: France 2020-10-19                       forecast   1.13 1.13 0.16     1.14
#> 78: France 2020-10-20                       forecast   1.13 1.13 0.16     1.14
#> 79: France 2020-10-21                       forecast   1.13 1.13 0.16     1.14
#> 80: France 2020-10-22                       forecast   1.13 1.13 0.16     1.14
#> 81: France 2020-10-23                       forecast   1.13 1.13 0.16     1.14
#> 82: France 2020-10-24                       forecast   1.13 1.13 0.16     1.14
#> 83: France 2020-10-25                       forecast   1.13 1.13 0.16     1.14
#> 84: France 2020-10-26                       forecast   1.13 1.13 0.16     1.14
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.21     1.16     1.22     1.13     1.28
#>  2:     1.20     1.17     1.22     1.13     1.28
#>  3:     1.20     1.17     1.22     1.14     1.29
#>  4:     1.19     1.17     1.22     1.13     1.29
#>  5:     1.19     1.16     1.22     1.13     1.29
#>  6:     1.20     1.16     1.22     1.12     1.29
#>  7:     1.19     1.16     1.22     1.13     1.30
#>  8:     1.20     1.17     1.22     1.12     1.29
#>  9:     1.19     1.16     1.22     1.12     1.29
#> 10:     1.19     1.16     1.21     1.11     1.29
#> 11:     1.18     1.16     1.21     1.12     1.29
#> 12:     1.18     1.15     1.21     1.11     1.29
#> 13:     1.18     1.16     1.21     1.12     1.29
#> 14:     1.18     1.15     1.21     1.12     1.29
#> 15:     1.18     1.15     1.21     1.11     1.28
#> 16:     1.18     1.15     1.20     1.11     1.28
#> 17:     1.19     1.15     1.20     1.11     1.28
#> 18:     1.18     1.15     1.20     1.10     1.28
#> 19:     1.18     1.15     1.20     1.10     1.28
#> 20:     1.18     1.15     1.20     1.08     1.26
#> 21:     1.18     1.15     1.20     1.08     1.25
#> 22:     1.18     1.14     1.20     1.08     1.26
#> 23:     1.18     1.14     1.20     1.08     1.25
#> 24:     1.18     1.14     1.20     1.08     1.25
#> 25:     1.18     1.14     1.19     1.07     1.25
#> 26:     1.18     1.14     1.19     1.07     1.24
#> 27:     1.17     1.14     1.19     1.07     1.24
#> 28:     1.17     1.13     1.19     1.07     1.24
#> 29:     1.18     1.13     1.19     1.06     1.23
#> 30:     1.18     1.13     1.18     1.06     1.22
#> 31:     1.18     1.13     1.19     1.05     1.22
#> 32:     1.18     1.13     1.18     1.05     1.22
#> 33:     1.17     1.13     1.18     1.04     1.21
#> 34:     1.17     1.12     1.18     1.04     1.22
#> 35:     1.17     1.12     1.18     1.04     1.21
#> 36:     1.18     1.12     1.18     1.03     1.21
#> 37:     1.17     1.12     1.18     1.03     1.21
#> 38:     1.18     1.12     1.18     1.04     1.22
#> 39:     1.17     1.11     1.18     1.03     1.21
#> 40:     1.15     1.11     1.18     1.02     1.21
#> 41:     1.18     1.11     1.18     1.02     1.21
#> 42:     1.18     1.11     1.18     1.02     1.21
#> 43:     1.18     1.11     1.18     1.01     1.20
#> 44:     1.18     1.10     1.18     1.01     1.20
#> 45:     1.18     1.10     1.18     1.02     1.21
#> 46:     1.17     1.10     1.18     1.01     1.20
#> 47:     1.18     1.10     1.18     1.01     1.21
#> 48:     1.18     1.10     1.18     1.01     1.21
#> 49:     1.18     1.10     1.18     1.01     1.21
#> 50:     1.18     1.10     1.18     1.00     1.21
#> 51:     1.17     1.10     1.18     1.00     1.21
#> 52:     1.18     1.10     1.18     1.00     1.22
#> 53:     1.18     1.10     1.18     1.00     1.22
#> 54:     1.18     1.10     1.19     1.00     1.23
#> 55:     1.17     1.10     1.19     0.99     1.23
#> 56:     1.18     1.10     1.19     0.99     1.24
#> 57:     1.18     1.10     1.19     0.98     1.23
#> 58:     1.18     1.09     1.19     0.98     1.25
#> 59:     1.19     1.09     1.19     0.97     1.25
#> 60:     1.18     1.09     1.19     0.96     1.25
#> 61:     1.19     1.09     1.19     0.95     1.26
#> 62:     1.18     1.09     1.20     0.94     1.26
#> 63:     1.18     1.08     1.19     0.94     1.27
#> 64:     1.17     1.08     1.20     0.93     1.28
#> 65:     1.18     1.08     1.19     0.94     1.30
#> 66:     1.18     1.08     1.20     0.92     1.30
#> 67:     1.19     1.07     1.20     0.91     1.30
#> 68:     1.18     1.07     1.20     0.90     1.31
#> 69:     1.18     1.07     1.20     0.89     1.32
#> 70:     1.18     1.07     1.20     0.88     1.32
#> 71:     1.18     1.07     1.20     0.88     1.32
#> 72:     1.18     1.07     1.20     0.88     1.32
#> 73:     1.18     1.07     1.20     0.88     1.32
#> 74:     1.18     1.07     1.20     0.88     1.32
#> 75:     1.18     1.07     1.20     0.88     1.32
#> 76:     1.18     1.07     1.20     0.88     1.32
#> 77:     1.18     1.07     1.20     0.88     1.32
#> 78:     1.18     1.07     1.20     0.88     1.32
#> 79:     1.18     1.07     1.20     0.88     1.32
#> 80:     1.18     1.07     1.20     0.88     1.32
#> 81:     1.18     1.07     1.20     0.88     1.32
#> 82:     1.18     1.07     1.20     0.88     1.32
#> 83:     1.18     1.07     1.20     0.88     1.32
#> 84:     1.18     1.07     1.20     0.88     1.32
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData
#>     region       date                           type median   mean      sd
#>  1: France 2020-08-04                       estimate   2179   2244     408
#>  2: France 2020-08-05                       estimate   2294   2367     436
#>  3: France 2020-08-06                       estimate   2420   2497     466
#>  4: France 2020-08-07                       estimate   2548   2633     498
#>  5: France 2020-08-08                       estimate   2684   2776     531
#>  6: France 2020-08-09                       estimate   2830   2927     566
#>  7: France 2020-08-10                       estimate   2979   3084     603
#>  8: France 2020-08-11                       estimate   3133   3249     641
#>  9: France 2020-08-12                       estimate   3300   3421     682
#> 10: France 2020-08-13                       estimate   3475   3600     724
#> 11: France 2020-08-14                       estimate   3662   3787     769
#> 12: France 2020-08-15                       estimate   3853   3981     816
#> 13: France 2020-08-16                       estimate   4054   4182     865
#> 14: France 2020-08-17                       estimate   4255   4390     915
#> 15: France 2020-08-18                       estimate   4463   4605     967
#> 16: France 2020-08-19                       estimate   4673   4827    1021
#> 17: France 2020-08-20                       estimate   4893   5056    1076
#> 18: France 2020-08-21                       estimate   5125   5291    1132
#> 19: France 2020-08-22                       estimate   5355   5532    1189
#> 20: France 2020-08-23                       estimate   5594   5780    1247
#> 21: France 2020-08-24                       estimate   5845   6034    1307
#> 22: France 2020-08-25                       estimate   6097   6294    1367
#> 23: France 2020-08-26                       estimate   6339   6560    1429
#> 24: France 2020-08-27                       estimate   6597   6832    1491
#> 25: France 2020-08-28                       estimate   6860   7109    1554
#> 26: France 2020-08-29                       estimate   7147   7393    1616
#> 27: France 2020-08-30                       estimate   7426   7681    1677
#> 28: France 2020-08-31                       estimate   7710   7975    1736
#> 29: France 2020-09-01                       estimate   8009   8274    1792
#> 30: France 2020-09-02                       estimate   8308   8577    1845
#> 31: France 2020-09-03                       estimate   8630   8886    1896
#> 32: France 2020-09-04                       estimate   8925   9199    1944
#> 33: France 2020-09-05                       estimate   9231   9517    1991
#> 34: France 2020-09-06                       estimate   9575   9840    2037
#> 35: France 2020-09-07                       estimate   9894  10168    2084
#> 36: France 2020-09-08                       estimate  10217  10502    2133
#> 37: France 2020-09-09                       estimate  10534  10842    2185
#> 38: France 2020-09-10                       estimate  10884  11188    2242
#> 39: France 2020-09-11                       estimate  11250  11541    2305
#> 40: France 2020-09-12                       estimate  11606  11900    2374
#> 41: France 2020-09-13                       estimate  11989  12268    2450
#> 42: France 2020-09-14                       estimate  12349  12644    2534
#> 43: France 2020-09-15                       estimate  12686  13028    2626
#> 44: France 2020-09-16                       estimate  13080  13423    2727
#> 45: France 2020-09-17                       estimate  13461  13829    2838
#> 46: France 2020-09-18                       estimate  13854  14246    2961
#> 47: France 2020-09-19                       estimate  14282  14677    3097
#> 48: France 2020-09-20                       estimate  14709  15123    3247
#> 49: France 2020-09-21                       estimate  15135  15584    3414
#> 50: France 2020-09-22                       estimate  15568  16064    3600
#> 51: France 2020-09-23                       estimate  16050  16564    3808
#> 52: France 2020-09-24                       estimate  16539  17086    4041
#> 53: France 2020-09-25                       estimate  17075  17632    4302
#> 54: France 2020-09-26                       estimate  17614  18204    4596
#> 55: France 2020-09-27                       estimate  18148  18806    4927
#> 56: France 2020-09-28                       estimate  18742  19439    5299
#> 57: France 2020-09-29                       estimate  19347  20107    5720
#> 58: France 2020-09-30                       estimate  19965  20814    6194
#> 59: France 2020-10-01                       estimate  20610  21562    6729
#> 60: France 2020-10-02 estimate based on partial data  21345  22357    7336
#> 61: France 2020-10-03 estimate based on partial data  22062  23202    8024
#> 62: France 2020-10-04 estimate based on partial data  22819  24103    8810
#> 63: France 2020-10-05 estimate based on partial data  23613  25068    9711
#> 64: France 2020-10-06 estimate based on partial data  24478  26102   10755
#> 65: France 2020-10-07 estimate based on partial data  25336  27217   11978
#> 66: France 2020-10-08 estimate based on partial data  26317  28424   13434
#> 67: France 2020-10-09 estimate based on partial data  27192  29738   15205
#> 68: France 2020-10-10 estimate based on partial data  28188  31180   17424
#> 69: France 2020-10-11 estimate based on partial data  29239  32776   20309
#> 70: France 2020-10-12 estimate based on partial data  30317  34565   24240
#> 71: France 2020-10-13                       forecast  31396  36525   29292
#> 72: France 2020-10-14                       forecast  32537  38723   36154
#> 73: France 2020-10-15                       forecast  33737  41210   45603
#> 74: France 2020-10-16                       forecast  34915  44051   58764
#> 75: France 2020-10-17                       forecast  36147  47336   77250
#> 76: France 2020-10-18                       forecast  37415  51183  103381
#> 77: France 2020-10-19                       forecast  38637  55751  140495
#> 78: France 2020-10-20                       forecast  40085  61257  193418
#> 79: France 2020-10-21                       forecast  41448  67998  269134
#> 80: France 2020-10-22                       forecast  43067  76381  377778
#> 81: France 2020-10-23                       forecast  44468  86971  534084
#> 82: France 2020-10-24                       forecast  46070 100553  759493
#> 83: France 2020-10-25                       forecast  47563 118226 1085249
#> 84: France 2020-10-26                       forecast  49269 141531 1556917
#>     region       date                           type median   mean      sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1976     2131     1890     2329     1590     2863
#>  2:     2061     2230     1968     2445     1731     3080
#>  3:     2216     2399     2090     2603     1798     3223
#>  4:     2302     2501     2165     2714     1908     3406
#>  5:     2449     2663     2279     2870     1991     3577
#>  6:     2545     2775     2351     2989     2081     3760
#>  7:     2701     2947     2435     3116     2219     4003
#>  8:     2886     3149     2566     3296     2294     4191
#>  9:     3044     3327     2679     3462     2404     4421
#> 10:     3133     3434     2832     3668     2483     4644
#> 11:     3189     3512     3048     3942     2604     4912
#> 12:     3402     3743     3179     4128     2735     5182
#> 13:     3553     3907     3292     4305     2898     5486
#> 14:     3708     4084     3482     4561     3074     5818
#> 15:     3905     4313     3621     4765     3105     6003
#> 16:     4019     4454     3700     4922     3370     6426
#> 17:     4291     4766     3915     5200     3511     6729
#> 18:     4410     4910     4093     5451     3640     7008
#> 19:     4581     5111     4373     5786     3812     7320
#> 20:     4701     5252     4667     6128     3964     7642
#> 21:     4923     5501     4896     6409     4132     7979
#> 22:     5168     5764     5100     6679     4283     8315
#> 23:     5396     6015     5186     6824     4406     8606
#> 24:     5575     6218     5536     7237     4579     8955
#> 25:     6341     7009     5651     7417     4916     9452
#> 26:     6553     7242     5936     7766     5023     9711
#> 27:     6877     7590     6114     8012     5269    10135
#> 28:     7306     8046     6477     8433     5478    10538
#> 29:     7695     8471     6685     8720     5575    10830
#> 30:     7845     8642     6973     9078     5789    11204
#> 31:     7598     8406     7245     9410     6005    11585
#> 32:     7854     8702     7518     9757     6211    11981
#> 33:     9041     9920     7722    10034     6585    12534
#> 34:     8804     9700     8109    10473     6836    13001
#> 35:     9353    10261     8476    10916     6822    13179
#> 36:     9611    10540     8665    11162     7267    13826
#> 37:     9576    10527     8711    11287     7192    13967
#> 38:    10062    11040     9090    11756     7493    14495
#> 39:    10052    11066     9387    12151     8028    15253
#> 40:    10316    11367     9730    12562     7890    15350
#> 41:    10635    11725     9994    12931     8427    16146
#> 42:    11107    12237    10198    13259     8453    16434
#> 43:    11478    12616    10730    13901     8609    16856
#> 44:    11769    12943    11248    14558     9058    17594
#> 45:    12252    13475    11584    15023     9321    18210
#> 46:    12477    13745    11507    15122     9732    18980
#> 47:    12941    14289    11920    15719     9712    19427
#> 48:    12967    14362    12172    16178    10007    20196
#> 49:    13289    14753    12381    16573    10151    20896
#> 50:    13864    15393    12882    17250    10395    21623
#> 51:    13950    15599    12948    17610    10528    22254
#> 52:    14731    16498    12792    17749    10946    23377
#> 53:    15759    17623    13282    18541    11030    24314
#> 54:    15996    17971    13464    19125    11113    25242
#> 55:    16275    18409    13963    20049    11397    26541
#> 56:    16750    19033    14235    20728    10938    27220
#> 57:    16999    19520    14641    21620    11048    28553
#> 58:    17330    20062    14660    22193    11324    30110
#> 59:    18038    20997    14482    22647    11083    31346
#> 60:    19233    22377    14522    23318    10802    32600
#> 61:    19124    22480    14562    24103    10509    34091
#> 62:    19806    23459    16157    26452    10641    36113
#> 63:    20766    24785    16289    27540    10503    38150
#> 64:    21434    25825    16298    28534    11012    41079
#> 65:    21471    26316    16718    29938     9762    42305
#> 66:    22335    27683    16543    30828     9711    45024
#> 67:    22165    28012    16806    32257     9267    48082
#> 68:    17600    23947    17593    34235     8072    50361
#> 69:    17654    24452    17712    35689     7850    53522
#> 70:    17642    25013    16539    35960     7729    57230
#> 71:    18344    26254    17004    37828     7462    60856
#> 72:    20292    28755    16968    39285     7517    65704
#> 73:    19197    28262    17089    41026     7082    69951
#> 74:    18635    28193    15743    41284     6306    73760
#> 75:    18731    28771    16153    43329     5721    78765
#> 76:    18666    29250    15901    44890     5691    84453
#> 77:    18529    29678    15572    46519     2742    87316
#> 78:    18228    30013    15766    48670     2505    93470
#> 79:    19359    31802    16848    51831     2288   100323
#> 80:    17776    30889    12390    49333     1335   105797
#> 81:    17219    31001    11725    50728     1491   113246
#> 82:    17722    32166    10539    51606     1087   120987
#> 83:    17524    32650    10663    53821     1079   130086
#> 84:    17522    33261    10208    55605      885   140170
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesReportData
#>     region       date                           type median  mean     sd
#>  1: France 2020-08-04                       estimate   1153  1775   1978
#>  2: France 2020-08-05                       estimate    754  1174   1326
#>  3: France 2020-08-06                       estimate   1140  1740   1893
#>  4: France 2020-08-07                       estimate   1272  1943   2186
#>  5: France 2020-08-08                       estimate   1394  2154   2473
#>  6: France 2020-08-09                       estimate   1056  1672   1894
#>  7: France 2020-08-10                       estimate    958  1513   1802
#>  8: France 2020-08-11                       estimate   1582  2507   2860
#>  9: France 2020-08-12                       estimate   1134  1752   1961
#> 10: France 2020-08-13                       estimate   1692  2594   2908
#> 11: France 2020-08-14                       estimate   1793  2735   3054
#> 12: France 2020-08-15                       estimate   1938  3024   3339
#> 13: France 2020-08-16                       estimate   1583  2460   2724
#> 14: France 2020-08-17                       estimate   1344  2111   2514
#> 15: France 2020-08-18                       estimate   2370  3596   3971
#> 16: France 2020-08-19                       estimate   1632  2522   2902
#> 17: France 2020-08-20                       estimate   2399  3753   4220
#> 18: France 2020-08-21                       estimate   2639  4025   4593
#> 19: France 2020-08-22                       estimate   2832  4526   4961
#> 20: France 2020-08-23                       estimate   2248  3513   4035
#> 21: France 2020-08-24                       estimate   1920  3009   3391
#> 22: France 2020-08-25                       estimate   3210  4955   5497
#> 23: France 2020-08-26                       estimate   2223  3491   4061
#> 24: France 2020-08-27                       estimate   3331  5406   6252
#> 25: France 2020-08-28                       estimate   3634  5781   6610
#> 26: France 2020-08-29                       estimate   4064  6367   7268
#> 27: France 2020-08-30                       estimate   3131  4854   5552
#> 28: France 2020-08-31                       estimate   2595  4180   5006
#> 29: France 2020-09-01                       estimate   4344  6982   8191
#> 30: France 2020-09-02                       estimate   3025  4751   5619
#> 31: France 2020-09-03                       estimate   4626  7293   8296
#> 32: France 2020-09-04                       estimate   4707  7544   9182
#> 33: France 2020-09-05                       estimate   5098  8098   9165
#> 34: France 2020-09-06                       estimate   4118  6442   7486
#> 35: France 2020-09-07                       estimate   3462  5462   6491
#> 36: France 2020-09-08                       estimate   5866  9441  11046
#> 37: France 2020-09-09                       estimate   4030  6355   7610
#> 38: France 2020-09-10                       estimate   6120  9507  10998
#> 39: France 2020-09-11                       estimate   6483 10072  11793
#> 40: France 2020-09-12                       estimate   7168 10831  11696
#> 41: France 2020-09-13                       estimate   5076  8075   9269
#> 42: France 2020-09-14                       estimate   4577  7315   8405
#> 43: France 2020-09-15                       estimate   7297 11878  13545
#> 44: France 2020-09-16                       estimate   5102  7840   8660
#> 45: France 2020-09-17                       estimate   7219 11435  13396
#> 46: France 2020-09-18                       estimate   8074 12543  14826
#> 47: France 2020-09-19                       estimate   8752 13986  15786
#> 48: France 2020-09-20                       estimate   6622 10188  11458
#> 49: France 2020-09-21                       estimate   5606  8885  10509
#> 50: France 2020-09-22                       estimate   9506 15123  17453
#> 51: France 2020-09-23                       estimate   6304 10395  12273
#> 52: France 2020-09-24                       estimate   9230 14629  16778
#> 53: France 2020-09-25                       estimate   9500 15259  17231
#> 54: France 2020-09-26                       estimate  10738 16658  19102
#> 55: France 2020-09-27                       estimate   8290 12729  14307
#> 56: France 2020-09-28                       estimate   6967 10809  12412
#> 57: France 2020-09-29                       estimate  11490 18756  21641
#> 58: France 2020-09-30                       estimate   7554 11975  13608
#> 59: France 2020-10-01                       estimate  11797 18471  21252
#> 60: France 2020-10-02 estimate based on partial data  12460 19599  22111
#> 61: France 2020-10-03 estimate based on partial data  13064 20808  24919
#> 62: France 2020-10-04 estimate based on partial data   9942 15445  17418
#> 63: France 2020-10-05 estimate based on partial data   8720 13979  16445
#> 64: France 2020-10-06 estimate based on partial data  14329 23323  27893
#> 65: France 2020-10-07 estimate based on partial data   9578 16105  20216
#> 66: France 2020-10-08 estimate based on partial data  14581 23480  28909
#> 67: France 2020-10-09 estimate based on partial data  14812 24029  28953
#> 68: France 2020-10-10 estimate based on partial data  16354 26199  31684
#> 69: France 2020-10-11 estimate based on partial data  12690 20410  24066
#> 70: France 2020-10-12 estimate based on partial data  10728 17569  22515
#> 71: France 2020-10-13                       forecast  16642 29341  38183
#> 72: France 2020-10-14                       forecast  11768 20140  26275
#> 73: France 2020-10-15                       forecast  17382 30611  41642
#> 74: France 2020-10-16                       forecast  18702 33171  44716
#> 75: France 2020-10-17                       forecast  19900 35026  47918
#> 76: France 2020-10-18                       forecast  15080 27450  38635
#> 77: France 2020-10-19                       forecast  12684 23416  40226
#> 78: France 2020-10-20                       forecast  21781 42799  81837
#> 79: France 2020-10-21                       forecast  14088 29344  51364
#> 80: France 2020-10-22                       forecast  21673 47822 217378
#> 81: France 2020-10-23                       forecast  23018 48900  99108
#> 82: France 2020-10-24                       forecast  24454 53830 128186
#> 83: France 2020-10-25                       forecast  19107 42151 106431
#> 84: France 2020-10-26                       forecast  16261 39760 133516
#>     region       date                           type median  mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:        0      345        0     1153        0     4201
#>  2:        0      232        0      755        0     2746
#>  3:        0      346        0     1141        0     4016
#>  4:        0      373        0     1273        0     4607
#>  5:        0      408        0     1394        0     4884
#>  6:        0      322        0     1056        0     3996
#>  7:        0      308        0      959        0     3488
#>  8:        2      478        0     1582        0     5984
#>  9:        0      346        0     1134        0     4122
#> 10:       10      520        0     1692        0     6040
#> 11:        0      519        0     1793        0     6306
#> 12:        0      585        0     1938        0     7163
#> 13:        0      486        0     1583        0     5743
#> 14:        0      393        0     1346        0     4960
#> 15:        0      702        0     2371        0     8340
#> 16:        0      498        0     1632        0     5796
#> 17:        0      711        0     2399        0     8697
#> 18:        0      766       17     2656        0     9349
#> 19:        0      860        0     2833        0    10790
#> 20:        0      665        0     2250        0     8239
#> 21:        0      563        0     1921        0     7101
#> 22:        0      976        0     3211        0    11641
#> 23:        0      658        0     2225        0     8095
#> 24:        0     1051        0     3332        0    12839
#> 25:        0     1079        1     3635        0    13786
#> 26:        2     1216        0     4065        0    14919
#> 27:        0      907        0     3138        0    11331
#> 28:        0      774        0     2596        0     9892
#> 29:        1     1257        0     4344        0    16912
#> 30:        0      907        0     3025        0    11002
#> 31:        0     1343        0     4628        0    17002
#> 32:        0     1412        0     4708        0    17797
#> 33:        0     1534        0     5103        0    19225
#> 34:        0     1301        0     4118        0    15017
#> 35:        3      967        0     3465        0    12650
#> 36:        1     1713        1     5868        1    21984
#> 37:        0     1195        0     4030        0    14801
#> 38:        0     1721        0     6121        0    22105
#> 39:        0     1929        7     6488        5    23718
#> 40:        0     2198        0     7171        0    25419
#> 41:       11     1562        3     5077        3    18504
#> 42:        0     1386        0     4577        0    17120
#> 43:        0     2236        0     7298        0    27934
#> 44:        0     1556        0     5104        0    18291
#> 45:        1     2181        1     7222        1    26859
#> 46:        1     2398        1     8080        1    29430
#> 47:        6     2680        9     8759        1    32937
#> 48:        2     1866        0     6630        0    23890
#> 49:        0     1651        0     5606        0    20545
#> 50:        0     2782        0     9507        0    36074
#> 51:        2     1942        0     6308        0    25248
#> 52:      175     3084        0     9236        0    34431
#> 53:        1     2860        1     9500        1    36651
#> 54:       43     3236        7    10740        7    38407
#> 55:        2     2482        2     8293        3    29044
#> 56:        0     2181        0     6972        0    24849
#> 57:        0     3569        0    11495        0    44244
#> 58:        4     2424        4     7555        4    27607
#> 59:        0     3429        0    11798        0    43370
#> 60:        2     3604        2    12464        2    47102
#> 61:        2     3972        2    13064        2    48653
#> 62:        3     2986        3     9950        3    36269
#> 63:        0     2566        7     8733        0    32796
#> 64:        0     4529        0    14330        0    53955
#> 65:        0     2745        0     9580        0    38011
#> 66:        1     4278        1    14584        1    55042
#> 67:        0     4303        0    14813        0    56447
#> 68:        0     4912        0    16363        0    59886
#> 69:        1     4010        1    12693        1    48025
#> 70:        0     3164        0    10729        0    40626
#> 71:        3     5005       16    16659        3    70478
#> 72:        0     3496        0    11782        0    47378
#> 73:       42     5089       10    17394        4    71956
#> 74:        0     5345        0    18718        0    79874
#> 75:        0     5572        0    19902        0    83292
#> 76:        6     4461        6    15085        1    64553
#> 77:        0     3642        0    12685        0    55582
#> 78:        4     6101        4    21802        4   100904
#> 79:        1     4148        1    14092        1    70196
#> 80:      197     6641        0    21679        0   100715
#> 81:        3     6280       10    23030        3   110207
#> 82:        2     6372        2    24472        2   121536
#> 83:        0     5117        0    19108        0    90333
#> 84:       14     4107        0    16262        0    82657
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-07-20       0
#>  2: France 2020-07-21    2080
#>  3: France 2020-07-22     584
#>  4: France 2020-07-23     998
#>  5: France 2020-07-24    1062
#>  6: France 2020-07-25    1130
#>  7: France 2020-07-26       0
#>  8: France 2020-07-27       0
#>  9: France 2020-07-28    2551
#> 10: France 2020-07-29     725
#> 11: France 2020-07-30    1392
#> 12: France 2020-07-31    1377
#> 13: France 2020-08-01    1346
#> 14: France 2020-08-02       0
#> 15: France 2020-08-03       0
#> 16: France 2020-08-04    3376
#> 17: France 2020-08-05    1039
#> 18: France 2020-08-06    1695
#> 19: France 2020-08-07    1604
#> 20: France 2020-08-08    2288
#> 21: France 2020-08-09       0
#> 22: France 2020-08-10       0
#> 23: France 2020-08-11    4854
#> 24: France 2020-08-12    1397
#> 25: France 2020-08-13    2524
#> 26: France 2020-08-14    2669
#> 27: France 2020-08-15    2846
#> 28: France 2020-08-16    3310
#> 29: France 2020-08-17    3015
#> 30: France 2020-08-18     493
#> 31: France 2020-08-19    2238
#> 32: France 2020-08-20    3776
#> 33: France 2020-08-21    4771
#> 34: France 2020-08-22    4586
#> 35: France 2020-08-23    3602
#> 36: France 2020-08-24    4897
#> 37: France 2020-08-25    1955
#> 38: France 2020-08-26    3304
#> 39: France 2020-08-27    5429
#> 40: France 2020-08-28    6111
#> 41: France 2020-08-29    7379
#> 42: France 2020-08-30    5453
#> 43: France 2020-08-31    5413
#> 44: France 2020-09-01    3082
#> 45: France 2020-09-02    4982
#> 46: France 2020-09-03    7017
#> 47: France 2020-09-04    7157
#> 48: France 2020-09-05    8975
#> 49: France 2020-09-06    8550
#> 50: France 2020-09-07    7071
#> 51: France 2020-09-08    4203
#> 52: France 2020-09-09    6544
#> 53: France 2020-09-10    8577
#> 54: France 2020-09-11    9843
#> 55: France 2020-09-12    9406
#> 56: France 2020-09-13   10561
#> 57: France 2020-09-14    7183
#> 58: France 2020-09-15    6158
#> 59: France 2020-09-16    7852
#> 60: France 2020-09-17    9784
#> 61: France 2020-09-18   10593
#> 62: France 2020-09-19   13215
#> 63: France 2020-09-20   13498
#> 64: France 2020-09-21   11569
#> 65: France 2020-09-22    4298
#> 66: France 2020-09-23   10008
#> 67: France 2020-09-24   13072
#> 68: France 2020-09-25   16096
#> 69: France 2020-09-26   15797
#> 70: France 2020-09-27   14412
#> 71: France 2020-09-28   11123
#> 72: France 2020-09-29    4070
#> 73: France 2020-09-30    8051
#> 74: France 2020-10-01   12845
#> 75: France 2020-10-02   13970
#> 76: France 2020-10-03   12148
#> 77: France 2020-10-04   16972
#> 78: France 2020-10-05   12545
#> 79: France 2020-10-06    5104
#> 80: France 2020-10-07   10489
#> 81: France 2020-10-08   18746
#> 82: France 2020-10-09   18129
#> 83: France 2020-10-10   20339
#> 84: France 2020-10-11   26896
#> 85: France 2020-10-12   16101
#>     region       date confirm
#>