Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

rtData

#> $summaryData
#>              region New confirmed cases by infection date
#>   1:    Afghanistan                        65 (11 -- 186)
#>   2:        Albania                      475 (252 -- 772)
#>   3:        Algeria                      328 (159 -- 529)
#>   4:        Andorra                        91 (12 -- 216)
#>   5:         Angola                       231 (85 -- 452)
#>  ---                                                     
#> 193:        Vietnam                           6 (0 -- 25)
#> 194: Western Sahara                          1 (0 -- 289)
#> 195:          Yemen                            0 (0 -- 1)
#> 196:         Zambia                        47 (18 -- 100)
#> 197:       Zimbabwe                          27 (2 -- 93)
#>      Expected change in daily cases Effective reproduction no.
#>   1:                         Unsure             1 (0.8 -- 1.4)
#>   2:                     Increasing             1.2 (1 -- 1.5)
#>   3:              Likely increasing           1.2 (0.9 -- 1.5)
#>   4:              Likely increasing           1.1 (0.8 -- 1.4)
#>   5:              Likely increasing           1.1 (0.9 -- 1.3)
#>  ---                                                          
#> 193:                         Unsure           1.2 (0.5 -- 2.1)
#> 194:                         Unsure             0.6 (0 -- 1.4)
#> 195:              Likely decreasing           0.8 (0.5 -- 1.1)
#> 196:                         Unsure           0.9 (0.8 -- 1.2)
#> 197:                         Unsure           1.1 (0.7 -- 1.7)
#>             Rate of growth Doubling/halving time (days)
#>   1:  0.01 (-0.06 -- 0.11)          88.8 (6.6 -- -11.5)
#>   2:  0.06 (-0.01 -- 0.13)         11.4 (5.4 -- -127.6)
#>   3:  0.05 (-0.01 -- 0.12)          13.4 (5.7 -- -55.4)
#>   4:   0.04 (-0.05 -- 0.1)          19.4 (7.2 -- -14.9)
#>   5:  0.02 (-0.02 -- 0.08)          29.4 (8.9 -- -31.2)
#>  ---                                                   
#> 193:  0.05 (-0.16 -- 0.26)             13 (2.6 -- -4.3)
#> 194:  -0.1 (-0.36 -- 0.14)           -6.7 (4.8 -- -1.9)
#> 195: -0.05 (-0.16 -- 0.05)         -13.8 (14.8 -- -4.4)
#> 196: -0.02 (-0.07 -- 0.05)        -43.7 (15.1 -- -10.7)
#> 197:   0.02 (-0.09 -- 0.2)           29.4 (3.5 -- -7.9)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-08-12 estimate   0.99 0.99 0.06     0.99     1.00
#>     2: Afghanistan 2020-08-13 estimate   0.99 0.98 0.06     0.99     1.01
#>     3: Afghanistan 2020-08-14 estimate   0.99 0.98 0.06     0.98     1.00
#>     4: Afghanistan 2020-08-15 estimate   0.99 0.98 0.06     0.99     1.01
#>     5: Afghanistan 2020-08-16 estimate   0.99 0.97 0.06     0.99     1.00
#>    ---                                                                   
#> 16388:    Zimbabwe 2020-10-31 forecast   1.09 1.17 0.39     0.86     0.97
#> 16389:    Zimbabwe 2020-11-01 forecast   1.09 1.17 0.39     0.86     0.97
#> 16390:    Zimbabwe 2020-11-02 forecast   1.09 1.17 0.39     0.86     0.97
#> 16391:    Zimbabwe 2020-11-03 forecast   1.09 1.17 0.39     0.86     0.97
#> 16392:    Zimbabwe 2020-11-04 forecast   1.09 1.17 0.39     0.86     0.97
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.97     1.02     0.90     1.08
#>     2:     0.97     1.02     0.90     1.07
#>     3:     0.97     1.02     0.89     1.07
#>     4:     0.97     1.02     0.88     1.07
#>     5:     0.97     1.02     0.88     1.06
#>    ---                                    
#> 16388:     0.86     1.20     0.66     1.75
#> 16389:     0.86     1.20     0.66     1.75
#> 16390:     0.86     1.20     0.66     1.75
#> 16391:     0.86     1.20     0.66     1.75
#> 16392:     0.86     1.20     0.66     1.75
#> 
#> $casesInfectionData
#>             region       date     type median  mean     sd lower_20 upper_20
#>     1: Afghanistan 2020-08-12 estimate     50    53     13       47       51
#>     2: Afghanistan 2020-08-13 estimate     50    53     13       47       51
#>     3: Afghanistan 2020-08-14 estimate     50    52     13       46       50
#>     4: Afghanistan 2020-08-15 estimate     50    52     13       46       50
#>     5: Afghanistan 2020-08-16 estimate     49    51     13       47       51
#>    ---                                                                      
#> 16388:    Zimbabwe 2020-10-31 forecast     34  1682  49798        5       12
#> 16389:    Zimbabwe 2020-11-01 forecast     35  3106 102130        4       12
#> 16390:    Zimbabwe 2020-11-02 forecast     35  5908 212244        4       12
#> 16391:    Zimbabwe 2020-11-03 forecast     36 11503 445663        4       11
#> 16392:    Zimbabwe 2020-11-04 forecast     37 22813 943106        3       11
#>        lower_50 upper_50 lower_90 upper_90
#>     1:       42       54       35       72
#>     2:       43       54       34       71
#>     3:       43       54       32       69
#>     4:       43       55       33       70
#>     5:       42       54       32       69
#>    ---                                    
#> 16388:        2       36        0      426
#> 16389:        0       35        0      492
#> 16390:        0       35        0      566
#> 16391:        0       36        0      655
#> 16392:        0       37        0      761
#> 
#> $casesReportData
#>             region       date     type median mean    sd lower_20 upper_20
#>     1: Afghanistan 2020-08-12 estimate     31   63    90        0        5
#>     2: Afghanistan 2020-08-13 estimate     25   51    75        0        4
#>     3: Afghanistan 2020-08-14 estimate     21   44    62        0        4
#>     4: Afghanistan 2020-08-15 estimate     23   48    85        0        3
#>     5: Afghanistan 2020-08-16 estimate     25   51    72        0        4
#>    ---                                                                    
#> 16388:    Zimbabwe 2020-10-31 forecast     24  152  2605        0        7
#> 16389:    Zimbabwe 2020-11-01 forecast     24  197  3480        1        7
#> 16390:    Zimbabwe 2020-11-02 forecast     19  371 11957        0        5
#> 16391:    Zimbabwe 2020-11-03 forecast     28  310  6779        0        8
#> 16392:    Zimbabwe 2020-11-04 forecast     25  593 12972        0        7
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0       31        0      169
#>     2:        0       25        0      133
#>     3:        0       21        0      112
#>     4:        0       23        0      122
#>     5:        0       25        0      133
#>    ---                                    
#> 16388:        0       24        0      177
#> 16389:        0       24        0      178
#> 16390:        0       19        0      136
#> 16391:        0       28        0      256
#> 16392:        0       25        0      254
#> 
#> $obsCasesData
#>                region       date confirm
#>     1:    Afghanistan 2020-07-29     103
#>     2:        Albania 2020-07-29     117
#>     3:        Algeria 2020-07-29     642
#>     4:        Andorra 2020-07-29       0
#>     5:         Angola 2020-07-29      18
#>    ---                                  
#> 17767:        Vietnam 2020-10-21       1
#> 17768: Western Sahara 2020-10-21       0
#> 17769:          Yemen 2020-10-21       0
#> 17770:         Zambia 2020-10-21      85
#> 17771:       Zimbabwe 2020-10-21      28
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France                35546 (11619 -- 67774)              Likely increasing
#>    Effective reproduction no.       Rate of growth Doubling/halving time (days)
#> 1:           1.1 (0.9 -- 1.4) 0.03 (-0.03 -- 0.09)          20.4 (7.3 -- -27.1)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-08-12                       estimate   1.22 1.22 0.05     1.20
#>  2: France 2020-08-13                       estimate   1.21 1.22 0.05     1.20
#>  3: France 2020-08-14                       estimate   1.21 1.21 0.05     1.20
#>  4: France 2020-08-15                       estimate   1.20 1.21 0.05     1.19
#>  5: France 2020-08-16                       estimate   1.20 1.20 0.05     1.19
#>  6: France 2020-08-17                       estimate   1.20 1.20 0.05     1.18
#>  7: France 2020-08-18                       estimate   1.19 1.20 0.05     1.18
#>  8: France 2020-08-19                       estimate   1.19 1.19 0.05     1.17
#>  9: France 2020-08-20                       estimate   1.19 1.19 0.05     1.17
#> 10: France 2020-08-21                       estimate   1.18 1.18 0.05     1.16
#> 11: France 2020-08-22                       estimate   1.18 1.18 0.05     1.17
#> 12: France 2020-08-23                       estimate   1.18 1.18 0.05     1.16
#> 13: France 2020-08-24                       estimate   1.17 1.17 0.05     1.17
#> 14: France 2020-08-25                       estimate   1.17 1.17 0.05     1.16
#> 15: France 2020-08-26                       estimate   1.17 1.16 0.05     1.16
#> 16: France 2020-08-27                       estimate   1.16 1.16 0.05     1.15
#> 17: France 2020-08-28                       estimate   1.16 1.16 0.05     1.15
#> 18: France 2020-08-29                       estimate   1.16 1.15 0.05     1.15
#> 19: France 2020-08-30                       estimate   1.15 1.15 0.05     1.15
#> 20: France 2020-08-31                       estimate   1.15 1.15 0.05     1.14
#> 21: France 2020-09-01                       estimate   1.15 1.14 0.05     1.14
#> 22: France 2020-09-02                       estimate   1.14 1.14 0.05     1.14
#> 23: France 2020-09-03                       estimate   1.14 1.14 0.05     1.14
#> 24: France 2020-09-04                       estimate   1.14 1.13 0.05     1.13
#> 25: France 2020-09-05                       estimate   1.13 1.13 0.05     1.14
#> 26: France 2020-09-06                       estimate   1.13 1.13 0.05     1.12
#> 27: France 2020-09-07                       estimate   1.13 1.12 0.05     1.12
#> 28: France 2020-09-08                       estimate   1.12 1.12 0.05     1.12
#> 29: France 2020-09-09                       estimate   1.12 1.12 0.05     1.12
#> 30: France 2020-09-10                       estimate   1.12 1.11 0.05     1.12
#> 31: France 2020-09-11                       estimate   1.12 1.11 0.05     1.11
#> 32: France 2020-09-12                       estimate   1.11 1.11 0.05     1.11
#> 33: France 2020-09-13                       estimate   1.11 1.11 0.05     1.10
#> 34: France 2020-09-14                       estimate   1.11 1.11 0.05     1.10
#> 35: France 2020-09-15                       estimate   1.11 1.10 0.05     1.10
#> 36: France 2020-09-16                       estimate   1.11 1.10 0.06     1.10
#> 37: France 2020-09-17                       estimate   1.11 1.10 0.06     1.10
#> 38: France 2020-09-18                       estimate   1.10 1.10 0.06     1.10
#> 39: France 2020-09-19                       estimate   1.10 1.10 0.06     1.10
#> 40: France 2020-09-20                       estimate   1.10 1.10 0.06     1.10
#> 41: France 2020-09-21                       estimate   1.10 1.10 0.06     1.09
#> 42: France 2020-09-22                       estimate   1.10 1.10 0.05     1.09
#> 43: France 2020-09-23                       estimate   1.10 1.10 0.05     1.09
#> 44: France 2020-09-24                       estimate   1.10 1.10 0.05     1.10
#> 45: France 2020-09-25                       estimate   1.10 1.10 0.05     1.09
#> 46: France 2020-09-26                       estimate   1.10 1.10 0.05     1.08
#> 47: France 2020-09-27                       estimate   1.11 1.11 0.05     1.09
#> 48: France 2020-09-28                       estimate   1.11 1.11 0.05     1.09
#> 49: France 2020-09-29                       estimate   1.11 1.11 0.05     1.11
#> 50: France 2020-09-30                       estimate   1.11 1.11 0.06     1.11
#> 51: France 2020-10-01                       estimate   1.11 1.11 0.06     1.11
#> 52: France 2020-10-02                       estimate   1.12 1.12 0.06     1.12
#> 53: France 2020-10-03                       estimate   1.12 1.12 0.06     1.12
#> 54: France 2020-10-04                       estimate   1.12 1.12 0.06     1.11
#> 55: France 2020-10-05                       estimate   1.12 1.12 0.07     1.10
#> 56: France 2020-10-06                       estimate   1.12 1.12 0.07     1.11
#> 57: France 2020-10-07                       estimate   1.12 1.13 0.07     1.11
#> 58: France 2020-10-08                       estimate   1.12 1.13 0.08     1.10
#> 59: France 2020-10-09                       estimate   1.12 1.13 0.08     1.10
#> 60: France 2020-10-10                       estimate   1.13 1.13 0.09     1.09
#> 61: France 2020-10-11 estimate based on partial data   1.13 1.13 0.09     1.10
#> 62: France 2020-10-12 estimate based on partial data   1.13 1.13 0.10     1.09
#> 63: France 2020-10-13 estimate based on partial data   1.13 1.13 0.11     1.09
#> 64: France 2020-10-14 estimate based on partial data   1.13 1.14 0.11     1.09
#> 65: France 2020-10-15 estimate based on partial data   1.13 1.14 0.12     1.11
#> 66: France 2020-10-16 estimate based on partial data   1.13 1.14 0.13     1.10
#> 67: France 2020-10-17 estimate based on partial data   1.13 1.14 0.13     1.11
#> 68: France 2020-10-18 estimate based on partial data   1.13 1.14 0.14     1.10
#> 69: France 2020-10-19 estimate based on partial data   1.13 1.14 0.14     1.10
#> 70: France 2020-10-20 estimate based on partial data   1.13 1.14 0.15     1.11
#> 71: France 2020-10-21 estimate based on partial data   1.13 1.14 0.16     1.10
#> 72: France 2020-10-22                       forecast   1.13 1.14 0.16     1.10
#> 73: France 2020-10-23                       forecast   1.13 1.14 0.16     1.10
#> 74: France 2020-10-24                       forecast   1.13 1.14 0.16     1.10
#> 75: France 2020-10-25                       forecast   1.13 1.14 0.16     1.10
#> 76: France 2020-10-26                       forecast   1.13 1.14 0.16     1.10
#> 77: France 2020-10-27                       forecast   1.13 1.14 0.16     1.10
#> 78: France 2020-10-28                       forecast   1.13 1.14 0.16     1.10
#> 79: France 2020-10-29                       forecast   1.13 1.14 0.16     1.10
#> 80: France 2020-10-30                       forecast   1.13 1.14 0.16     1.10
#> 81: France 2020-10-31                       forecast   1.13 1.14 0.16     1.10
#> 82: France 2020-11-01                       forecast   1.13 1.14 0.16     1.10
#> 83: France 2020-11-02                       forecast   1.13 1.14 0.16     1.10
#> 84: France 2020-11-03                       forecast   1.13 1.14 0.16     1.10
#> 85: France 2020-11-04                       forecast   1.13 1.14 0.16     1.10
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.22     1.18     1.24     1.15     1.30
#>  2:     1.22     1.18     1.24     1.14     1.30
#>  3:     1.22     1.17     1.23     1.13     1.30
#>  4:     1.21     1.17     1.22     1.13     1.29
#>  5:     1.21     1.17     1.22     1.12     1.29
#>  6:     1.20     1.16     1.22     1.12     1.28
#>  7:     1.20     1.16     1.21     1.11     1.27
#>  8:     1.19     1.17     1.22     1.11     1.28
#>  9:     1.19     1.16     1.21     1.11     1.28
#> 10:     1.18     1.16     1.21     1.10     1.27
#> 11:     1.19     1.15     1.20     1.09     1.26
#> 12:     1.18     1.15     1.20     1.09     1.26
#> 13:     1.18     1.14     1.20     1.09     1.26
#> 14:     1.18     1.14     1.20     1.09     1.25
#> 15:     1.18     1.14     1.19     1.08     1.25
#> 16:     1.17     1.14     1.19     1.07     1.24
#> 17:     1.17     1.13     1.18     1.08     1.25
#> 18:     1.17     1.13     1.18     1.07     1.23
#> 19:     1.17     1.13     1.18     1.07     1.23
#> 20:     1.16     1.13     1.18     1.06     1.23
#> 21:     1.16     1.12     1.18     1.05     1.22
#> 22:     1.16     1.12     1.18     1.05     1.21
#> 23:     1.16     1.11     1.17     1.05     1.21
#> 24:     1.15     1.11     1.17     1.05     1.21
#> 25:     1.16     1.11     1.17     1.04     1.21
#> 26:     1.14     1.11     1.17     1.03     1.20
#> 27:     1.14     1.10     1.16     1.04     1.20
#> 28:     1.14     1.10     1.16     1.03     1.20
#> 29:     1.14     1.10     1.16     1.03     1.20
#> 30:     1.14     1.10     1.16     1.03     1.20
#> 31:     1.13     1.10     1.16     1.03     1.19
#> 32:     1.13     1.09     1.16     1.02     1.19
#> 33:     1.13     1.09     1.15     1.02     1.18
#> 34:     1.13     1.08     1.15     1.02     1.19
#> 35:     1.12     1.09     1.15     1.02     1.19
#> 36:     1.12     1.07     1.14     1.01     1.18
#> 37:     1.12     1.08     1.14     1.01     1.18
#> 38:     1.12     1.08     1.15     1.01     1.18
#> 39:     1.12     1.08     1.14     1.01     1.18
#> 40:     1.13     1.08     1.14     1.02     1.19
#> 41:     1.11     1.08     1.14     1.02     1.19
#> 42:     1.11     1.08     1.14     1.02     1.19
#> 43:     1.11     1.07     1.14     1.02     1.19
#> 44:     1.12     1.08     1.14     1.02     1.19
#> 45:     1.11     1.08     1.14     1.02     1.18
#> 46:     1.11     1.07     1.14     1.02     1.18
#> 47:     1.11     1.08     1.14     1.02     1.18
#> 48:     1.12     1.08     1.14     1.02     1.19
#> 49:     1.14     1.08     1.15     1.02     1.19
#> 50:     1.14     1.08     1.15     1.03     1.20
#> 51:     1.14     1.08     1.15     1.03     1.20
#> 52:     1.15     1.08     1.15     1.02     1.21
#> 53:     1.15     1.08     1.15     1.02     1.21
#> 54:     1.14     1.08     1.16     1.01     1.21
#> 55:     1.13     1.08     1.16     1.01     1.22
#> 56:     1.14     1.08     1.16     1.01     1.23
#> 57:     1.14     1.08     1.16     1.01     1.25
#> 58:     1.14     1.08     1.17     1.01     1.26
#> 59:     1.14     1.08     1.18     1.00     1.26
#> 60:     1.13     1.08     1.17     0.99     1.27
#> 61:     1.14     1.07     1.18     0.99     1.29
#> 62:     1.13     1.07     1.18     0.98     1.30
#> 63:     1.13     1.07     1.18     0.95     1.29
#> 64:     1.13     1.07     1.18     0.95     1.30
#> 65:     1.15     1.07     1.18     0.95     1.31
#> 66:     1.15     1.07     1.19     0.94     1.32
#> 67:     1.15     1.06     1.18     0.93     1.32
#> 68:     1.15     1.06     1.18     0.92     1.33
#> 69:     1.15     1.06     1.18     0.90     1.33
#> 70:     1.16     1.06     1.19     0.91     1.35
#> 71:     1.15     1.06     1.19     0.90     1.35
#> 72:     1.15     1.06     1.19     0.90     1.35
#> 73:     1.15     1.06     1.19     0.90     1.35
#> 74:     1.15     1.06     1.19     0.90     1.35
#> 75:     1.15     1.06     1.19     0.90     1.35
#> 76:     1.15     1.06     1.19     0.90     1.35
#> 77:     1.15     1.06     1.19     0.90     1.35
#> 78:     1.15     1.06     1.19     0.90     1.35
#> 79:     1.15     1.06     1.19     0.90     1.35
#> 80:     1.15     1.06     1.19     0.90     1.35
#> 81:     1.15     1.06     1.19     0.90     1.35
#> 82:     1.15     1.06     1.19     0.90     1.35
#> 83:     1.15     1.06     1.19     0.90     1.35
#> 84:     1.15     1.06     1.19     0.90     1.35
#> 85:     1.15     1.06     1.19     0.90     1.35
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData
#>     region       date                           type median   mean     sd
#>  1: France 2020-08-12                       estimate   3263   3357    622
#>  2: France 2020-08-13                       estimate   3441   3539    659
#>  3: France 2020-08-14                       estimate   3625   3728    698
#>  4: France 2020-08-15                       estimate   3815   3922    736
#>  5: France 2020-08-16                       estimate   4019   4122    775
#>  6: France 2020-08-17                       estimate   4223   4328    815
#>  7: France 2020-08-18                       estimate   4427   4539    856
#>  8: France 2020-08-19                       estimate   4641   4755    897
#>  9: France 2020-08-20                       estimate   4856   4977    939
#> 10: France 2020-08-21                       estimate   5087   5204    982
#> 11: France 2020-08-22                       estimate   5312   5436   1025
#> 12: France 2020-08-23                       estimate   5549   5672   1068
#> 13: France 2020-08-24                       estimate   5782   5914   1111
#> 14: France 2020-08-25                       estimate   6023   6159   1154
#> 15: France 2020-08-26                       estimate   6276   6408   1196
#> 16: France 2020-08-27                       estimate   6525   6662   1238
#> 17: France 2020-08-28                       estimate   6776   6919   1279
#> 18: France 2020-08-29                       estimate   7016   7179   1320
#> 19: France 2020-08-30                       estimate   7278   7443   1361
#> 20: France 2020-08-31                       estimate   7532   7710   1401
#> 21: France 2020-09-01                       estimate   7802   7980   1441
#> 22: France 2020-09-02                       estimate   8072   8252   1482
#> 23: France 2020-09-03                       estimate   8342   8526   1523
#> 24: France 2020-09-04                       estimate   8620   8803   1564
#> 25: France 2020-09-05                       estimate   8902   9081   1605
#> 26: France 2020-09-06                       estimate   9180   9361   1645
#> 27: France 2020-09-07                       estimate   9445   9643   1683
#> 28: France 2020-09-08                       estimate   9717   9926   1721
#> 29: France 2020-09-09                       estimate  10009  10211   1758
#> 30: France 2020-09-10                       estimate  10314  10497   1796
#> 31: France 2020-09-11                       estimate  10602  10787   1836
#> 32: France 2020-09-12                       estimate  10880  11079   1879
#> 33: France 2020-09-13                       estimate  11184  11374   1927
#> 34: France 2020-09-14                       estimate  11497  11674   1979
#> 35: France 2020-09-15                       estimate  11805  11979   2036
#> 36: France 2020-09-16                       estimate  12115  12291   2098
#> 37: France 2020-09-17                       estimate  12446  12610   2163
#> 38: France 2020-09-18                       estimate  12785  12937   2232
#> 39: France 2020-09-19                       estimate  13116  13275   2305
#> 40: France 2020-09-20                       estimate  13452  13623   2381
#> 41: France 2020-09-21                       estimate  13826  13983   2462
#> 42: France 2020-09-22                       estimate  14230  14358   2548
#> 43: France 2020-09-23                       estimate  14582  14747   2639
#> 44: France 2020-09-24                       estimate  14987  15154   2736
#> 45: France 2020-09-25                       estimate  15369  15578   2839
#> 46: France 2020-09-26                       estimate  15836  16023   2948
#> 47: France 2020-09-27                       estimate  16273  16489   3065
#> 48: France 2020-09-28                       estimate  16767  16978   3189
#> 49: France 2020-09-29                       estimate  17204  17492   3323
#> 50: France 2020-09-30                       estimate  17721  18033   3468
#> 51: France 2020-10-01                       estimate  18274  18601   3626
#> 52: France 2020-10-02                       estimate  18790  19199   3800
#> 53: France 2020-10-03                       estimate  19391  19829   3993
#> 54: France 2020-10-04                       estimate  20026  20492   4210
#> 55: France 2020-10-05                       estimate  20646  21189   4455
#> 56: France 2020-10-06                       estimate  21293  21924   4737
#> 57: France 2020-10-07                       estimate  22022  22699   5063
#> 58: France 2020-10-08                       estimate  22761  23517   5443
#> 59: France 2020-10-09                       estimate  23510  24382   5890
#> 60: France 2020-10-10                       estimate  24378  25297   6420
#> 61: France 2020-10-11 estimate based on partial data  25221  26269   7051
#> 62: France 2020-10-12 estimate based on partial data  26132  27303   7804
#> 63: France 2020-10-13 estimate based on partial data  27039  28406   8706
#> 64: France 2020-10-14 estimate based on partial data  27998  29588   9792
#> 65: France 2020-10-15 estimate based on partial data  29012  30859  11108
#> 66: France 2020-10-16 estimate based on partial data  30079  32231  12715
#> 67: France 2020-10-17 estimate based on partial data  31084  33719  14696
#> 68: France 2020-10-18 estimate based on partial data  32218  35343  17158
#> 69: France 2020-10-19 estimate based on partial data  33337  37127  20250
#> 70: France 2020-10-20 estimate based on partial data  34452  39099  24170
#> 71: France 2020-10-21 estimate based on partial data  35546  41299  29192
#> 72: France 2020-10-22                       forecast  36682  43724  35416
#> 73: France 2020-10-23                       forecast  37848  46431  43343
#> 74: France 2020-10-24                       forecast  39163  49468  53476
#> 75: France 2020-10-25                       forecast  40456  52896  66480
#> 76: France 2020-10-26                       forecast  41825  56789  83226
#> 77: France 2020-10-27                       forecast  43099  61240 104858
#> 78: France 2020-10-28                       forecast  44518  66365 132877
#> 79: France 2020-10-29                       forecast  46043  72309 169256
#> 80: France 2020-10-30                       forecast  47725  79254 216595
#> 81: France 2020-10-31                       forecast  49362  87431 278317
#> 82: France 2020-11-01                       forecast  51048  97131 358936
#> 83: France 2020-11-02                       forecast  52717 108725 464414
#> 84: France 2020-11-03                       forecast  54581 122688 602630
#> 85: France 2020-11-04                       forecast  56415 139625 784003
#>     region       date                           type median   mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     2906     3165     2792     3512     2441     4263
#>  2:     3052     3336     2932     3706     2549     4484
#>  3:     3234     3544     3089     3921     2726     4789
#>  4:     3557     3883     3309     4193     2838     5025
#>  5:     3723     4068     3483     4421     2971     5282
#>  6:     3749     4112     3619     4611     3040     5485
#>  7:     4116     4496     3799     4848     3178     5749
#>  8:     4344     4740     3900     4998     3349     6040
#>  9:     4563     4980     4039     5193     3529     6348
#> 10:     4784     5215     4218     5412     3717     6690
#> 11:     5062     5511     4402     5647     3841     6939
#> 12:     5303     5781     4612     5918     4010     7230
#> 13:     5464     5965     4882     6237     4235     7604
#> 14:     5758     6276     5009     6422     4302     7822
#> 15:     6051     6581     5211     6662     4546     8220
#> 16:     6289     6852     5449     6954     4782     8593
#> 17:     6353     6934     5854     7407     4932     8858
#> 18:     6634     7232     5912     7508     5074     9108
#> 19:     6907     7504     6094     7727     5318     9480
#> 20:     7175     7795     6411     8091     5459     9774
#> 21:     7037     7674     6687     8414     5774    10193
#> 22:     7382     8037     6914     8686     5921    10486
#> 23:     7684     8365     7102     8931     6240    10902
#> 24:     7935     8638     7478     9347     6365    11141
#> 25:     8065     8794     7805     9731     6631    11495
#> 26:     8247     8995     8002     9972     6724    11730
#> 27:     8571     9334     8277    10312     6887    12039
#> 28:     8847     9630     8453    10537     7409    12676
#> 29:     9176     9971     8692    10822     7656    13061
#> 30:     9331    10148     8930    11121     7485    13030
#> 31:     9868    10701     9194    11443     7793    13426
#> 32:    10216    11066     9563    11857     8024    13866
#> 33:    10459    11332     9739    12083     8305    14288
#> 34:    10441    11349    10053    12464     8523    14647
#> 35:    11318    12243    10262    12745     8555    14833
#> 36:    11405    12352    10503    13032     8872    15327
#> 37:    11906    12864    10764    13395     9207    15886
#> 38:    12242    13252    11260    14020     9364    16258
#> 39:    12445    13517    11317    14202     9542    16688
#> 40:    12867    13970    11643    14656     9771    17187
#> 41:    12971    14134    11947    15089     9918    17688
#> 42:    13290    14476    12210    15480    10212    18313
#> 43:    13727    14955    12571    15958    10262    18658
#> 44:    14351    15614    12886    16397    10566    19314
#> 45:    14583    15922    13246    16885    10988    20013
#> 46:    14725    16096    13615    17374    11430    20820
#> 47:    15101    16516    13890    17774    11680    21420
#> 48:    15632    17119    14429    18431    12051    22157
#> 49:    15456    16999    14775    18962    12026    22611
#> 50:    15916    17497    14935    19308    12651    23649
#> 51:    16380    18047    15228    19805    12705    24147
#> 52:    17140    18887    15481    20250    13371    25375
#> 53:    17350    19169    16158    21161    13401    25946
#> 54:    18851    20729    16694    21922    13683    27006
#> 55:    18808    20795    16513    22009    14061    28247
#> 56:    19603    21684    17582    23317    14301    29334
#> 57:    19892    22112    18336    24421    14679    30653
#> 58:    20508    22831    18238    24735    14984    32061
#> 59:    21006    23478    19312    26264    15225    33615
#> 60:    21505    24212    19338    26902    15658    35613
#> 61:    22222    25187    19432    27665    15681    37438
#> 62:    23128    26386    20724    29736    15649    39413
#> 63:    22900    26468    20860    30736    15662    41821
#> 64:    23413    27355    21377    32230    14976    43668
#> 65:    23882    28252    21625    33542    15112    46867
#> 66:    24986    29794    22238    35314    14242    49427
#> 67:    23741    29034    22327    36679    13648    52648
#> 68:    24080    29901    21946    37604    13200    55552
#> 69:    24347    30714    21651    38553    13244    59734
#> 70:    25237    32183    22042    40486    11646    62657
#> 71:    27935    35365    20000    40041    11619    67774
#> 72:    28277    36191    20034    41673    10673    71993
#> 73:    28456    36874    20750    44025     9987    77320
#> 74:    27709    36715    20888    45861     9241    82789
#> 75:    28435    38136    19873    46616     8900    89408
#> 76:    28503    38872    20935    49513     6450    94437
#> 77:    29016    40195    20236    50655     7387   102279
#> 78:    30273    42188    18938    51185     6985   109669
#> 79:    30951    43596    18441    52601     6560   117696
#> 80:    30459    44024    18668    54738     5576   126154
#> 81:    30900    45289    18731    56908      906   131209
#> 82:    31199    46363    18459    58986      761   140941
#> 83:    19767    35810    18629    61488     2335   153450
#> 84:    19669    36423    17184    62344     1825   164420
#> 85:    19929    37406    16884    64531     1890   177639
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesReportData
#>     region       date                           type median  mean     sd
#>  1: France 2020-08-12                       estimate   1186  1624   1542
#>  2: France 2020-08-13                       estimate   1931  2590   2385
#>  3: France 2020-08-14                       estimate   2048  2812   2637
#>  4: France 2020-08-15                       estimate   2154  2977   2751
#>  5: France 2020-08-16                       estimate   1937  2642   2491
#>  6: France 2020-08-17                       estimate   1772  2343   2132
#>  7: France 2020-08-18                       estimate   1784  2461   2277
#>  8: France 2020-08-19                       estimate   1835  2502   2311
#>  9: France 2020-08-20                       estimate   2737  3755   3509
#> 10: France 2020-08-21                       estimate   3112  4190   3835
#> 11: France 2020-08-22                       estimate   3298  4509   4262
#> 12: France 2020-08-23                       estimate   2930  3933   3712
#> 13: France 2020-08-24                       estimate   2470  3365   3101
#> 14: France 2020-08-25                       estimate   2532  3452   3160
#> 15: France 2020-08-26                       estimate   2624  3542   3254
#> 16: France 2020-08-27                       estimate   3890  5307   4974
#> 17: France 2020-08-28                       estimate   4252  5733   5287
#> 18: France 2020-08-29                       estimate   4615  6307   6121
#> 19: France 2020-08-30                       estimate   3970  5328   4914
#> 20: France 2020-08-31                       estimate   3507  4747   4417
#> 21: France 2020-09-01                       estimate   3461  4634   4154
#> 22: France 2020-09-02                       estimate   3577  4870   4603
#> 23: France 2020-09-03                       estimate   5308  7258   6861
#> 24: France 2020-09-04                       estimate   5624  7670   7206
#> 25: France 2020-09-05                       estimate   5984  8131   7532
#> 26: France 2020-09-06                       estimate   5128  6983   6561
#> 27: France 2020-09-07                       estimate   4550  6136   5696
#> 28: France 2020-09-08                       estimate   4572  6283   5884
#> 29: France 2020-09-09                       estimate   4488  6136   5658
#> 30: France 2020-09-10                       estimate   6796  9152   8477
#> 31: France 2020-09-11                       estimate   7370 10050   9624
#> 32: France 2020-09-12                       estimate   7486 10381  10072
#> 33: France 2020-09-13                       estimate   6582  8911   8333
#> 34: France 2020-09-14                       estimate   5568  7663   7104
#> 35: France 2020-09-15                       estimate   5757  7823   7328
#> 36: France 2020-09-16                       estimate   5670  7778   7425
#> 37: France 2020-09-17                       estimate   8581 11459  10422
#> 38: France 2020-09-18                       estimate   8996 12152  11475
#> 39: France 2020-09-19                       estimate   9111 12503  11555
#> 40: France 2020-09-20                       estimate   8147 10592   9443
#> 41: France 2020-09-21                       estimate   7095  9376   8220
#> 42: France 2020-09-22                       estimate   7030  9466   8903
#> 43: France 2020-09-23                       estimate   6794  9218   8461
#> 44: France 2020-09-24                       estimate  10000 13578  12447
#> 45: France 2020-09-25                       estimate  10975 14789  13831
#> 46: France 2020-09-26                       estimate  11390 15243  13823
#> 47: France 2020-09-27                       estimate  10062 13535  12471
#> 48: France 2020-09-28                       estimate   8176 11099  10291
#> 49: France 2020-09-29                       estimate   8111 11111  10587
#> 50: France 2020-09-30                       estimate   8170 11295  10651
#> 51: France 2020-10-01                       estimate  12678 17049  15429
#> 52: France 2020-10-02                       estimate  13476 18287  17041
#> 53: France 2020-10-03                       estimate  14185 18971  17849
#> 54: France 2020-10-04                       estimate  11602 15663  14322
#> 55: France 2020-10-05                       estimate  10134 13753  13018
#> 56: France 2020-10-06                       estimate  10104 13745  12904
#> 57: France 2020-10-07                       estimate   9957 13475  12548
#> 58: France 2020-10-08                       estimate  15670 21308  20257
#> 59: France 2020-10-09                       estimate  16127 21533  19718
#> 60: France 2020-10-10                       estimate  17025 23269  21690
#> 61: France 2020-10-11 estimate based on partial data  14802 19674  18141
#> 62: France 2020-10-12 estimate based on partial data  12497 16730  15729
#> 63: France 2020-10-13 estimate based on partial data  12709 17325  16494
#> 64: France 2020-10-14 estimate based on partial data  12331 17021  16060
#> 65: France 2020-10-15 estimate based on partial data  19128 25977  24464
#> 66: France 2020-10-16 estimate based on partial data  19904 28204  27500
#> 67: France 2020-10-17 estimate based on partial data  21640 30554  29695
#> 68: France 2020-10-18 estimate based on partial data  17938 25252  24660
#> 69: France 2020-10-19 estimate based on partial data  15786 21678  20331
#> 70: France 2020-10-20 estimate based on partial data  16708 23340  22922
#> 71: France 2020-10-21 estimate based on partial data  16185 22859  22719
#> 72: France 2020-10-22                       forecast  23548 34066  35949
#> 73: France 2020-10-23                       forecast  25448 36559  36625
#> 74: France 2020-10-24                       forecast  26422 38807  41643
#> 75: France 2020-10-25                       forecast  22658 34310  39061
#> 76: France 2020-10-26                       forecast  20060 30099  34535
#> 77: France 2020-10-27                       forecast  20132 31357  45352
#> 78: France 2020-10-28                       forecast  19522 31871  39732
#> 79: France 2020-10-29                       forecast  29804 48320  73656
#> 80: France 2020-10-30                       forecast  31746 52313  72057
#> 81: France 2020-10-31                       forecast  33861 59371 133962
#> 82: France 2020-11-01                       forecast  29172 54804 120989
#> 83: France 2020-11-02                       forecast  24102 45437 103721
#> 84: France 2020-11-03                       forecast  25463 50906 178853
#> 85: France 2020-11-04                       forecast  25382 50567 121650
#>     region       date                           type median  mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:      264      672       85     1253        1     3539
#>  2:      522     1180      264     2151        4     5477
#>  3:      316     1034       29     2061        2     6083
#>  4:      545     1312      173     2283       12     6574
#>  5:      376     1060       93     2010        1     5722
#>  6:      396     1012      111     1860        0     4987
#>  7:      548     1160      147     1902       10     5419
#>  8:      208      853      100     1897        0     5533
#>  9:      623     1598      144     2817        2     8067
#> 10:      432     1494      188     3270        0     9042
#> 11:      812     1964      163     3433       21     9647
#> 12:      350     1376      163     3026       23     8472
#> 13:      485     1354      117     2563        0     7302
#> 14:      321     1223      129     2616        8     7567
#> 15:      640     1551      194     2746        2     7664
#> 16:      462     1824      157     3978        0    11510
#> 17:     1152     2589      396     4553        1    12260
#> 18:      712     2283      242     4807       33    13612
#> 19:     1000     2361      105     4047        5    11476
#> 20:      522     1804      288     3753        3    10491
#> 21:     1014     2221      363     3767       14     9940
#> 22:      413     1732       78     3590        4    10606
#> 23:     1399     3286      271     5426        0    15648
#> 24:      967     2969      578     6131       16    16776
#> 25:      634     2731      239     6135       48    17635
#> 26:      741     2551      145     5169       34    15337
#> 27:      586     2116      286     4715        9    13514
#> 28:      388     1945      160     4646        7    13611
#> 29:     1295     2833      126     4559       32    13496
#> 30:     1469     3804      254     7004       15    19814
#> 31:     1643     4073      752     8018       11    21536
#> 32:     1427     4038      588     7959       12    22151
#> 33:      885     3264       71     6603       19    19126
#> 34:     1870     3803       67     5595        6    16706
#> 35:     1546     3468      598     6161       22    16669
#> 36:     1009     3014       63     5708        0    16710
#> 37:     1456     4497      718     9222        3    24997
#> 38:     3243     6465      579     9353        9    26162
#> 39:     1915     5104      565     9568       16    27711
#> 40:     2225     5049      741     8749       41    22380
#> 41:     1414     3849      596     7540       18    20551
#> 42:     1111     3537      291     7226        2    20514
#> 43:     2181     4562      305     6980       43    20136
#> 44:     3016     6483      504    10396        0    29712
#> 45:     2478     6251     1376    12126       45    31453
#> 46:     2012     5877      977    12155       67    33422
#> 47:     1282     4543      357    10207       11    29760
#> 48:     1063     4045      481     8610        8    23987
#> 49:     1365     4112      207     8192        0    24059
#> 50:     1694     4633      230     8283       44    24559
#> 51:     2223     6718     1446    13951       43    37536
#> 52:     2231     6679     1475    14799       23    39918
#> 53:     2052     6774     1124    14936       10    40407
#> 54:     1708     5882      838    12174       44    33943
#> 55:     2844     6490      359    10355       49    29024
#> 56:     2011     5542      824    10709       27    30113
#> 57:     2007     5321      483    10282       16    29595
#> 58:     4531     9931     1277    16398       27    44644
#> 59:     5751    11633     1492    17108       75    45382
#> 60:     4132    10100      445    17336       27    49910
#> 61:     2733     7945      402    15022       11    41982
#> 62:     2583     6679     1312    13422        7    36264
#> 63:     1640     6213      498    13023        3    36269
#> 64:     2131     6533      578    12724        6    37561
#> 65:     5257    11701      882    19631      154    56834
#> 66:     6397    13450     1123    20531      101    62864
#> 67:     3381    11104     1108    22504       18    65897
#> 68:     5135    11514     1519    19250       80    55130
#> 69:     3545     9155     1145    16674       29    47081
#> 70:     4947    10695      507    16923      108    51201
#> 71:     4242     9884      420    16541        1    49244
#> 72:     4414    12345      764    24183       51    73467
#> 73:     3699    12793     1253    26342       34    81641
#> 74:     3443    12550     1196    27363       17    84671
#> 75:     4078    11821      306    22777       14    76605
#> 76:     1610     8670      239    20214        7    66414
#> 77:     1146     7874      779    20777       25    65279
#> 78:     2591     9599      640    19976       46    71635
#> 79:     2855    13262      581    30252       63   103435
#> 80:     5734    16332      872    32379       31   113775
#> 81:     3051    14625      962    34476      236   122567
#> 82:     4507    14328      573    29474       12   110143
#> 83:     3760    11347      327    24312       11    92728
#> 84:     1329     9494      288    25526      158   101742
#> 85:     1517     9725      199    25473        6   100811
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-07-29     725
#>  2: France 2020-07-30    1392
#>  3: France 2020-07-31    1377
#>  4: France 2020-08-01    1346
#>  5: France 2020-08-02       0
#>  6: France 2020-08-03       0
#>  7: France 2020-08-04    3376
#>  8: France 2020-08-05    1039
#>  9: France 2020-08-06    1695
#> 10: France 2020-08-07    1604
#> 11: France 2020-08-08    2288
#> 12: France 2020-08-09       0
#> 13: France 2020-08-10       0
#> 14: France 2020-08-11    4854
#> 15: France 2020-08-12    1397
#> 16: France 2020-08-13    2524
#> 17: France 2020-08-14    2669
#> 18: France 2020-08-15    2846
#> 19: France 2020-08-16    3310
#> 20: France 2020-08-17    3015
#> 21: France 2020-08-18     493
#> 22: France 2020-08-19    2238
#> 23: France 2020-08-20    3776
#> 24: France 2020-08-21    4771
#> 25: France 2020-08-22    4586
#> 26: France 2020-08-23    3602
#> 27: France 2020-08-24    4897
#> 28: France 2020-08-25    1955
#> 29: France 2020-08-26    3304
#> 30: France 2020-08-27    5429
#> 31: France 2020-08-28    6111
#> 32: France 2020-08-29    7379
#> 33: France 2020-08-30    5453
#> 34: France 2020-08-31    5413
#> 35: France 2020-09-01    3082
#> 36: France 2020-09-02    4982
#> 37: France 2020-09-03    7017
#> 38: France 2020-09-04    7157
#> 39: France 2020-09-05    8975
#> 40: France 2020-09-06    8550
#> 41: France 2020-09-07    7071
#> 42: France 2020-09-08    4203
#> 43: France 2020-09-09    6544
#> 44: France 2020-09-10    8577
#> 45: France 2020-09-11    9843
#> 46: France 2020-09-12    9406
#> 47: France 2020-09-13   10561
#> 48: France 2020-09-14    7183
#> 49: France 2020-09-15    6158
#> 50: France 2020-09-16    7852
#> 51: France 2020-09-17    9784
#> 52: France 2020-09-18   10593
#> 53: France 2020-09-19   13215
#> 54: France 2020-09-20   13498
#> 55: France 2020-09-21   11569
#> 56: France 2020-09-22    4298
#> 57: France 2020-09-23   10008
#> 58: France 2020-09-24   13072
#> 59: France 2020-09-25   16096
#> 60: France 2020-09-26   15797
#> 61: France 2020-09-27   14412
#> 62: France 2020-09-28   11123
#> 63: France 2020-09-29    4070
#> 64: France 2020-09-30    8051
#> 65: France 2020-10-01   12845
#> 66: France 2020-10-02   13970
#> 67: France 2020-10-03   12148
#> 68: France 2020-10-04   16972
#> 69: France 2020-10-05   12545
#> 70: France 2020-10-06    5104
#> 71: France 2020-10-07   10489
#> 72: France 2020-10-08   18746
#> 73: France 2020-10-09   18129
#> 74: France 2020-10-10   20339
#> 75: France 2020-10-11   26896
#> 76: France 2020-10-12   16101
#> 77: France 2020-10-13    8505
#> 78: France 2020-10-14   12993
#> 79: France 2020-10-15   22591
#> 80: France 2020-10-16   30621
#> 81: France 2020-10-17   25086
#> 82: France 2020-10-18   32427
#> 83: France 2020-10-19   29837
#> 84: France 2020-10-20   13243
#> 85: France 2020-10-21   20468
#>     region       date confirm
#>