Read in Results from EpiNow2

Reads in results from EpiNow2 and converts them into the RtD3 format. Supports either input via a list object or from a file path/url.

readInEpiNow2(input_list, path, region_var = "region", regions)

Arguments

input_list
A list of results as returned by EpiNow2::regional_summary

path
A character string indicating the path (either file or URL) to the summary results

region_var
A character string that identifies the region name used.

regions
A character string indicating the regions of interest to returns. Defaults to all regions.

Value

A named list in the format required by summaryWidget along with a summary table.

Examples

# Read in each summary folder

base_path <- "https://raw.githubusercontent.com/epiforecasts/covid-rt-estimates/"
rtData <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country")

rtData

#> $summaryData
#>              region New confirmed cases by infection date
#>   1:    Afghanistan                          19 (0 -- 42)
#>   2:        Albania                       118 (75 -- 162)
#>   3:        Algeria                       111 (67 -- 161)
#>   4:        Andorra                         46 (3 -- 116)
#>   5:         Angola                       145 (57 -- 269)
#>  ---                                                     
#> 192:        Vietnam                         26 (0 -- 293)
#> 193: Western Sahara                          1 (0 -- 289)
#> 194:          Yemen                            0 (0 -- 1)
#> 195:         Zambia                          35 (1 -- 95)
#> 196:       Zimbabwe                           6 (0 -- 21)
#>      Expected change in daily cases Effective reproduction no.
#>   1:              Likely decreasing        0.91 (0.55 -- 1.11)
#>   2:                         Unsure        0.97 (0.84 -- 1.06)
#>   3:              Likely decreasing        0.88 (0.74 -- 1.03)
#>   4:              Likely increasing        1.14 (0.85 -- 1.42)
#>   5:              Likely increasing        1.08 (0.93 -- 1.27)
#>  ---                                                          
#> 192:              Likely increasing        2.27 (0.43 -- 5.21)
#> 193:                         Unsure             0.6 (0 -- 1.4)
#> 194:                     Decreasing        0.69 (0.32 -- 0.96)
#> 195:              Likely decreasing        0.84 (0.46 -- 1.14)
#> 196:              Likely decreasing         0.73 (0.3 -- 1.19)
#>              Rate of growth Doubling/halving time (days)
#>   1:  -0.03 (-0.13 -- 0.03)         -27.1 (21.2 -- -5.3)
#>   2:  -0.01 (-0.05 -- 0.02)        -78.1 (40.2 -- -15.1)
#>   3:  -0.03 (-0.08 -- 0.01)           -21 (72.2 -- -9.2)
#>   4:   0.04 (-0.04 -- 0.11)          18.9 (6.2 -- -16.5)
#>   5:   0.02 (-0.02 -- 0.07)            31.8 (9.6 -- -34)
#>  ---                                                    
#> 192:    0.31 (-0.14 -- 0.9)            2.3 (0.8 -- -5.1)
#> 193:   -0.1 (-0.36 -- 0.14)           -6.7 (4.8 -- -1.9)
#> 194: -0.09 (-0.21 -- -0.01)         -7.8 (-69.4 -- -3.3)
#> 195:  -0.04 (-0.14 -- 0.06)         -15.5 (12.4 -- -4.9)
#> 196:  -0.08 (-0.22 -- 0.06)          -8.9 (12.3 -- -3.2)
#> 
#> $rtData
#>             region       date     type median mean   sd lower_20 upper_20
#>     1: Afghanistan 2020-07-23 estimate   0.91 0.90 0.06     0.90     0.92
#>     2: Afghanistan 2020-07-24 estimate   0.91 0.90 0.06     0.90     0.92
#>     3: Afghanistan 2020-07-25 estimate   0.91 0.90 0.06     0.91     0.92
#>     4: Afghanistan 2020-07-26 estimate   0.91 0.90 0.06     0.91     0.93
#>     5: Afghanistan 2020-07-27 estimate   0.91 0.90 0.06     0.91     0.92
#>    ---                                                                   
#> 16189:    Zimbabwe 2020-10-10 forecast   0.73 0.78 0.35     0.64     0.76
#> 16190:    Zimbabwe 2020-10-11 forecast   0.73 0.78 0.35     0.64     0.76
#> 16191:    Zimbabwe 2020-10-12 forecast   0.73 0.78 0.35     0.64     0.76
#> 16192:    Zimbabwe 2020-10-13 forecast   0.73 0.78 0.35     0.64     0.76
#> 16193:    Zimbabwe 2020-10-14 forecast   0.73 0.78 0.35     0.64     0.76
#>        lower_50 upper_50 lower_90 upper_90
#>     1:     0.89     0.94     0.81     0.99
#>     2:     0.89     0.94     0.81     0.99
#>     3:     0.89     0.94     0.82     1.00
#>     4:     0.89     0.94     0.82     1.00
#>     5:     0.89     0.94     0.82     1.00
#>    ---                                    
#> 16189:     0.54     0.86     0.30     1.19
#> 16190:     0.54     0.86     0.30     1.19
#> 16191:     0.54     0.86     0.30     1.19
#> 16192:     0.54     0.86     0.30     1.19
#> 16193:     0.54     0.86     0.30     1.19
#> 
#> $casesInfectionData
#>             region       date     type median  mean      sd lower_20 upper_20
#>     1: Afghanistan 2020-07-23 estimate     90    91      21       86       94
#>     2: Afghanistan 2020-07-24 estimate     88    89      20       83       91
#>     3: Afghanistan 2020-07-25 estimate     86    86      20       81       88
#>     4: Afghanistan 2020-07-26 estimate     84    84      20       79       87
#>     5: Afghanistan 2020-07-27 estimate     82    82      19       77       85
#>    ---                                                                       
#> 16189:    Zimbabwe 2020-10-10 forecast      3  1241   56156        0        1
#> 16190:    Zimbabwe 2020-10-11 forecast      3  2597  121853        0        1
#> 16191:    Zimbabwe 2020-10-12 forecast      2  5476  264560        0        0
#> 16192:    Zimbabwe 2020-10-13 forecast      2 11607  574665        0        0
#> 16193:    Zimbabwe 2020-10-14 forecast      2 24700 1248746        0        0
#>        lower_50 upper_50 lower_90 upper_90
#>     1:       81      104       57      120
#>     2:       77       99       55      117
#>     3:       76       98       52      113
#>     4:       72       94       50      110
#>     5:       71       93       49      109
#>    ---                                    
#> 16189:        0        3        0       26
#> 16190:        0        3        0       27
#> 16191:        0        2        0       28
#> 16192:        0        2        0       28
#> 16193:        0        2        0       29
#> 
#> $casesReportData
#>             region       date     type median mean    sd lower_20 upper_20
#>     1: Afghanistan 2020-07-23 estimate     59  118   166        0       11
#>     2: Afghanistan 2020-07-24 estimate     44   92   135        0        8
#>     3: Afghanistan 2020-07-25 estimate     58  122   180        0       11
#>     4: Afghanistan 2020-07-26 estimate     57  112   158        0        9
#>     5: Afghanistan 2020-07-27 estimate     34   70   103        0        6
#>    ---                                                                    
#> 16189:    Zimbabwe 2020-10-10 forecast      4   53  2128        0        1
#> 16190:    Zimbabwe 2020-10-11 forecast      4   74  2911        0        1
#> 16191:    Zimbabwe 2020-10-12 forecast      3  143  5560        0        1
#> 16192:    Zimbabwe 2020-10-13 forecast      5  191  5396        0        1
#> 16193:    Zimbabwe 2020-10-14 forecast      3  596 27332        0        1
#>        lower_50 upper_50 lower_90 upper_90
#>     1:        0       59        0      307
#>     2:        0       44        0      242
#>     3:        0       58        0      307
#>     4:        0       57        0      287
#>     5:        0       34        0      180
#>    ---                                    
#> 16189:        0        4        0       23
#> 16190:        0        4        0       23
#> 16191:        0        3        0       19
#> 16192:        0        5        0       31
#> 16193:        0        3        0       24
#> 
#> $obsCasesData
#>                region       date confirm
#>     1:    Afghanistan 2020-07-09      59
#>     2:        Albania 2020-07-09      68
#>     3:        Algeria 2020-07-09     469
#>     4:        Andorra 2020-07-09       0
#>     5:         Angola 2020-07-09      33
#>    ---                                  
#> 17761:        Vietnam 2020-10-01       0
#> 17762: Western Sahara 2020-10-01       0
#> 17763:          Yemen 2020-10-01       0
#> 17764:         Zambia 2020-10-01      44
#> 17765:       Zimbabwe 2020-10-01       1
#> 

france <- readInEpiNow2(
  path = paste0(base_path, "master/national/cases/summary"),
  region_var = "country",
  regions = "France")

france

#> $summaryData
#>    region New confirmed cases by infection date Expected change in daily cases
#> 1: France                 29741 (3431 -- 63134)              Likely increasing
#>    Effective reproduction no.       Rate of growth Doubling/halving time (days)
#> 1:        1.16 (0.86 -- 1.35) 0.04 (-0.04 -- 0.09)          15.6 (7.5 -- -18.1)
#> 
#> $rtData
#>     region       date                           type median mean   sd lower_20
#>  1: France 2020-07-23                       estimate   1.20 1.20 0.06     1.18
#>  2: France 2020-07-24                       estimate   1.20 1.21 0.06     1.19
#>  3: France 2020-07-25                       estimate   1.20 1.21 0.06     1.19
#>  4: France 2020-07-26                       estimate   1.20 1.21 0.06     1.19
#>  5: France 2020-07-27                       estimate   1.20 1.21 0.06     1.19
#>  6: France 2020-07-28                       estimate   1.21 1.21 0.06     1.19
#>  7: France 2020-07-29                       estimate   1.21 1.21 0.06     1.19
#>  8: France 2020-07-30                       estimate   1.21 1.21 0.06     1.19
#>  9: France 2020-07-31                       estimate   1.21 1.22 0.07     1.19
#> 10: France 2020-08-01                       estimate   1.21 1.22 0.07     1.18
#> 11: France 2020-08-02                       estimate   1.21 1.22 0.07     1.18
#> 12: France 2020-08-03                       estimate   1.21 1.22 0.07     1.19
#> 13: France 2020-08-04                       estimate   1.21 1.22 0.07     1.19
#> 14: France 2020-08-05                       estimate   1.21 1.22 0.07     1.19
#> 15: France 2020-08-06                       estimate   1.21 1.22 0.07     1.19
#> 16: France 2020-08-07                       estimate   1.21 1.22 0.07     1.19
#> 17: France 2020-08-08                       estimate   1.21 1.22 0.07     1.19
#> 18: France 2020-08-09                       estimate   1.21 1.22 0.07     1.19
#> 19: France 2020-08-10                       estimate   1.21 1.22 0.07     1.19
#> 20: France 2020-08-11                       estimate   1.21 1.22 0.07     1.19
#> 21: France 2020-08-12                       estimate   1.20 1.22 0.07     1.18
#> 22: France 2020-08-13                       estimate   1.20 1.22 0.07     1.18
#> 23: France 2020-08-14                       estimate   1.20 1.22 0.07     1.18
#> 24: France 2020-08-15                       estimate   1.20 1.21 0.07     1.19
#> 25: France 2020-08-16                       estimate   1.20 1.21 0.07     1.18
#> 26: France 2020-08-17                       estimate   1.20 1.21 0.07     1.19
#> 27: France 2020-08-18                       estimate   1.20 1.21 0.07     1.18
#> 28: France 2020-08-19                       estimate   1.20 1.20 0.07     1.18
#> 29: France 2020-08-20                       estimate   1.19 1.20 0.07     1.18
#> 30: France 2020-08-21                       estimate   1.19 1.20 0.07     1.18
#> 31: France 2020-08-22                       estimate   1.19 1.19 0.06     1.18
#> 32: France 2020-08-23                       estimate   1.19 1.19 0.06     1.18
#> 33: France 2020-08-24                       estimate   1.19 1.19 0.06     1.18
#> 34: France 2020-08-25                       estimate   1.19 1.19 0.06     1.18
#> 35: France 2020-08-26                       estimate   1.19 1.18 0.06     1.18
#> 36: France 2020-08-27                       estimate   1.18 1.18 0.06     1.18
#> 37: France 2020-08-28                       estimate   1.18 1.18 0.06     1.18
#> 38: France 2020-08-29                       estimate   1.18 1.17 0.06     1.18
#> 39: France 2020-08-30                       estimate   1.18 1.17 0.06     1.18
#> 40: France 2020-08-31                       estimate   1.18 1.17 0.06     1.17
#> 41: France 2020-09-01                       estimate   1.17 1.17 0.06     1.17
#> 42: France 2020-09-02                       estimate   1.17 1.16 0.06     1.17
#> 43: France 2020-09-03                       estimate   1.17 1.16 0.07     1.17
#> 44: France 2020-09-04                       estimate   1.17 1.16 0.07     1.17
#> 45: France 2020-09-05                       estimate   1.17 1.16 0.07     1.18
#> 46: France 2020-09-06                       estimate   1.17 1.16 0.07     1.18
#> 47: France 2020-09-07                       estimate   1.17 1.15 0.07     1.17
#> 48: France 2020-09-08                       estimate   1.17 1.15 0.07     1.17
#> 49: France 2020-09-09                       estimate   1.16 1.15 0.07     1.17
#> 50: France 2020-09-10                       estimate   1.16 1.15 0.08     1.17
#> 51: France 2020-09-11                       estimate   1.16 1.15 0.08     1.17
#> 52: France 2020-09-12                       estimate   1.16 1.15 0.08     1.17
#> 53: France 2020-09-13                       estimate   1.16 1.15 0.08     1.18
#> 54: France 2020-09-14                       estimate   1.16 1.15 0.09     1.18
#> 55: France 2020-09-15                       estimate   1.16 1.14 0.09     1.17
#> 56: France 2020-09-16                       estimate   1.16 1.14 0.09     1.17
#> 57: France 2020-09-17                       estimate   1.16 1.14 0.10     1.17
#> 58: France 2020-09-18                       estimate   1.16 1.14 0.10     1.17
#> 59: France 2020-09-19                       estimate   1.16 1.14 0.10     1.17
#> 60: France 2020-09-20                       estimate   1.16 1.14 0.11     1.17
#> 61: France 2020-09-21 estimate based on partial data   1.16 1.14 0.11     1.18
#> 62: France 2020-09-22 estimate based on partial data   1.16 1.14 0.12     1.18
#> 63: France 2020-09-23 estimate based on partial data   1.16 1.14 0.12     1.17
#> 64: France 2020-09-24 estimate based on partial data   1.16 1.14 0.13     1.17
#> 65: France 2020-09-25 estimate based on partial data   1.16 1.14 0.13     1.17
#> 66: France 2020-09-26 estimate based on partial data   1.16 1.14 0.14     1.17
#> 67: France 2020-09-27 estimate based on partial data   1.16 1.14 0.14     1.17
#> 68: France 2020-09-28 estimate based on partial data   1.16 1.14 0.15     1.17
#> 69: France 2020-09-29 estimate based on partial data   1.16 1.14 0.15     1.17
#> 70: France 2020-09-30 estimate based on partial data   1.16 1.14 0.16     1.17
#> 71: France 2020-10-01 estimate based on partial data   1.16 1.14 0.16     1.16
#> 72: France 2020-10-02                       forecast   1.16 1.14 0.16     1.16
#> 73: France 2020-10-03                       forecast   1.16 1.14 0.16     1.16
#> 74: France 2020-10-04                       forecast   1.16 1.14 0.16     1.16
#> 75: France 2020-10-05                       forecast   1.16 1.14 0.16     1.16
#> 76: France 2020-10-06                       forecast   1.16 1.14 0.16     1.16
#> 77: France 2020-10-07                       forecast   1.16 1.14 0.16     1.16
#> 78: France 2020-10-08                       forecast   1.16 1.14 0.16     1.16
#> 79: France 2020-10-09                       forecast   1.16 1.14 0.16     1.16
#> 80: France 2020-10-10                       forecast   1.16 1.14 0.16     1.16
#> 81: France 2020-10-11                       forecast   1.16 1.14 0.16     1.16
#> 82: France 2020-10-12                       forecast   1.16 1.14 0.16     1.16
#> 83: France 2020-10-13                       forecast   1.16 1.14 0.16     1.16
#> 84: France 2020-10-14                       forecast   1.16 1.14 0.16     1.16
#> 85: France 2020-10-15                       forecast   1.16 1.14 0.16     1.16
#>     region       date                           type median mean   sd lower_20
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1.20     1.18     1.23     1.11     1.28
#>  2:     1.21     1.18     1.23     1.12     1.29
#>  3:     1.21     1.17     1.22     1.12     1.30
#>  4:     1.21     1.17     1.22     1.12     1.30
#>  5:     1.21     1.18     1.23     1.12     1.30
#>  6:     1.21     1.18     1.23     1.12     1.31
#>  7:     1.21     1.18     1.23     1.12     1.31
#>  8:     1.21     1.18     1.23     1.12     1.32
#>  9:     1.21     1.18     1.23     1.12     1.32
#> 10:     1.20     1.17     1.23     1.12     1.33
#> 11:     1.20     1.18     1.23     1.12     1.33
#> 12:     1.20     1.17     1.22     1.12     1.34
#> 13:     1.20     1.17     1.23     1.13     1.34
#> 14:     1.20     1.17     1.23     1.13     1.34
#> 15:     1.21     1.17     1.23     1.13     1.35
#> 16:     1.21     1.17     1.22     1.13     1.34
#> 17:     1.20     1.17     1.23     1.13     1.35
#> 18:     1.20     1.18     1.23     1.13     1.35
#> 19:     1.20     1.17     1.23     1.13     1.35
#> 20:     1.21     1.17     1.23     1.12     1.34
#> 21:     1.20     1.17     1.22     1.12     1.34
#> 22:     1.20     1.18     1.23     1.11     1.33
#> 23:     1.20     1.17     1.23     1.10     1.32
#> 24:     1.20     1.17     1.23     1.10     1.32
#> 25:     1.20     1.17     1.23     1.11     1.32
#> 26:     1.21     1.17     1.23     1.10     1.32
#> 27:     1.20     1.17     1.22     1.10     1.31
#> 28:     1.20     1.16     1.22     1.09     1.31
#> 29:     1.20     1.16     1.22     1.09     1.30
#> 30:     1.20     1.16     1.22     1.09     1.30
#> 31:     1.20     1.16     1.22     1.08     1.29
#> 32:     1.20     1.16     1.22     1.08     1.29
#> 33:     1.20     1.16     1.21     1.08     1.28
#> 34:     1.20     1.15     1.21     1.08     1.28
#> 35:     1.20     1.16     1.22     1.08     1.27
#> 36:     1.20     1.15     1.21     1.08     1.28
#> 37:     1.20     1.15     1.21     1.08     1.27
#> 38:     1.20     1.15     1.21     1.06     1.26
#> 39:     1.20     1.15     1.21     1.06     1.26
#> 40:     1.19     1.15     1.21     1.06     1.26
#> 41:     1.19     1.15     1.22     1.06     1.26
#> 42:     1.19     1.15     1.21     1.06     1.27
#> 43:     1.19     1.15     1.22     1.06     1.27
#> 44:     1.19     1.15     1.22     1.06     1.27
#> 45:     1.20     1.14     1.22     1.04     1.26
#> 46:     1.20     1.15     1.22     1.04     1.26
#> 47:     1.20     1.15     1.22     1.03     1.26
#> 48:     1.19     1.14     1.22     1.03     1.26
#> 49:     1.19     1.13     1.21     1.02     1.26
#> 50:     1.20     1.13     1.21     1.01     1.26
#> 51:     1.20     1.13     1.22     1.01     1.26
#> 52:     1.20     1.13     1.22     1.00     1.26
#> 53:     1.21     1.13     1.22     0.99     1.26
#> 54:     1.21     1.13     1.22     1.00     1.28
#> 55:     1.21     1.13     1.22     0.98     1.27
#> 56:     1.20     1.13     1.23     0.98     1.28
#> 57:     1.20     1.13     1.23     0.97     1.28
#> 58:     1.20     1.13     1.23     0.96     1.28
#> 59:     1.20     1.13     1.23     0.94     1.27
#> 60:     1.21     1.13     1.23     0.94     1.28
#> 61:     1.21     1.13     1.23     0.94     1.29
#> 62:     1.21     1.12     1.23     0.93     1.29
#> 63:     1.21     1.12     1.23     0.91     1.29
#> 64:     1.20     1.12     1.23     0.92     1.31
#> 65:     1.20     1.12     1.23     0.90     1.31
#> 66:     1.21     1.12     1.23     0.92     1.34
#> 67:     1.21     1.12     1.24     0.90     1.35
#> 68:     1.20     1.11     1.23     0.88     1.33
#> 69:     1.20     1.12     1.24     0.87     1.33
#> 70:     1.21     1.11     1.23     0.86     1.34
#> 71:     1.20     1.11     1.23     0.86     1.35
#> 72:     1.20     1.11     1.23     0.86     1.35
#> 73:     1.20     1.11     1.23     0.86     1.35
#> 74:     1.20     1.11     1.23     0.86     1.35
#> 75:     1.20     1.11     1.23     0.86     1.35
#> 76:     1.20     1.11     1.23     0.86     1.35
#> 77:     1.20     1.11     1.23     0.86     1.35
#> 78:     1.20     1.11     1.23     0.86     1.35
#> 79:     1.20     1.11     1.23     0.86     1.35
#> 80:     1.20     1.11     1.23     0.86     1.35
#> 81:     1.20     1.11     1.23     0.86     1.35
#> 82:     1.20     1.11     1.23     0.86     1.35
#> 83:     1.20     1.11     1.23     0.86     1.35
#> 84:     1.20     1.11     1.23     0.86     1.35
#> 85:     1.20     1.11     1.23     0.86     1.35
#>     upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesInfectionData
#>     region       date                           type median   mean      sd
#>  1: France 2020-07-23                       estimate   1209   1235     280
#>  2: France 2020-07-24                       estimate   1274   1302     300
#>  3: France 2020-07-25                       estimate   1341   1374     320
#>  4: France 2020-07-26                       estimate   1413   1450     341
#>  5: France 2020-07-27                       estimate   1491   1531     363
#>  6: France 2020-07-28                       estimate   1573   1617     386
#>  7: France 2020-07-29                       estimate   1658   1708     410
#>  8: France 2020-07-30                       estimate   1747   1805     436
#>  9: France 2020-07-31                       estimate   1845   1907     463
#> 10: France 2020-08-01                       estimate   1944   2016     492
#> 11: France 2020-08-02                       estimate   2049   2132     523
#> 12: France 2020-08-03                       estimate   2165   2255     556
#> 13: France 2020-08-04                       estimate   2288   2385     592
#> 14: France 2020-08-05                       estimate   2419   2524     630
#> 15: France 2020-08-06                       estimate   2552   2670     671
#> 16: France 2020-08-07                       estimate   2694   2824     716
#> 17: France 2020-08-08                       estimate   2845   2987     763
#> 18: France 2020-08-09                       estimate   3007   3159     812
#> 19: France 2020-08-10                       estimate   3177   3341     865
#> 20: France 2020-08-11                       estimate   3360   3531     920
#> 21: France 2020-08-12                       estimate   3549   3731     980
#> 22: France 2020-08-13                       estimate   3744   3941    1044
#> 23: France 2020-08-14                       estimate   3947   4160    1112
#> 24: France 2020-08-15                       estimate   4157   4389    1185
#> 25: France 2020-08-16                       estimate   4376   4628    1264
#> 26: France 2020-08-17                       estimate   4605   4876    1346
#> 27: France 2020-08-18                       estimate   4844   5134    1433
#> 28: France 2020-08-19                       estimate   5093   5402    1522
#> 29: France 2020-08-20                       estimate   5348   5678    1613
#> 30: France 2020-08-21                       estimate   5622   5964    1704
#> 31: France 2020-08-22                       estimate   5907   6259    1797
#> 32: France 2020-08-23                       estimate   6195   6562    1889
#> 33: France 2020-08-24                       estimate   6487   6874    1980
#> 34: France 2020-08-25                       estimate   6792   7195    2072
#> 35: France 2020-08-26                       estimate   7094   7526    2164
#> 36: France 2020-08-27                       estimate   7422   7865    2257
#> 37: France 2020-08-28                       estimate   7751   8214    2350
#> 38: France 2020-08-29                       estimate   8101   8573    2445
#> 39: France 2020-08-30                       estimate   8464   8942    2543
#> 40: France 2020-08-31                       estimate   8835   9322    2643
#> 41: France 2020-09-01                       estimate   9212   9712    2746
#> 42: France 2020-09-02                       estimate   9611  10114    2855
#> 43: France 2020-09-03                       estimate  10023  10529    2970
#> 44: France 2020-09-04                       estimate  10430  10956    3093
#> 45: France 2020-09-05                       estimate  10820  11397    3226
#> 46: France 2020-09-06                       estimate  11261  11852    3371
#> 47: France 2020-09-07                       estimate  11714  12323    3532
#> 48: France 2020-09-08                       estimate  12165  12811    3712
#> 49: France 2020-09-09                       estimate  12646  13316    3913
#> 50: France 2020-09-10                       estimate  13115  13841    4140
#> 51: France 2020-09-11                       estimate  13625  14386    4394
#> 52: France 2020-09-12                       estimate  14190  14954    4678
#> 53: France 2020-09-13                       estimate  14721  15546    4996
#> 54: France 2020-09-14                       estimate  15253  16164    5350
#> 55: France 2020-09-15                       estimate  15851  16811    5743
#> 56: France 2020-09-16                       estimate  16474  17489    6179
#> 57: France 2020-09-17                       estimate  17067  18202    6663
#> 58: France 2020-09-18                       estimate  17751  18952    7202
#> 59: France 2020-09-19                       estimate  18420  19744    7805
#> 60: France 2020-09-20                       estimate  19096  20583    8481
#> 61: France 2020-09-21 estimate based on partial data  19822  21474    9244
#> 62: France 2020-09-22 estimate based on partial data  20598  22424   10112
#> 63: France 2020-09-23 estimate based on partial data  21388  23439   11105
#> 64: France 2020-09-24 estimate based on partial data  22354  24528   12252
#> 65: France 2020-09-25 estimate based on partial data  23202  25701   13589
#> 66: France 2020-09-26 estimate based on partial data  24173  26969   15165
#> 67: France 2020-09-27 estimate based on partial data  25248  28345   17054
#> 68: France 2020-09-28 estimate based on partial data  26298  29846   19360
#> 69: France 2020-09-29 estimate based on partial data  27420  31491   22240
#> 70: France 2020-09-30 estimate based on partial data  28575  33306   25929
#> 71: France 2020-10-01 estimate based on partial data  29741  35325   30805
#> 72: France 2020-10-02                       forecast  30997  37538   37063
#> 73: France 2020-10-03                       forecast  32433  40003   45429
#> 74: France 2020-10-04                       forecast  33881  42767   56798
#> 75: France 2020-10-05                       forecast  35424  45891   72496
#> 76: France 2020-10-06                       forecast  36855  49455   94484
#> 77: France 2020-10-07                       forecast  38403  53564  125665
#> 78: France 2020-10-08                       forecast  40077  58357  170326
#> 79: France 2020-10-09                       forecast  41725  64024  234794
#> 80: France 2020-10-10                       forecast  43652  70822  328395
#> 81: France 2020-10-11                       forecast  45688  79101  464871
#> 82: France 2020-10-12                       forecast  47618  89350  664473
#> 83: France 2020-10-13                       forecast  49932 102250  957039
#> 84: France 2020-10-14                       forecast  52148 118755 1386549
#> 85: France 2020-10-15                       forecast  54572 140219 2017833
#>     region       date                           type median   mean      sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:     1130     1223     1061     1330      745     1607
#>  2:     1184     1284     1110     1401      808     1725
#>  3:     1238     1348     1171     1483      852     1820
#>  4:     1293     1411     1238     1576      898     1925
#>  5:     1356     1486     1297     1659      948     2039
#>  6:     1424     1565     1364     1753     1009     2160
#>  7:     1589     1738     1437     1853     1077     2295
#>  8:     1625     1784     1458     1898     1156     2451
#>  9:     1713     1883     1540     2005     1213     2581
#> 10:     1769     1948     1641     2137     1289     2733
#> 11:     1852     2039     1747     2273     1338     2866
#> 12:     1968     2168     1825     2382     1427     3037
#> 13:     2100     2316     1937     2523     1489     3178
#> 14:     2147     2378     2037     2661     1618     3415
#> 15:     2284     2527     2134     2804     1717     3641
#> 16:     2414     2671     2294     3015     1786     3836
#> 17:     2562     2838     2332     3104     1915     4078
#> 18:     2765     3060     2431     3249     2009     4310
#> 19:     2820     3139     2585     3449     2128     4585
#> 20:     2949     3292     2708     3629     2206     4817
#> 21:     3070     3432     2895     3883     2345     5119
#> 22:     3260     3652     3058     4115     2441     5387
#> 23:     3451     3868     3159     4295     2561     5704
#> 24:     3744     4181     3305     4518     2727     6064
#> 25:     3998     4454     3514     4799     2885     6423
#> 26:     4184     4676     3631     4992     2908     6665
#> 27:     4478     5000     3744     5186     3111     7085
#> 28:     4536     5094     3919     5453     3252     7441
#> 29:     4785     5365     4126     5732     3392     7841
#> 30:     5079     5691     4298     5994     3531     8244
#> 31:     5125     5773     4631     6435     3698     8683
#> 32:     5303     5991     4691     6593     3877     9127
#> 33:     5576     6292     4898     6908     4158     9700
#> 34:     5897     6644     5180     7299     4224    10032
#> 35:     6270     7071     5506     7734     4555    10609
#> 36:     6499     7345     5822     8155     4736    11059
#> 37:     6652     7532     6059     8497     5017    11654
#> 38:     6922     7839     6341     8897     5105    11984
#> 39:     7322     8294     6561     9254     5273    12468
#> 40:     7861     8884     6968     9771     5472    12962
#> 41:     8186     9259     7183    10113     5662    13461
#> 42:     8002     9132     7520    10579     5979    14115
#> 43:     8331     9506     7856    11052     6186    14683
#> 44:     8958    10188     8174    11508     6406    15347
#> 45:     9406    10691     8529    12034     6669    16000
#> 46:     9600    10923     8623    12302     6806    16546
#> 47:     9759    11133     9070    12923     7012    17233
#> 48:    10342    11762     9632    13721     7002    17797
#> 49:    10583    12095    10019    14307     7366    18690
#> 50:    11137    12779    10513    15052     7672    19610
#> 51:    11774    13542    10195    14999     8015    20717
#> 52:    12097    13984    10610    15688     8166    21734
#> 53:    11871    13863    10942    16333     8473    22903
#> 54:    13063    15198    11252    17027     8644    24105
#> 55:    13049    15313    11115    17342     8458    24998
#> 56:    13549    15949    12210    18938     8820    26630
#> 57:    13899    16548    11935    19192     8737    27950
#> 58:    14031    16927    12280    20087     8629    29254
#> 59:    14252    17422    12608    21085     8192    30510
#> 60:    14239    17647    13184    22404     8394    32650
#> 61:    14793    18481    12445    22437     8507    34814
#> 62:    15559    19610    13407    24213     7945    36623
#> 63:    16548    20925    13509    25270     6864    38361
#> 64:    15411    20253    12664    25417     6599    40833
#> 65:    17648    22900    13134    26921     6383    43412
#> 66:    14987    20739    13767    28769     6069    46246
#> 67:    15359    21569    13654    30037     5625    49280
#> 68:    16235    22842    14126    31898     5431    52651
#> 69:    16129    23338    14092    33354     4814    55942
#> 70:    14953    22841    14514    35386     4303    59757
#> 71:    14942    23388    13364    35950     3431    63134
#> 72:    13606    22748    13669    37954     3721    67606
#> 73:    13960    23726    13400    39575     1703    70680
#> 74:    14243    24619    13467    41592     3012    77000
#> 75:    14183    25274    14018    44161     2507    81734
#> 76:    14091    25984    13408    45695     1127    86345
#> 77:    13971    26573    13938    48390      831    91596
#> 78:    14132    27446    12847    49725      464    97217
#> 79:    14410    28598     9056    48202      170   103424
#> 80:    13471    28562     9090    50765      273   110739
#> 81:    12905    28908     3853    47976      167   118561
#> 82:    10302    27218     3701    50131       75   126692
#> 83:    10147    27845     2262    51217       57   136301
#> 84:    10215    28916     1959    53326       43   146473
#> 85:     1803    21227     1873    55751       33   156380
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $casesReportData
#>     region       date                           type median  mean     sd
#>  1: France 2020-07-23                       estimate    470   974   1390
#>  2: France 2020-07-24                       estimate    481  1055   1714
#>  3: France 2020-07-25                       estimate    527  1121   1618
#>  4: France 2020-07-26                       estimate    290   641   1047
#>  5: France 2020-07-27                       estimate    272   606    931
#>  6: France 2020-07-28                       estimate    670  1451   2085
#>  7: France 2020-07-29                       estimate    348   769   1169
#>  8: France 2020-07-30                       estimate    652  1393   2018
#>  9: France 2020-07-31                       estimate    658  1473   2252
#> 10: France 2020-08-01                       estimate    766  1634   2411
#> 11: France 2020-08-02                       estimate    406   916   1391
#> 12: France 2020-08-03                       estimate    391   870   1339
#> 13: France 2020-08-04                       estimate    998  2101   3111
#> 14: France 2020-08-05                       estimate    522  1163   1847
#> 15: France 2020-08-06                       estimate   1017  2064   2967
#> 16: France 2020-08-07                       estimate   1100  2156   3061
#> 17: France 2020-08-08                       estimate   1152  2381   3719
#> 18: France 2020-08-09                       estimate    618  1359   2164
#> 19: France 2020-08-10                       estimate    568  1253   1955
#> 20: France 2020-08-11                       estimate   1500  3213   4839
#> 21: France 2020-08-12                       estimate    770  1719   2580
#> 22: France 2020-08-13                       estimate   1400  3025   4583
#> 23: France 2020-08-14                       estimate   1516  3179   4673
#> 24: France 2020-08-15                       estimate   1684  3601   5779
#> 25: France 2020-08-16                       estimate   1002  2035   2902
#> 26: France 2020-08-17                       estimate    856  1916   2987
#> 27: France 2020-08-18                       estimate   1996  4545   7137
#> 28: France 2020-08-19                       estimate   1086  2451   3739
#> 29: France 2020-08-20                       estimate   2056  4507   6871
#> 30: France 2020-08-21                       estimate   2036  4582   6959
#> 31: France 2020-08-22                       estimate   2444  5383   8317
#> 32: France 2020-08-23                       estimate   1321  2950   4657
#> 33: France 2020-08-24                       estimate   1213  2607   3807
#> 34: France 2020-08-25                       estimate   3016  6676  10943
#> 35: France 2020-08-26                       estimate   1702  3678   5490
#> 36: France 2020-08-27                       estimate   2950  6266   9178
#> 37: France 2020-08-28                       estimate   3112  6922  11024
#> 38: France 2020-08-29                       estimate   3390  7178  10453
#> 39: France 2020-08-30                       estimate   1872  4073   5846
#> 40: France 2020-08-31                       estimate   1709  3749   5623
#> 41: France 2020-09-01                       estimate   4130  9270  15646
#> 42: France 2020-09-02                       estimate   2392  5129   7674
#> 43: France 2020-09-03                       estimate   4318  8827  12991
#> 44: France 2020-09-04                       estimate   4304  9057  13542
#> 45: France 2020-09-05                       estimate   4749 10242  15167
#> 46: France 2020-09-06                       estimate   2773  5700   8264
#> 47: France 2020-09-07                       estimate   2328  5089   7545
#> 48: France 2020-09-08                       estimate   5752 12749  19728
#> 49: France 2020-09-09                       estimate   2996  7033  11395
#> 50: France 2020-09-10                       estimate   5777 12419  18797
#> 51: France 2020-09-11                       estimate   5954 12343  18665
#> 52: France 2020-09-12                       estimate   6158 13570  19761
#> 53: France 2020-09-13                       estimate   3419  7464  11418
#> 54: France 2020-09-14                       estimate   3066  6999  10836
#> 55: France 2020-09-15                       estimate   7380 17198  28489
#> 56: France 2020-09-16                       estimate   4022  8922  13844
#> 57: France 2020-09-17                       estimate   7252 15522  22443
#> 58: France 2020-09-18                       estimate   7047 15711  24969
#> 59: France 2020-09-19                       estimate   8260 17564  26656
#> 60: France 2020-09-20                       estimate   4498  9650  15877
#> 61: France 2020-09-21 estimate based on partial data   4125  8810  13025
#> 62: France 2020-09-22 estimate based on partial data   9910 23203  38679
#> 63: France 2020-09-23 estimate based on partial data   5274 12113  19462
#> 64: France 2020-09-24 estimate based on partial data   9238 20854  32111
#> 65: France 2020-09-25 estimate based on partial data   9417 21567  34480
#> 66: France 2020-09-26 estimate based on partial data  10951 24775  39824
#> 67: France 2020-09-27 estimate based on partial data   5636 13327  21791
#> 68: France 2020-09-28 estimate based on partial data   5184 11714  19720
#> 69: France 2020-09-29 estimate based on partial data  12608 29704  49605
#> 70: France 2020-09-30 estimate based on partial data   6442 15583  28040
#> 71: France 2020-10-01 estimate based on partial data  12232 28621  46652
#> 72: France 2020-10-02                       forecast  12393 29902  46536
#> 73: France 2020-10-03                       forecast  13098 32161  57546
#> 74: France 2020-10-04                       forecast   6728 18376  38523
#> 75: France 2020-10-05                       forecast   7027 17017  28440
#> 76: France 2020-10-06                       forecast  16886 42937  93395
#> 77: France 2020-10-07                       forecast   8038 21135  39063
#> 78: France 2020-10-08                       forecast  15153 40381  85904
#> 79: France 2020-10-09                       forecast  15653 41837  89761
#> 80: France 2020-10-10                       forecast  17709 50113 129857
#> 81: France 2020-10-11                       forecast   8910 26934  70030
#> 82: France 2020-10-12                       forecast   8394 27060  96419
#> 83: France 2020-10-13                       forecast  20569 78322 611688
#> 84: France 2020-10-14                       forecast  11407 36115 115287
#> 85: France 2020-10-15                       forecast  19648 69583 344669
#>     region       date                           type median  mean     sd
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#>  1:        0       76        0      470        0     2587
#>  2:        0       83        0      481        0     2685
#>  3:        0       91        0      527        0     2896
#>  4:        0       47        0      290        0     1611
#>  5:        0       48        0      272        0     1579
#>  6:        0      118        0      670        0     3702
#>  7:        0       60        0      348        0     2010
#>  8:        0      109        0      653        0     3626
#>  9:        0      122        0      659        0     3876
#> 10:        0      149        0      768        0     4176
#> 11:        0       70        0      407        0     2408
#> 12:        0       71        0      391        0     2295
#> 13:        0      175        0      999        0     5405
#> 14:        0       91        0      522        0     3028
#> 15:        0      177        0     1018        0     5213
#> 16:        0      194        0     1102        0     5578
#> 17:        0      205        0     1153        0     6092
#> 18:        0      113        0      618        0     3557
#> 19:        0       92        0      569        0     3265
#> 20:        0      259        0     1500        0     8358
#> 21:        0      134        0      770        0     4591
#> 22:        0      233        0     1401        0     7537
#> 23:        0      249        0     1516        0     8317
#> 24:        0      284        0     1684        0     8957
#> 25:        0      166        0     1002        0     5500
#> 26:        0      150        0      857        0     4851
#> 27:        0      361        0     1997        0    11926
#> 28:        0      199        0     1086        0     6487
#> 29:        0      381        0     2057        0    11444
#> 30:        0      370        0     2036        0    12050
#> 31:        0      446        0     2446        0    13870
#> 32:        0      234        0     1324        0     7747
#> 33:        0      219        0     1213        0     6756
#> 34:        0      512        0     3016        0    16862
#> 35:        0      263        0     1702        0     9367
#> 36:        0      477        0     2954        0    16750
#> 37:        0      532        0     3118        0    17797
#> 38:        0      598        0     3391        0    18438
#> 39:        0      301        0     1873        0    11028
#> 40:        0      314        0     1710        0     9831
#> 41:        0      776        0     4135        0    23173
#> 42:        0      411        0     2393        0    13522
#> 43:        0      741        0     4318        0    22879
#> 44:        0      743        0     4305        0    23206
#> 45:        0      753        0     4749        0    26267
#> 46:        0      469        0     2774        0    14780
#> 47:        0      406        0     2329        0    13840
#> 48:        0      937        0     5754        0    33924
#> 49:        0      561        0     2998        0    18217
#> 50:        0      937        0     5779        0    32465
#> 51:        0     1020        0     5957        0    31266
#> 52:        0     1195        0     6163        0    34527
#> 53:        0      624        0     3421        0    19137
#> 54:        0      558        0     3069        0    18286
#> 55:        0     1326        0     7381        0    43590
#> 56:        0      659        0     4024        0    22969
#> 57:        0     1201        0     7253        0    40446
#> 58:        0     1230        0     7059        0    39937
#> 59:        0     1447        0     8274        0    45495
#> 60:        0      719        0     4499        0    24882
#> 61:        0      690        0     4126        0    23093
#> 62:        0     1742        0     9913        0    59099
#> 63:        0      934        0     5279        0    31536
#> 64:        0     1587        0     9254        0    52529
#> 65:        0     1594        0     9418        0    54881
#> 66:        0     1884        0    10954        0    63029
#> 67:        0      997        0     5650        0    34435
#> 68:        0      881        0     5188        0    29832
#> 69:        0     2190        0    12617        0    76828
#> 70:        0     1151        0     6442        0    37691
#> 71:        0     2156        0    12239        0    75323
#> 72:        0     1992        0    12394        0    80725
#> 73:        0     2270        0    13100        0    83438
#> 74:        0     1091        0     6729        0    46331
#> 75:        0     1141        0     7037        0    44575
#> 76:        0     2612        0    16894        0   108131
#> 77:        0     1346        0     8042        0    53411
#> 78:        0     2544        0    15158        0    99026
#> 79:        0     2650        0    15660        0   105799
#> 80:        0     2738        0    17716        0   118830
#> 81:        0     1454        0     8919        0    63241
#> 82:        0     1286        0     8401        0    61810
#> 83:        0     3148        0    20607        0   163331
#> 84:        0     1823        0    11414        0    79720
#> 85:        0     2842        0    19651        0   154204
#>     lower_20 upper_20 lower_50 upper_50 lower_90 upper_90
#> 
#> $obsCasesData
#>     region       date confirm
#>  1: France 2020-07-09     663
#>  2: France 2020-07-10     621
#>  3: France 2020-07-11     658
#>  4: France 2020-07-12       0
#>  5: France 2020-07-13       0
#>  6: France 2020-07-14    1625
#>  7: France 2020-07-15       0
#>  8: France 2020-07-16     927
#>  9: France 2020-07-17     534
#> 10: France 2020-07-18     836
#> 11: France 2020-07-19       0
#> 12: France 2020-07-20       0
#> 13: France 2020-07-21    2080
#> 14: France 2020-07-22     584
#> 15: France 2020-07-23     998
#> 16: France 2020-07-24    1062
#> 17: France 2020-07-25    1130
#> 18: France 2020-07-26       0
#> 19: France 2020-07-27       0
#> 20: France 2020-07-28    2551
#> 21: France 2020-07-29     725
#> 22: France 2020-07-30    1392
#> 23: France 2020-07-31    1377
#> 24: France 2020-08-01    1346
#> 25: France 2020-08-02       0
#> 26: France 2020-08-03       0
#> 27: France 2020-08-04    3376
#> 28: France 2020-08-05    1039
#> 29: France 2020-08-06    1695
#> 30: France 2020-08-07    1604
#> 31: France 2020-08-08    2288
#> 32: France 2020-08-09       0
#> 33: France 2020-08-10       0
#> 34: France 2020-08-11    4854
#> 35: France 2020-08-12    1397
#> 36: France 2020-08-13    2524
#> 37: France 2020-08-14    2669
#> 38: France 2020-08-15    2846
#> 39: France 2020-08-16    3310
#> 40: France 2020-08-17    3015
#> 41: France 2020-08-18     493
#> 42: France 2020-08-19    2238
#> 43: France 2020-08-20    3776
#> 44: France 2020-08-21    4771
#> 45: France 2020-08-22    4586
#> 46: France 2020-08-23    3602
#> 47: France 2020-08-24    4897
#> 48: France 2020-08-25    1955
#> 49: France 2020-08-26    3304
#> 50: France 2020-08-27    5429
#> 51: France 2020-08-28    6111
#> 52: France 2020-08-29    7379
#> 53: France 2020-08-30    5453
#> 54: France 2020-08-31    5413
#> 55: France 2020-09-01    3082
#> 56: France 2020-09-02    4982
#> 57: France 2020-09-03    7017
#> 58: France 2020-09-04    7157
#> 59: France 2020-09-05    8975
#> 60: France 2020-09-06    8550
#> 61: France 2020-09-07    7071
#> 62: France 2020-09-08    4203
#> 63: France 2020-09-09    6544
#> 64: France 2020-09-10    8577
#> 65: France 2020-09-11    9843
#> 66: France 2020-09-12    9406
#> 67: France 2020-09-13   10561
#> 68: France 2020-09-14    7183
#> 69: France 2020-09-15    6158
#> 70: France 2020-09-16    7852
#> 71: France 2020-09-17    9784
#> 72: France 2020-09-18   10593
#> 73: France 2020-09-19   13215
#> 74: France 2020-09-20   13498
#> 75: France 2020-09-21   11569
#> 76: France 2020-09-22    4298
#> 77: France 2020-09-23   10008
#> 78: France 2020-09-24   13072
#> 79: France 2020-09-25   16096
#> 80: France 2020-09-26   15797
#> 81: France 2020-09-27   14412
#> 82: France 2020-09-28   11123
#> 83: France 2020-09-29    4070
#> 84: France 2020-09-30    8051
#> 85: France 2020-10-01   12845
#>     region       date confirm
#>